### 2022-11-16 #4940 @Kevin-Mattheus-Moerman
@alex-golts thanks for this submission to JOSS. I am the AEiC for this track and have started looking for a handling editor. While we are waiting for an editor to pick this up, please work on the following:

- [ ] check  those potentially missing DOI's :point_up: , you can call `@editorialbot check references` to check them again, and use `@editorialbot generate pdf` to update the paper draft.
- [ ] Can you add the city to your affiliation?
