+++ paper/paper.bib
@@ -62,7 +62,8 @@ publisher="Springer International Publishing",
 address="Cham",
 pages="34--43",
 abstract="Simulating the disease progression of a suspicious finding in mammography (MG) images could assist in the early detection of breast cancer, where the telltale signs of malignancy are subtle and hard to detect. It could also decrease both unnecessary biopsies and treatments of indolent or low-grade disease that might otherwise remain asymptomatic. We propose a novel approach to simulating disease progression, as a significant step towards those goals. Our architecture uses the powerful Wasserstein GAN in combination with a novel component that simulates the progression of the disease in deep feature space. This allows us to learn from unlabeled longitudinal MG pairs of current and prior studies, stabilize the learning procedure, and overcome misalignment between the MG pairs. Our output image replicates an actual MG image, maintains the prior's shape and general appearance while also containing a finding with characteristics that resemble the current image's suspicious finding. We simulate a progression of: (i) a full MG prior image in low-resolution, and (ii) a high-resolution patch in suspicious areas of the prior image. We demonstrate the effectiveness of our pipeline in achieving the above goals using quantitative and qualitative metrics and a reader study. Our results show the high quality of our simulation and the promise it holds for early risk stratification.",
-isbn="978-3-030-87592-3"
+isbn="978-3-030-87592-3",
+doi="10.1007/978-3-030-87592-3_4"
 }
 
 @InProceedings{tlusty2021pre,
@@ -89,7 +90,8 @@ publisher="Springer International Publishing",
 address="Cham",
 pages="277--286",
 abstract="Characterization of lesions by artificial intelligence (AI) has been the subject of extensive research. In recent years, many studies demonstrated the ability of convolution neural networks (CNNs) to successfully distinguish between malignant and benign breast lesions in mammography (MG) images. However, to date, no study has assessed the specific sub-type of lesions in MG images, as detailed in histolopathology reports. We present a method for finer classification of breast lesions in MG images into multiple pathology sub-types. Our approach works well with radiologists' diagnostic workflow, and uses data available in radiology reports. The proposed Dual-Radiology Dual-Resolution Network (Du-Rad Du-Res Net) receives dual input from the radiologist and dual image resolutions. The radiologist input includes annotation of the lesion area and semantic radiology features; the dual image resolutions comprise a low resolution of the entire mammogram and a high resolution of the lesion area. The network estimates the likelihood of malignancy, as well as the associated pathological sub-type. We show that the combined input of the lesion region of interest (ROI) and the entire mammogram is important for optimizing the model's performance. We tested the AI in a reader study on a dataset of 100 heldout cases. The AI outperformed three breast radiologists in the task of lesion histopathology sub-typing.",
-isbn="978-3-030-87589-3"
+isbn="978-3-030-87589-3",
+doi="10.1007/978-3-030-87589-3_29"
 }
 
 @InProceedings{golts2022ensemble,
@@ -111,7 +113,8 @@ publisher="Springer International Publishing",
 address="Cham",
 pages="103--115",
 abstract="Automatic segmentation of renal tumors and surrounding anatomy in computed tomography (CT) scans is a promising tool for assisting radiologists and surgeons in their efforts to study these scans and improve the prospect of treating kidney cancer. We describe our approach, which we used to compete in the 2021 Kidney and Kidney Tumor Segmentation (KiTS21) challenge. Our approach is based on the successful 3D U-Net architecture with our added innovations, including the use of transfer learning, an unsupervised regularized loss, custom postprocessing, and multi-annotator ground truth that mimics the evaluation protocol. Our submission has reached the 2nd place in the KiTS21 challenge.",
-isbn="978-3-030-98385-7"
+isbn="978-3-030-98385-7",
+doi="10.1007/978-3-030-98385-7_14"
 }
 
 @misc{knight,
