### 2022-11-17 #4940 @alex-golts
> @alex-golts thanks for this submission to JOSS. I am the AEiC for this track and have started looking for a handling editor. While we are waiting for an editor to pick this up, please work on the following:
>
> * [ ]  check  those potentially missing DOI's ☝️ , you can call `@editorialbot check references` to check them again, and use `@editorialbot generate pdf` to update the paper draft.
> * [ ]  Can you add the city to your affiliation?

done, thank you!
