### 2022-11-17 fad7df7
added axis to softmax
