+++ examples/fuse_examples/imaging/segmentation/PI-CAI/run_test.py
@@ -1,0 +1,205 @@
+import json
+from pathlib import Path
+
+import numpy as np
+import SimpleITK as sitk
+import torch
+from evalutils import SegmentationAlgorithm
+from evalutils.validators import (UniqueImagesValidator,
+                                  UniquePathIndicesValidator)
+
+from fuseimg.datasets.picai_test import PICAI
+from picai_baseline.unet.training_setup.default_hyperparam import \
+    get_default_hyperparams
+from picai_baseline.unet.training_setup.neural_network_selector import \
+    neural_network_for_run
+from fuse.data.utils.collates import CollateDefault
+from picai_baseline.unet.training_setup.preprocess_utils import z_score_norm
+from picai_prep.data_utils import atomic_image_write
+from picai_prep.preprocessing import Sample, PreprocessingSettings, crop_or_pad, resample_img
+from report_guided_annotation import extract_lesion_candidates
+from scipy.ndimage import gaussian_filter
+from runner import create_model
+from torch.utils.data.dataloader import DataLoader
+from pytorch_lightning import Trainer
+from fuse.dl.lightning.pl_module import LightningModuleDefault
+
+class csPCaAlgorithm(SegmentationAlgorithm):
+    """
+    Wrapper to deploy trained baseline U-Net model from
+    https://github.com/DIAGNijmegen/picai_baseline as a
+    grand-challenge.org algorithm.
+    """
+
+    def __init__(self):
+        super().__init__(
+            validators=dict(
+                input_image=(
+                    UniqueImagesValidator(),
+                    UniquePathIndicesValidator(),
+                )
+            ),
+        )
+
+        # set expected i/o paths in gc env (image i/p, algorithms, prediction o/p)
+        # see grand-challenge.org/algorithms/interfaces/ for expected path per i/o interface
+        # note: these are fixed paths that should not be modified
+
+        # directory to model weights
+        self.algorithm_weights_dir = Path("/opt/algorithm/weights/")
+
+        # path to image files
+        self.image_input_dirs = [
+            "/input/images/transverse-t2-prostate-mri/",
+            # "/input/images/transverse-adc-prostate-mri/",
+            # "/input/images/transverse-hbv-prostate-mri/",
+            # "/input/images/coronal-t2-prostate-mri/",  # not used in this algorithm
+            # "/input/images/sagittal-t2-prostate-mri/"  # not used in this algorithm
+        ]
+        self.image_input_paths = [list(Path(x).glob("*.mha"))[0] for x in self.image_input_dirs]
+
+        # load clinical information
+        with open("/input/clinical-information-prostate-mri.json") as fp:
+            self.clinical_info = json.load(fp)
+
+        # path to output files
+        self.detection_map_output_path = Path("/output/images/cspca-detection-map/cspca_detection_map.mha")
+        self.case_level_likelihood_output_file = Path("/output/cspca-case-level-likelihood.json")
+
+        # create output directory
+        self.detection_map_output_path.parent.mkdir(parents=True, exist_ok=True)
+
+        # define compute used for training/inference ('cpu' or 'cuda')
+        if torch.cuda.is_available():
+            self.device = 'cuda'
+        else:
+            self.device = 'cpu'
+
+        # extract available clinical metadata (not used for this example)
+        self.age = self.clinical_info["patient_age"]
+
+        if "PSA_report" in self.clinical_info:
+            self.psa = self.clinical_info["PSA_report"]
+        else:
+            self.psa = None  # value is missing, if not reported
+
+        if "PSAD_report" in self.clinical_info:
+            self.psad = self.clinical_info["PSAD_report"]
+        else:
+            self.psad = None  # value is missing, if not reported
+
[…]
