+++ examples/fuse_examples/imaging/segmentation/PI-CAI/runner.py
@@ -20,8 +20,6 @@ import sys
 import copy
 from fuse.eval.metrics.classification.metrics_thresholding_common import MetricApplyThresholds
 import numpy as np
-from pathlib import Path
-import pickle
 
 from fuse.utils.utils_debug import FuseDebug
 from fuse.utils.gpu import choose_and_enable_multiple_gpus
@@ -117,9 +115,19 @@ def pre_proc_batch(in_batch): # [N, C, D, H, W]
 
 def post_proc_batch(out_model): # [N * D, C, H, W]
     # return torch.unsqueeze(out_model,dim=0).transpose(1,2) # [N, C, D, H, W]
-    softmax_values =F.softmax(out_model,dim=1)
+    softmax_values =F.softmax(out_model,dim=0)
     return softmax_values
     
+
+    # n_slices = 23
+    # n_all, ch, h, w = out_model.shape
+    # nb = int(n_all / n_slices)
+    # if nb == 0:
+    #     import ipdb; ipdb.set_trace(context=7) # BREAKPOINT
+
+    # out_model = out_model.view(nb, n_slices, ch, h, w)
+    # return out_model.transpose(1,2) # [N, C, D, H, W]
+
 
 class Val_collate(CollateDefault):
 
@@ -176,7 +184,7 @@ class Val_collate(CollateDefault):
         return batch_dict
 
 
-def create_model(train: NDict) -> torch.nn.Module:
+def create_model(train: NDict,paths: NDict) -> torch.nn.Module:
     """
     creates the model
     See HeadGlobalPoolingClassifier for details
@@ -277,7 +285,7 @@ def run_train(paths: NDict, train: NDict) -> torch.nn.Module:
     # split to folds randomly - temp
     dataset_all = PICAI.dataset(
         paths = paths,
-        cfg = train,
+        train_cfg = train,
         reset_cache=False,
         train=True,
     )
@@ -299,7 +307,7 @@ def run_train(paths: NDict, train: NDict) -> torch.nn.Module:
 
     train_dataset = PICAI.dataset(
         paths = paths,
-        cfg = train,
+        train_cfg = train,
         reset_cache=False,
         sample_ids=train_sample_ids,  #[:10],
         train=True,
@@ -307,7 +315,7 @@ def run_train(paths: NDict, train: NDict) -> torch.nn.Module:
 
     validation_dataset = PICAI.dataset(
         paths = paths,
-        cfg = train,
+        train_cfg = train,
         reset_cache=False,
         sample_ids=validation_sample_ids  #[:10]
     )
@@ -448,10 +456,10 @@ def run_train(paths: NDict, train: NDict) -> torch.nn.Module:
 ######################################
 # Inference Template
 ######################################
-def run_infer(infer: NDict, paths: NDict):
+def run_infer(train: NDict, paths: NDict, infer: NDict):
     create_dir(paths["inference_dir"])
     #### Logger
-    # fuse_logger_start(output_path=paths["inference_dir"], console_verbose_level=logging.INFO)
+    fuse_logger_start(output_path=paths["inference_dir"], console_verbose_level=logging.INFO)
     lgr = logging.getLogger("Fuse")
     lgr.info("Fuse Inference", {"attrs": ["bold", "underline"]})
     infer_file = os.path.join(paths["inference_dir"], infer["infer_filename"])
@@ -472,20 +480,19 @@ def run_infer(infer: NDict, paths: NDict):
         infer_sample_ids += folds[fold]
 
     test_dataset = PICAI.dataset(
-        paths=paths,
-        cfg=infer,
+        paths["data_dir"],
+        paths["data_misc_dir"],
+        infer["target"],
+        paths["cache_dir"],
         sample_ids=infer_sample_ids,
-        reset_cache=False,
+        train=False,
     )
-
[…]
