+++ examples/fuse_examples/imaging/segmentation/PI-CAI/runner.py
@@ -1,5 +1,4 @@
 """
-
 (C) Copyright 2021 IBM Corp.
 Licensed under the Apache License, Version 2.0 (the "License");
 you may not use this file except in compliance with the License.
@@ -11,7 +10,6 @@ WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 See the License for the specific language governing permissions and
 limitations under the License.
 Created on June 30, 2021
-
 """
 
 from collections import OrderedDict
@@ -20,6 +18,8 @@ import sys
 import copy
 from fuse.eval.metrics.classification.metrics_thresholding_common import MetricApplyThresholds
 import numpy as np
+from pathlib import Path
+import pickle
 
 from fuse.utils.utils_debug import FuseDebug
 from fuse.utils.gpu import choose_and_enable_multiple_gpus
@@ -115,19 +115,9 @@ def pre_proc_batch(in_batch): # [N, C, D, H, W]
 
 def post_proc_batch(out_model): # [N * D, C, H, W]
     # return torch.unsqueeze(out_model,dim=0).transpose(1,2) # [N, C, D, H, W]
-    softmax_values =F.softmax(out_model,dim=0)
+    softmax_values =F.softmax(out_model,dim=1)
     return softmax_values
     
-
-    # n_slices = 23
-    # n_all, ch, h, w = out_model.shape
-    # nb = int(n_all / n_slices)
-    # if nb == 0:
-    #     import ipdb; ipdb.set_trace(context=7) # BREAKPOINT
-
-    # out_model = out_model.view(nb, n_slices, ch, h, w)
-    # return out_model.transpose(1,2) # [N, C, D, H, W]
-
 
 class Val_collate(CollateDefault):
 
@@ -184,7 +174,7 @@ class Val_collate(CollateDefault):
         return batch_dict
 
 
-def create_model(train: NDict,paths: NDict) -> torch.nn.Module:
+def create_model(target) -> torch.nn.Module:
     """
     creates the model
     See HeadGlobalPoolingClassifier for details
@@ -212,7 +202,7 @@ def create_model(train: NDict,paths: NDict) -> torch.nn.Module:
     #     gt_label = "data.gt.subtype"
     #     skip_keys = ["data.gt.classification"]
     #     class_names = ["Luminal A", "Luminal B", "HER2-enriched", "triple negative"]
-    if train['target'] == 'segmentation':
+    if target == 'segmentation':
         torch_model = UNet(n_channels=1, n_classes=1, bilinear=False)
 
         model = ModelWrapSeqToDict(model=torch_model,
@@ -222,7 +212,7 @@ def create_model(train: NDict,paths: NDict) -> torch.nn.Module:
                                 post_forward_processing_function=post_proc_batch
                                 )
 
-    elif train['target'] == 'seg3d':
+    elif target == 'seg3d':
 
         # define the model specifications used for initialization at train-time
         # note: if the default hyperparam listed in picai_baseline was used,
@@ -285,7 +275,7 @@ def run_train(paths: NDict, train: NDict) -> torch.nn.Module:
     # split to folds randomly - temp
     dataset_all = PICAI.dataset(
         paths = paths,
-        train_cfg = train,
+        cfg = train,
         reset_cache=False,
         train=True,
     )
@@ -307,7 +297,7 @@ def run_train(paths: NDict, train: NDict) -> torch.nn.Module:
 
     train_dataset = PICAI.dataset(
         paths = paths,
-        train_cfg = train,
+        cfg = train,
         reset_cache=False,
         sample_ids=train_sample_ids,  #[:10],
         train=True,
@@ -315,7 +305,7 @@ def run_train(paths: NDict, train: NDict) -> torch.nn.Module:
 
[…]
