+++ examples/fuse_examples/imaging/segmentation/PI-CAI/run_test.py
@@ -112,9 +112,7 @@ class csPCaAlgorithm(SegmentationAlgorithm):
         # 'model_type' via function 'get_default_hyperparams' is enough.
         # otherwise arguments 'model_strides' and 'model_features' must also
         # be explicitly passed directly to function 'neural_network_for_run'
-        train={}
-        train['target'] == 'seg3d'
-        model = create_model(train,{})
+        model = create_model('seg3d')
         # load python lightning module
         pl_module = LightningModuleDefault.load_from_checkpoint(
             checkpoint_path, model_dir=self.algorithm_weights_dir, model=model, map_location=self.device, strict=True
