+++ examples/fuse_examples/imaging/segmentation/PI-CAI/picai_test.py
@@ -1,0 +1,236 @@
+from fuse.data.datasets.dataset_default import DatasetDefault
+from fuse.data.datasets.caching.samples_cacher import SamplesCacher
+from fuseimg.data.ops.image_loader import OpLoadImage
+from fuseimg.data.ops.color import OpNormalizeAgainstSelf
+from fuseimg.data.ops.aug.geometry import OpAugAffine2D, OpAugSqueeze3Dto2D, OpAugUnsqueeze3DFrom2D
+from fuse.data import PipelineDefault, OpToTensor, OpRepeat
+from fuse.data.ops.ops_common import OpLambda
+from fuseimg.data.ops.aug.color import OpAugColor
+from fuseimg.data.ops.aug.geometry import OpAugAffine2D
+from fuse.data.ops.ops_common import OpConcat, OpLambda, OpLookup, OpToOneHot
+from fuse.data.ops.ops_aug_common import OpSample, OpRandApply, OpSampleAndRepeat
+from fuse.data.ops.ops_read import OpReadDataframe
+from fuse.data.ops.ops_cast import OpToNumpy
+from fuse.data.ops.op_base import OpBase
+from fuse.utils import NDict
+from functools import partial
+import nibabel as nib
+from typing import Hashable, Optional, Sequence
+import torch
+import pandas as pd
+import numpy as np
+import skimage
+import pydicom
+import os
+import glob
+from pathlib import Path
+from fuse.data.utils.sample import get_sample_id
+from medpy.io import load
+from fuse.utils.rand.param_sampler import Uniform, RandInt, RandBool
+
+
+class OpSample3dImg(OpBase):
+    """
+    Squeeze selected axis of volume image into channel dimension, in order to fit the 2D augmentation functions
+    """
+
+    def __init__(self, verify_arguments: bool = True):
+        """
+        :param verify_arguments: this ops expects torch tensor with 4 dimensions. Set to False to disable verification
+        """
+        super().__init__()
+        self._verify_arguments = verify_arguments
+
+    def __call__(self, sample_dict: NDict, key: str, axis: int, slice: float) -> NDict:
+        """
+        :param key: key to a tensor stored in sample_dict: 3D tensor representing an image to augment, shape [num_channels, spatial axis 1, spatial axis 2, spatial axis 3]
+        :param axis_squeeze: the axis (1, 2 or 3) to squeeze into channel dimension - typically z axis
+        :param slice: random number in [0, 1] that define the selected slice
+        """
+        # print(key, slice)
+        aug_input = sample_dict[key]
+        n_slices = aug_input.shape[axis]
+        n = int(n_slices * slice)
+
+        # verify
+        if self._verify_arguments:
+            assert isinstance(
+                aug_input, torch.Tensor
+            ), f"Error: OpSample3dImg expects torch Tensor, got {type(aug_input)}"
+            # assert (
+            #     len(aug_input.shape) == 4
+            # ), f"Error: OpSample3dImg expects tensor with 4 dimensions. got {aug_input.shape}"
+
+        aug_output = aug_input[n]
+
+        # add channel dim - TODO -check in case of multiple channels
+        aug_output = torch.unsqueeze(aug_output, dim=0)
+
+        sample_dict[key] = aug_output
+        return sample_dict
+
+
+class OpPICAISampleIDDecode(OpBase):
+    """
+    decodes sample id into image and segmentation filename
+    """
+
+    def __call__(self, sample_dict: NDict) -> NDict:
+        """ """
+        sid = get_sample_id(sample_dict)
+
+        img_filename_key = "data.input.img_path"
+        sample_dict[img_filename_key] = sid
+
+        return sample_dict
+
[…]
