### 2022-11-17 #4943 @editorialbot
[REVIEW]: FuseMedML: a framework for accelerated discovery in machine learning based biomedicine
**Submitting author:** @alex-golts (<a href="http://orcid.org/0000-0002-1478-3881">Alex Golts</a>)
**Repository:** https://github.com/BiomedSciAI/fuse-med-ml
**Branch with paper.md** (empty if default branch): joss_paper
**Version:** 0.2.9
