+++ examples/fuse_examples/imaging/classification/PI-CAI/conf/config.yaml
@@ -11,12 +11,12 @@ paths:
 run: 
   running_modes : ['train', 'infer', 'eval']
 train:
-  target : classification
+  target : subtype
   reset_cache: False
   num_workers : 12
   num_folds : 5
-  train_folds : [0,1,2]
-  validation_folds : [3]
+  train_folds : [0,1,2,4]
+  validation_folds : [4]
   batch_size: 8
   learning_rate : 1e-4
   weight_decay :  0
