+++ examples/fuse_examples/imaging/classification/PI-CAI/conf/config.yaml
@@ -15,7 +15,7 @@ train:
   reset_cache: False
   num_workers : 12
   num_folds : 5
-  train_folds : [0,1,2,4]
+  train_folds : [0,1,2,3]
   validation_folds : [4]
   batch_size: 8
   learning_rate : 1e-4
