### 2022-11-21 5400e25
feat: added support for logits in some classification metrics
