+++ fuse/eval/metrics/libs/classification.py
@@ -23,6 +23,8 @@ import numpy as np
 
 from sklearn import metrics
 import sklearn
+from sklearn.utils.multiclass import type_of_target
+
 
 import matplotlib.pyplot as plt
 
@@ -138,14 +140,20 @@ class MetricsLibClass:
     ):
         """
         Compute accuracy score
-        :param pred: class prediction. Each element is an integer in range [0 - num_classes)
-        :param target: the target class. Each element is an integer in range [0 - num_classes)
+        :param pred: class prediction. Each element is an integer in range [0 - num_classes) or a float in range [0-1]
+        :param target: the target class. Each element is an integer in range [0 - num_classes) or its one hot encoded version
         :param sample_weight: Optional - weight per sample for a weighted score. Each element is  float in range [0-1]
         :return: accuracy score
         """
 
         pred = np.array(pred)
         target = np.array(target)
+
+        if type_of_target(pred) in ("continuous", "continuous-multioutput"):
+            pred = np.argmax(pred, -1)
+        if type_of_target(target) in ("multilabel-indicator", "multiclass-multioutput"):
+            target = np.argmax(target, -1)
+
         return metrics.accuracy_score(target, pred, sample_weight=sample_weight)
 
     @staticmethod
@@ -159,8 +167,8 @@ class MetricsLibClass:
         """
         Compute metrics derived from one-vs-rest confusion matrix such as 'sensitivity', 'recall', 'tpr', 'specificity',  'selectivity', 'npr', 'precision', 'ppv', 'f1'
         Assuming that there are positive cases and negative cases in targets
-        :param pred: class prediction. Each element is an integer in range [0 - num_classes)
-        :param target: the target class. Each element is an integer in range [0 - num_classes)
+        :param pred: class prediction. Each element is an integer in range [0 - num_classes) or a float in range [0-1]
+        :param target: the target class. Each element is an integer in range [0 - num_classes) or its one hot encoded version
         :param pos_class_index: the class to compute the metrics in one vs rest manner - set to 1 in binary classification
         :param metrics: required metrics names, options: 'sensitivity', 'recall', 'tpr', 'specificity',  'selectivity', 'npr', 'precision', 'ppv', 'f1'
         :param sample_weight: Optional - weight per sample for a weighted score. Each element is  float in range [0-1]
@@ -169,6 +177,11 @@ class MetricsLibClass:
         """
         pred = np.array(pred)
         target = np.array(target)
+
+        if type_of_target(pred) in ("continuous", "continuous-multioutput"):
+            pred = np.argmax(pred, -1)
+        if type_of_target(target) in ("multilabel-indicator", "multiclass-multioutput"):
+            target = np.argmax(target, -1)
 
         class_target_t = np.where(target == pos_class_index, 1, 0)
         class_pred_t = np.where(pred == pos_class_index, 1, 0)
@@ -212,13 +225,19 @@ class MetricsLibClass:
     ) -> Dict[str, pd.DataFrame]:
         """
         Calculates Confusion Matrix (multi class version)
-        :param cls_pred: sequence of class prediction
-        :param target: sequence of labels
+        :param cls_pred: sequence of class prediction or a floats in range [0-1]
+        :param target: sequence of labels or their one hot encoded versions
         :param class_names: string name per class
         :param sample_weight: optional, weight per sample.
         :return: {"count": <confusion matrix>, "percent" : <confusion matrix - percent>)
         Confusion matrix whose i-th row and j-th column entry indicates the number of samples with true label being i-th class and predicted label being j-th class.
         """
+
[…]
