### 2022-11-21 7366e3c
adding check for class attribute, fixes issue #217
