### 2022-11-21 7d5e6c6
fuse function for clearml logger
