### 2022-11-22 700b732
adding check for class attribute, fixes issue #217 (#218)
