+++ fuse/dl/models/backbones/backbone_resnet_3d.py
@@ -80,7 +80,7 @@ class BackboneResnet3D(VideoResNet):
         x = self.layer2(x)
         x = self.layer3(x)
         x = self.layer4(x)
-        if self.pool:
+        if hasattr(self, "pool") and self.pool:
             x = self.gmp(x).flatten(1)
         return x
 
