### 2022-11-22 904734e
train both on classification and regression
