+++ examples/fuse_examples/imaging/classification/PI-CAI/runner.py
@@ -70,12 +70,10 @@ def create_model(train: NDict, paths: NDict) -> torch.nn.Module:
     if train["target"] == "classification":
         num_classes = 2
         gt_label = "data.gt.classification"
-        skip_keys = ["data.gt.subtype"]
         class_names = ["Benign", "Malignant"]
     elif train["target"] == "subtype":
         num_classes = 6
         gt_label = "data.gt.subtype"
-        skip_keys = ["data.gt.classification"]
         class_names = ["Healthy -0 ", "ISUP 1", "ISUP 2", "ISUP 3", "ISUP 4", "ISUP 5"]
     else:
         raise ("unsuported target!!")
@@ -85,6 +83,7 @@ def create_model(train: NDict, paths: NDict) -> torch.nn.Module:
         heads=[
             Head3D(head_name='head_0',
                              conv_inputs=[("model.backbone_features", 512)],
+                             mode = "classification",
                              #  dropout_rate=train_params['imaging_dropout'],
                              #  append_dropout_rate=train_params['clinical_dropout'],
                              #  fused_dropout_rate=train_params['fused_dropout'],
@@ -92,8 +91,18 @@ def create_model(train: NDict, paths: NDict) -> torch.nn.Module:
                              #  append_features=[("data.input.clinical", 8)],
                              #  append_layers_description=(256,128),
                              ),
+            Head3D(head_name='head_1',
+                             conv_inputs=[("model.backbone_features", 512)],
+                             mode = "regression",
+                             #  dropout_rate=train_params['imaging_dropout'],
+                             #  append_dropout_rate=train_params['clinical_dropout'],
+                             #  fused_dropout_rate=train_params['fused_dropout'],
+                             num_outputs=1,
+                             #  append_features=[("data.input.clinical", 8)],
+                             #  append_layers_description=(256,128),
+                             ),
         ])
-    return model, num_classes, gt_label, skip_keys, class_names
+    return model, num_classes, gt_label, class_names
 
 
 #################################
@@ -112,7 +121,7 @@ def run_train(paths: NDict, train: NDict) -> torch.nn.Module:
     # ==============================================================================
     lgr.info("Model:", {"attrs": "bold"})
 
-    model, num_classes, gt_label, skip_keys, class_names = create_model(train, paths)
+    model, num_classes, gt_label, class_names = create_model(train, paths)
     lgr.info("Model: Done", {"attrs": "bold"})
 
     lgr.info("\nFuse Train", {"attrs": ["bold", "underline"]})
@@ -125,7 +134,7 @@ def run_train(paths: NDict, train: NDict) -> torch.nn.Module:
     dataset_all = PICAI.dataset(
         paths["data_dir"],
         paths["clinical_file"],
-        train["target"],
+        train,
         paths["cache_dir"],
         reset_cache=False,
         num_workers=train["num_workers"],
@@ -151,7 +160,7 @@ def run_train(paths: NDict, train: NDict) -> torch.nn.Module:
     train_dataset = PICAI.dataset(
         paths["data_dir"],
         paths["clinical_file"],
-        train["target"],
+        train,
         paths["cache_dir"],
         reset_cache=False,
         num_workers=train["num_workers"],
@@ -163,7 +172,7 @@ def run_train(paths: NDict, train: NDict) -> torch.nn.Module:
     validation_dataset = PICAI.dataset(
         paths["data_dir"],
         paths["clinical_file"],
-        train["target"],
+        train,
         paths["cache_dir"],
         reset_cache=False,
         num_workers=train["num_workers"],
[…]
