+++ examples/fuse_examples/imaging/classification/PI-CAI/conf/config.yaml
@@ -11,7 +11,9 @@ paths:
 run: 
   running_modes : ['train', 'infer', 'eval']
 train:
-  target : subtype
+  target : classification
+  series_config : ['_t2w']
+  resize_resolution : [23,384,384]
   reset_cache: False
   num_workers : 12
   num_folds : 5
