+++ examples/fuse_examples/imaging/classification/isic/isic_runner.py
@@ -46,12 +46,15 @@ from fuse.data.utils.collates import CollateDefault
 
 import pytorch_lightning as pl
 from fuse.dl.lightning.pl_module import LightningModuleDefault
-from fuse.dl.lightning.pl_funcs import convert_predictions_to_dataframe
+from fuse.dl.lightning.pl_funcs import convert_predictions_to_dataframe, start_clearml_logger
 from pytorch_lightning.utilities.rank_zero import rank_zero_only
 
 from fuseimg.datasets.isic import ISIC, ISICDataModule
 from fuse_examples.imaging.classification.isic.golden_members import FULL_GOLDEN_MEMBERS
 
+import torch.nn as nn
+from fuse.dl.models.model_wrapper import ModelWrapSeqToDict
+from fuse.dl.models.backbones.backbone_vit import ViT
 
 ###########################################################################################################
 # Fuse
@@ -66,20 +69,24 @@ debug = FuseDebug(mode)
 # GPUs and Workers
 ##########################################
 NUM_GPUS = 1  # supports multiple gpu training with DDP strategy
-NUM_WORKERS = 8
+NUM_WORKERS = 16
 
 ##########################################
 # Modality
 ##########################################
+multimodality = True # Set: 'False' to use only imaging, 'True' to use imaging & meta-data
 
-multimodality = True  # Set: 'False' to use only imaging, 'True' to use imaging & meta-data
+##########################################
+# Model Type
+##########################################
+model_type = "CNN" # Set: 'Transformer' to use ViT/MMViT, 'CNN' to use InceptionResNet
 
 ##########################################
 # Output Paths
 ##########################################
 
 
-ROOT = "./_examples/isic/"
+ROOT = "/dccstor/mm_hcls/shatz/_examples/isic/"
 DATA = os.environ["ISIC19_DATA_PATH"] if "ISIC19_DATA_PATH" in os.environ else os.path.join(ROOT, "data_dir")
 modality = "multimodality" if multimodality else "imaging"
 model_dir = os.path.join(ROOT, f"model_dir_{modality}")
@@ -100,7 +107,7 @@ TRAIN_COMMON_PARAMS = {}
 # ============
 # Data
 # ============
-TRAIN_COMMON_PARAMS["data.batch_size"] = 32  # effective batch size = batch_size * num_gpus
+TRAIN_COMMON_PARAMS["data.batch_size"] = 64  # effective batch size = batch_size * num_gpus
 TRAIN_COMMON_PARAMS["data.num_workers"] = NUM_WORKERS
 TRAIN_COMMON_PARAMS["data.num_folds"] = 5
 TRAIN_COMMON_PARAMS["data.train_folds"] = [0, 1, 2]
@@ -126,15 +133,63 @@ TRAIN_COMMON_PARAMS["opt.weight_decay"] = 1e-3
 # ===============
 # Model
 # ===============
-TRAIN_COMMON_PARAMS["model"] = dict(
-    dropout_rate=0.5,
-    layers_description=(256,),
-    tabular_data_inputs=[("data.input.clinical.all", 19)] if multimodality else None,
-    tabular_layers_description=(128,) if multimodality else tuple(),
-)
+if model_type == "CNN":
+    TRAIN_COMMON_PARAMS["model"] = dict(
+        dropout_rate=0.5,
+        layers_description=(256,),
+        tabular_data_inputs=[("data.input.clinical.all", 19)] if multimodality else None,
+        tabular_layers_description=(128,) if multimodality else tuple(),
+    )
+elif model_type == "Transformer":
+    TRAIN_COMMON_PARAMS["model"] = dict(
+        random_dict = None
+    )
 
[…]
