### 2022-11-23 21a0d2f
fix pl_funcs args
