### 2022-11-23 e5de772
black, flake8, and fixed some bugs in inference
