### 2022-11-23 31a8f07
made some args for clearml logger mandatory and added clearml to requirements
