### 2022-11-23 8e9a212
removed clearml and cleaned up code
