### 2022-11-27 e690e91
added some extra instructions for clearml
