+++ fuse/eval/metrics/libs/survival.py
@@ -10,6 +10,7 @@ class MetricsSurvival:
         event_observed: np.ndarray,
         event_class_index: int = -1,
         time_unit: int = 1,
+        time_followup: int = None,
     ) -> float:
         """
         Compute c-index (concordance index) score using lifelines
@@ -23,6 +24,12 @@ class MetricsSurvival:
             pred = np.asarray(pred)[:, event_class_index]
 
         event_times = (np.array(event_times) / time_unit).astype(int)
+
+        if time_followup is not None:
+            event_observed = np.array(event_observed)
+            event_times = np.array(event_times)
+            print(f"C-index time_follow_up={time_followup}: ignored events ={event_observed[event_times>time_followup].sum()} remaining_events={event_observed[event_times<=time_followup].sum()}")
+            event_observed[event_times>time_followup] = 0
 
         return concordance_index(event_times, -pred, event_observed)
 
