### 2022-12-04 145ec67
mod for peptide design
