### 2022-12-05 0289ada
avoid from batching when there are missing values
