### 2022-12-08 81a7a70
added hints and better error messages for common mistakes when providing pipeline ops list
