+++ fuse/data/ops/op_base.py
@@ -23,6 +23,7 @@ from fuse.utils.ndict import NDict
 from fuse.data.ops.hashable_class import HashableClass
 import inspect
 
+
 class OpBase(HashableClass):
     """
     Operator Base Class
@@ -93,23 +94,26 @@ class OpReversibleBase(OpBase):
 
 
 def op_call(op: OpBase, sample_dict: NDict, op_id: str, **kwargs):
-    
-    if inspect.isclass(op):                    
-        raise Exception(f'Error: expected an instance object, not a class object for {op}\n' 
-        'When creating a pipeline, such error can happen when you provide the following ops list description:\n'
-        '[SomeOp, {}]\n'
-        'instead of \n'
-        '[SomeOp(), {}]\n'
+
+    if inspect.isclass(op):
+        raise Exception(
+            f"Error: expected an instance object, not a class object for {op}\n"
+            "When creating a pipeline, such error can happen when you provide the following ops list description:\n"
+            "[SomeOp, {}]\n"
+            "instead of \n"
+            "[SomeOp(), {}]\n"
         )
 
     try:
         if isinstance(op, OpReversibleBase):
             return op(sample_dict, op_id=op_id, **kwargs)
-        elif isinstance(op, OpBase): # OpBase but not reversible
+        elif isinstance(op, OpBase):  # OpBase but not reversible
             return op(sample_dict, **kwargs)
         else:
-            raise Exception(f'Ops are expected to be instances of classes or subclasses of OpBase. The following op is not: {op}')
-           
+            raise Exception(
+                f"Ops are expected to be instances of classes or subclasses of OpBase. The following op is not: {op}"
+            )
+
     except:
         # error messages are cryptic without this. For example, you can get "TypeError: __call__() got an unexpected keyword argument 'key_out_input'" , without any reference to the relevant op!
         print(
