### 2022-12-11 034ece1
support arbitrary (not sequenctial and even not ints) balanced class values
