+++ fuse/data/ops/ops_common.py
@@ -509,3 +509,14 @@ class OpCond(OpBase):
             return self._op(sample_dict, **kwargs)
         else:
             return sample_dict
+
+class OpSet(OpBase):
+    """Add/override key-value pair into sample_dict"""
+    def __call__(self, sample_dict: NDict, key: str, value: Any) -> Union[None, dict, List[dict]]:
+        """
+        :param key: where to store the value
+        :param value: the value to store
+        """
+        sample_dict[key] = value
+        return sample_dict
+
