+++ fuse/data/utils/samplers.py
@@ -1,258 +1,0 @@
-"""
-(C) Copyright 2021 IBM Corp.
-
-Licensed under the Apache License, Version 2.0 (the "License");
-you may not use this file except in compliance with the License.
-You may obtain a copy of the License at
-
-   http://www.apache.org/licenses/LICENSE-2.0
-
-Unless required by applicable law or agreed to in writing, software
-distributed under the License is distributed on an "AS IS" BASIS,
-WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-See the License for the specific language governing permissions and
-limitations under the License.
-
-Created on June 30, 2021
-
-"""
-
-import math
-import numpy as np
-from typing import Any, List, Optional, Union
-from torch.utils.data.sampler import Sampler, BatchSampler
-from fuse.data.datasets.dataset_base import DatasetBase
-
-
-class BatchSamplerDefault(BatchSampler):
-    """
-    Torch batch sampler - balancing per batch
-    """
-
-    def __init__(
-        self,
-        dataset: DatasetBase,
-        balanced_class_name: str,
-        num_balanced_classes: int,
-        sampler: Optional[Sampler] = None,
-        batch_size: Optional[int] = None,
-        mode: str = "exact",
-        balanced_class_weights: Union[List[int], List[float], None] = None,
-        num_batches: Optional[int] = None,
-        verbose: bool = False,
-        **dataset_get_multi_kwargs,
-    ) -> None:
-        """
-        :param dataset: dataset used to extract the balanced class from each sample
-        :param balanced_class_name:  the name of balanced class to extract from dataset
-        :param num_balanced_classes: number of classes to balance between
-        :param sampler: Optional - pytorch sampler for collecting the data.
-                        In use in DDP strategy, when PL trainer re-instantiate a custom batch_sampler with a DistributedSampler.
-        :param batch_size: batch_size.
-                        - In "exact" mode
-                            If balanced_class_weights is None, must be set and divided by num_balanced_classes. Otherwise keep None.
-                        - In "approx" mode
-                          Must be set
-        :param mode: either 'exact' or 'approx'. if 'exact each element in balanced_class_weights will specify the exact number of samples from this class.
-                        if 'approx' - each element will specify the a probability that a sample will be from this class
-        :param balanced_class_weights: Optional, integer/float per balanced class, Expected length is num_balanced_classes.
-                                        In mode 'exact' expecting list of integers that sums up to batch dict.
-                                        In mode 'approx' expecting list of floats that sums up to ~1
-                                        If not specified and equal number of samples from each class will be used.
-        :param num_batches: optional - if set will force num_batches, otherwise num_batches will be automatically to go over each sample at least once (exactly or approximately).
-        :param verbose:
-        :param dataset_get_multi_kwargs: extra parameters for dataset.get_multi() to optimize the running time.
-        """
-        super().__init__(sampler, batch_size, False)
-
-        # store input
-        self._dataset = dataset
-        self._balanced_class_name = balanced_class_name
-        self._num_balanced_classes = num_balanced_classes
-        self._sampler = sampler
-        self._batch_size = batch_size
-        self._mode = mode
-        self._balanced_class_weights = balanced_class_weights
-        self._num_batches = num_batches
-        self._dataset_get_multi_kwargs = dataset_get_multi_kwargs
-        self._verbose = verbose
-        # modify relevant keys
-        if self._balanced_class_name not in self._dataset_get_multi_kwargs:
-            self._dataset_get_multi_kwargs["keys"] = [self._balanced_class_name]
-
-        # validate input
-        # modes
-        if self._mode not in ["exact", "approx"]:
-            raise Exception("Error, expected sampler mode to be either 'exact' or 'approx', got {mode}")
-
-        # weights
[…]
