### 2022-12-13 787d50e
mark unused argument + add option to few ccls tranformer tokens
