### 2022-12-14 BiomedSciAI/fuse-med-ml#228 @anupamajha1
References missing in the Fuse-med-ML paper
The [fuse-med-ml paper](https://github.com/BiomedSciAI/fuse-med-ml/blob/joss_paper/paper/paper.md) is missing references. Please add[ proper citations](https://pandoc.org/MANUAL.html#extension-citations) for the papers referenced in the text (e.g. the following papers:[@raboh2022context], [@rabinovici2022early], [@rabinovici2022multimodal], [@jubran2021glimpse], [@tlusty2021pre], [@golts2022ensemble], [@radiology])
