### 2022-12-14 BiomedSciAI/fuse-med-ml#229 @anupamajha1
State of the field section missing from the fuse-med-ML paper
According to the JOSS review guidelines, the authors should provide a section titled 'State of the field' which describes how this software compares to other commonly-used packages. This section is currently missing from the [Fuse-Med-ML paper](https://github.com/BiomedSciAI/fuse-med-ml/blob/joss_paper/paper/paper.md).
