### 2022-12-14 BiomedSciAI/fuse-med-ml#230 @anupamajha1
Add example to Fuse.dl section in the Fuse-med ML paper
Examples in fuse.data and fuse.ml sections are extremely useful. Please expand fuse.dl section to include a similar example to explain its functionality.
