### 2022-12-14 #4943 @anupamajha1
Issues opened regarding
1. [missing references](https://github.com/BiomedSciAI/fuse-med-ml/issues/228).
2. [missing state of the field section](https://github.com/BiomedSciAI/fuse-med-ml/issues/229) in the paper.
3. [adding an example to the fuse.dl section](https://github.com/BiomedSciAI/fuse-med-ml/issues/230)
