### 2022-12-14 c44f9d6
remove torchvision requirement
