### 2022-12-14 6df870d
improved NDict's print_tree func
