### 2022-12-14 18927d1
support arbitrary (not sequential and even not ints) balanced class values (#227)
