+++ paper/paper.bib
@@ -162,3 +162,52 @@ eprint = {
 ,
     abstract = { Background Computational models based on artificial intelligence (AI) are increasingly used to diagnose malignant breast lesions. However, assessment from radiologic images of the specific pathologic lesion subtypes, as detailed in the results of biopsy procedures, remains a challenge. Purpose To develop an AI-based model to identify breast lesion subtypes with mammograms and linked electronic health records labeled with histopathologic information. Materials and Methods In this retrospective study, 26 569 images were collected in 9234 women who underwent digital mammography to pretrain the algorithms. The training data included individuals who had at least 1 year of clinical and imaging history followed by biopsy-based histopathologic diagnosis from March 2013 to November 2018. A model that combined convolutional neural networks with supervised learning algorithms was independently trained to make breast lesion predictions with data from 2120 women in Israel and 1642 women in the United States. Results were reported using the area under the receiver operating characteristic curve (AUC) with the 95\% DeLong approach to estimate CIs. Significance was tested with bootstrapping. Results The Israeli model was validated in 456 women and tested in 441 women (mean age, 51 years ± 11 [SD]). The U.S. model was validated in 350 women and tested in 344 women (mean age, 60 years ± 12). For predicting malignancy in the test sets (consisting of 220 Israeli patient examinations and 126 U.S. patient examinations with ductal carcinoma in situ or invasive cancer), the algorithms obtained an AUC of 0.88 (95\% CI: 0.85, 0.91) and 0.80 (95\% CI: 0.74, 0.85) for Israeli and U.S. patients, respectively (P = .006). These results may not hold for other cohorts of patients, and generalizability across populations should be further investigated. Conclusion The results offer supporting evidence that artificial intelligence applied to clinical and mammographic images can identify breast lesion subtypes when the data are sufficiently large, which may help assess diagnostic workflow and reduce biopsy sampling errors. Published under a CC BY 4.0 license. Online supplemental material is available for this article. }
 }
+
+@inproceedings{NEURIPS2019_bdbca288,
+ author = {Paszke, Adam and Gross, Sam and Massa, Francisco and Lerer, Adam and Bradbury, James and Chanan, Gregory and Killeen, Trevor and Lin, Zeming and Gimelshein, Natalia and Antiga, Luca and Desmaison, Alban and Kopf, Andreas and Yang, Edward and DeVito, Zachary and Raison, Martin and Tejani, Alykhan and Chilamkurthy, Sasank and Steiner, Benoit and Fang, Lu and Bai, Junjie and Chintala, Soumith},
+ booktitle = {Advances in Neural Information Processing Systems},
+ editor = {H. Wallach and H. Larochelle and A. Beygelzimer and F. d\textquotesingle Alch\'{e}-Buc and E. Fox and R. Garnett},
+ pages = {},
+ publisher = {Curran Associates, Inc.},
+ title = {PyTorch: An Imperative Style, High-Performance Deep Learning Library},
+ url = {https://proceedings.neurips.cc/paper/2019/file/bdbca288fee7f92f2bfa9f7012727740-Paper.pdf},
+ volume = {32},
+ year = {2019}
+}
+
+@software{Falcon_PyTorch_Lightning_2019,
[…]
