+++ paper/paper.md
@@ -52,9 +52,9 @@ Before open sourcing it, we used *fuse* internally in multiple research projects
 We were also able to meaningfully measure our progress and statistical significance of our results with off-the-shelf *fuse.eval* components that facilitate metrics' confidence interval calculation and model comparison. These tools have enabled us to organize two challenges as part of the 2022 International Symposium on Biomedical Imaging (ISBI) [@knight], [@bright].
 
 # State of the field
-FuseMedML is a comprehensive ML library with an emphasis on the biomedical domain. It provides a broad set of tools spanning the whole development pipeline, from data preparation, through model training to evaluation. It is built on top of leading ML frameworks such as PyTorch [@NEURIPS2019_bdbca288] and PyTorch Lightning [@Falcon_PyTorch_Lightning_2019], and attempts to complement them where needed, as well as introduce domain specific layers.  
-One way in which *fuse* is unique is in it's flexible design concept of storing data in a specialized nested dictionary. This is a key driver of flexibility, allowing minimal code modifications when moving building blocks between different projects.  
-There are existing PyTorch-based ML libraries that similarly to *fuse* cater to researchers in the biomedical domain. Two such prominent libraries are MONAI [@Cardoso_MONAI_An_open-source_2022] and PyHealth [@DBLP:journals/corr/abs-2101-04209]. One aspect in which *fuse* may complement MONAI is in the availability of tools for comparing between models as well as analyzing statistical significance of metric results. In PyHealth, unlike *fuse* the main focus appears to be on health records data, with not enough emphasis on medical imaging data and models. 
+FuseMedML is a comprehensive ML library with an emphasis on the biomedical domain. It provides a broad set of tools spanning the whole development pipeline, from data preparation, through model training to evaluation. It is built on top of leading ML frameworks such as PyTorch [@NEURIPS2019_bdbca288] and PyTorch Lightning [@Falcon_PyTorch_Lightning_2019], and attempts to complement them with optional flexible capabilities, as well as introduce domain specific layers.  
+One way in which *fuse* can complement PyTorch is through its design concept of storing data in a specialized nested dictionary. This is a key driver of flexibility, allowing minimal code modifications when moving building blocks between different projects. Concretely, *fuse* has a dataset class that extends the PyTorch dataset, and model wrapper class that enables PyTorch models to operate on `batch_dict`s.    
+There are existing PyTorch-based ML libraries that similarly to *fuse* cater to researchers in the biomedical domain. Two such prominent libraries are MONAI [@Cardoso_MONAI_An_open-source_2022] and PyHealth [@DBLP:journals/corr/abs-2101-04209]. One aspect in which *fuse* may complement MONAI is in the availability of tools for comparing between models as well as analyzing statistical significance of metric results. In PyHealth, unlike *fuse* the main focus appears to be on health records data, with not enough emphasis on medical imaging data and models. Another difference with respect to both libraries is *fuse*'s flexible caching mechanism, which allows the user to separate their processing into a static and dynamic pipelines, controlling the desired layers to be cached.
 
 # Packages
 
