+++ paper/paper.md
@@ -52,11 +52,11 @@ Before open sourcing it, we used *fuse* internally in multiple research projects
 We were also able to meaningfully measure our progress and statistical significance of our results with off-the-shelf *fuse.eval* components that facilitate metrics' confidence interval calculation and model comparison. These tools have enabled us to organize two challenges as part of the 2022 International Symposium on Biomedical Imaging (ISBI) [@knight], [@bright].
 
 # State of the field
-FuseMedML is a comprehensive ML library with an emphasis on the biomedical domain. It provides a broad set of tools spanning the whole development pipeline, from data preparation, through model training to evaluation. It is built on top of leading ML frameworks such as PyTorch [@NEURIPS2019_bdbca288] and PyTorch Lightning [@Falcon_PyTorch_Lightning_2019]. It is not meant as a substitute to popular frameworks, but offers to complement them with optional flexible capabilities, as well as introduce alternative components, including domain specific ones.  
+FuseMedML is a comprehensive ML library with an emphasis on the biomedical domain. It provides a broad set of tools spanning the whole development pipeline, from data preparation, through model training to evaluation. It is built on top of leading ML frameworks such as PyTorch [@NEURIPS2019_bdbca288] and PyTorch Lightning [@Falcon_PyTorch_Lightning_2019]. It is not meant as a substitute to these popular frameworks, but offers to complement them with optional flexible capabilities, as well as introduce alternative components, including domain specific ones.  
 One way in which *fuse* can complement PyTorch is through its generic design concept of storing arbitrary types of data in a specialized nested dictionary. This is a key driver of flexibility, allowing minimal code modifications when moving building blocks between different projects. Concretely, *fuse* has a dataset class that extends the PyTorch dataset, and a model wrapper class that enables PyTorch models to operate on `batch_dict`s rather than tensors.  
 In the case of PyTorch Lightning, *fuse* integrates with it directly as it builds upon its comprehensive trainer class, also allowing users to define their models and data modules in PyTorch Lightning style, with flexible levels of customizability.
 
-There are existing PyTorch-based ML libraries that similarly to *fuse* cater to researchers in the biomedical domain. Two such prominent libraries are MONAI [@Cardoso_MONAI_An_open-source_2022] and PyHealth [@DBLP:journals/corr/abs-2101-04209]. MONAI is primarily focused on medical imaging applications. PyHealth on the other hand mainly focuses on health records data. *fuse* is designed to support different types of medical data and multimodal use cases involving imaging, clinical and biochemical data.  
+There are existing PyTorch-based ML libraries that similarly to *fuse* cater to researchers in the biomedical domain. Two examples of such prominent libraries are MONAI [@Cardoso_MONAI_An_open-source_2022] and PyHealth [@DBLP:journals/corr/abs-2101-04209]. MONAI is primarily focused on medical imaging applications. PyHealth on the other hand mainly focuses on health records data. *fuse* is designed to support different types of medical data and multimodal use cases involving imaging, clinical and biochemical data.  
 As with generic ML frameworks like PyTorch and PyTorch Lightning, *fuse* can also coexist with the more specific libraries like MONAI, PyHealth or others. A user may opt to borrowing parts from different libraries and complement them with components from *fuse*. As another example, a user may want to use the data *ops* of *fuse* which are generic and flexible, or its data caching mechanism, which allows to separate processing into a static and dynamic pipelines, controlling the desired stages to be cached.  
 
 # Packages
