+++ paper/paper.md
@@ -57,7 +57,7 @@ One way in which *fuse* can complement PyTorch is through its generic design con
 In the case of PyTorch Lightning, *fuse* integrates with it directly as it builds upon its comprehensive trainer class, also allowing users to define their models and data modules in PyTorch Lightning style, with flexible levels of customizability.
 
 There are existing PyTorch-based ML libraries that similarly to *fuse* cater to researchers in the biomedical domain. Two examples of such prominent libraries are MONAI [@Cardoso_MONAI_An_open-source_2022] and PyHealth [@DBLP:journals/corr/abs-2101-04209]. MONAI is primarily focused on medical imaging applications. PyHealth on the other hand mainly focuses on health records data. *fuse* is designed to support different types of medical data and multimodal use cases involving imaging, clinical and biochemical data.  
-As with generic ML frameworks like PyTorch and PyTorch Lightning, *fuse* can also coexist with the more specific libraries like MONAI, PyHealth or others. A user may opt to borrowing parts from different libraries and complement them with components from *fuse*. As another example, a user may want to use the data *ops* of *fuse* which are generic and flexible, or its data caching mechanism, which allows to separate processing into a static and dynamic pipelines, controlling the desired stages to be cached.  
+As with generic ML frameworks like PyTorch and PyTorch Lightning, *fuse* can also coexist with the more specific libraries like MONAI, PyHealth or others. A user may opt to borrow parts from different libraries and complement them with components from *fuse*. As another example, a user may want to use the data *ops* of *fuse* which are generic and flexible, or its data caching mechanism, which allows to separate processing into a static and dynamic pipelines, controlling the desired stages to be cached.  
 
 # Packages
 
