### 2022-12-19 493bdfe
support ensemble on multi-modal survival data
