### 2022-12-21 BiomedSciAI/fuse-med-ml#229 @alex-golts
Added a "state of the field" section. [see here](https://github.com/BiomedSciAI/fuse-med-ml/blob/joss_paper/paper/paper.md)
