+++ paper/paper.md
@@ -75,7 +75,10 @@ The data package also includes generic utilities such as a PyTorch based sampler
 ![In this example a medical image loader is the *fuse* component reused in projects A and B. Different projects can have different formats for their data samples, but they can all use OpMedicalImageLoader by providing the appropriate key names when calling it. In Project B the same key name is used for the input and output, resulting in the loaded image data overriding the image paths in the updated sample.\label{fig:diagram}](figures/diagram.png){width=100%}
 
 ## *fuse.dl*
-FuseMedML's DL package works with PyTorch models, only modifying them to interact with a `batch_dict`. For training, *fuse.dl* utilizes PyTorch-Lightning, either through an already made `LightningModuleDefault` class that inherits from Pytorch-Lightning's `LightningModule` class, or by allowing users who seek maximal customizability to implement their own custom `LightningModule` and operate in close resemblance to the standard PyTorch-Lightning workflow or use alternative training loop implementations.   
+FuseMedML's DL package works with PyTorch models, only modifying them to interact with a `batch_dict`. For training, *fuse.dl* utilizes PyTorch-Lightning, either through an already made `LightningModuleDefault` class that inherits from Pytorch-Lightning's `LightningModule` class, or by allowing users who seek maximal customizability to implement their own custom `LightningModule` and operate in close resemblance to the standard PyTorch-Lightning workflow or use alternative training loop implementations.  
+*fuse.dl* also offers generic core DL components such as model architectures and losses, implemented in *fuse* style. See an example model architecture definition in \autoref{fig:model_multi_head}.
+
+![In this example a model architecture is defined using the `ModelMultiHead` class. It contains of a 3D ResNet backbone represented by the `BackboneResnet3D` class and a 3D classification head represented by the `Head3D` class. Note the user can define a list of heads, to support a multi task use case. The inputs to the backbone and classification heads are defined in the *fuse* style described earlier, using the `batch_dict` key names with the relevant data. This enables easy reuse of similar model architectures between projects.\label{fig:model_multi_head}](figures/model_multi_head.png){width=100%}
 
 ## *fuse.eval*
 The evaluation package of fuse is a standalone library for evaluating ML models for various performance metrics and comparing between models. It comes with advanced capabilities such as generic confidence interval wrapper for any metric, generic one vs. all extension for any binary metric to a multi-class scenario, metrics for model comparison taking into account statistical significance, model calibration tools, pipelining for combining a sequence of metrics with possible dependencies, automatic per-fold evaluation and sub-group analysis, automaric handling of massive data through batching and multiprocessing, and more. See \autoref{fig:metric_pipeline} for an example of an evaluation metric pipeline that can be reused across projects.  
