### 2022-12-21 BiomedSciAI/fuse-med-ml#230 @alex-golts
It later occurred to me that you most likely meant adding an example to the `fuse.dl` section in the paper itself and not in the code readme. So I also added one. please see [updated file here](https://github.com/BiomedSciAI/fuse-med-ml/blob/joss_paper/paper/paper.md).
