### 2022-12-21 3d31274
made the print keys optional
