### 2022-12-21 32b8652
fixed ndict's print_tree
