### 2022-12-21 8cf3661
fixed mypy & added documentation
