+++ fuse/utils/ndict.py
@@ -293,14 +293,14 @@ class NDict(dict):
 
         return ans
 
-    def print_tree(self, print_keys=False) -> None:
+    def print_tree(self, print_keys: bool = False) -> None:
         """
         print the inner structure of the nested dict with a tree-like structure.
         """
         self._print_tree_static(self._stored, print_keys=print_keys)
 
     @staticmethod
-    def _print_tree_static(data_dict: dict, level: int = 0, print_keys=False) -> None:
+    def _print_tree_static(data_dict: dict, level: int = 0, print_keys: bool = False) -> None:
         """
         static-method to print the inner structure of a dict in a tree-like structure.
         """
