### 2022-12-22 333b01f
addressed inline comment on documentation
