+++ fuse/dl/models/backbones/backbone_mlp.py
@@ -47,7 +47,6 @@ class BackboneMultilayerPerceptron(torch.nn.Module):
         self.mlp = nn.ModuleList(mlp_layers)
 
     def forward(self, input_tensor: torch.Tensor) -> torch.Tensor:
-        input_tensor = input_tensor.float()
 
         for layer in self.mlp:
             input_tensor = layer(input_tensor)
