### 2022-12-22 96bdcbb
raise exception in case of str input/outputs
