+++ fuse/dl/models/model_wrapper.py
@@ -61,6 +61,8 @@ class ModelWrapSeqToDict(torch.nn.Module):
         self.pre_forward_processing_function = pre_forward_processing_function
         self.post_forward_processing_function = post_forward_processing_function
         self.model_outputs = model_outputs
+        if isinstance(self.model_inputs, str) or isinstance(self.model_outputs, str):
+            raise Exception("Model Inputs and Outputs should be a Sequence. Not str.")
 
     def forward(self, batch_dict: NDict, *args, **kwargs) -> Dict:
         # convert input to the model's expected input
