### 2022-12-22 1e9c44d
samples_chacer write pipeline desc for cmp
