### 2022-12-23 BiomedSciAI/fuse-med-ml#229 @anupamajha1
The state of the field section compares with existing tools.
