### 2022-12-23 BiomedSciAI/fuse-med-ml#230 @anupamajha1
Thanks for adding the figure.
