### 2022-12-23 #4943 @anupamajha1
Thanks@alex-golts for addressing my comments. Here are some suggestions before I can finish my review checklist:

1. I installed FuseMedML and tried running it on the multiple examples you provided. The download instructions for datasets other than MNIST need to be improved. Currently, the READMEs for a lot of the examples simply point to the website with the data. You should include steps for download in all the example data READMEs.
2. The writing quality of the paper can be improved.
