+++ examples/fuse_examples/imaging/classification/knight/baseline/fuse_baseline_improve.py
@@ -7,7 +7,9 @@ import yaml
 import pandas as pd
 from fuse.dl.models import ModelMultiHead
 from fuse.dl.models.backbones.backbone_resnet_3d import BackboneResnet3D
-from fuse.dl.models.heads.heads_3D import Head3D
+# from fuse.dl.models.heads.heads_3D import Head3D
+from fuse.dl.models.heads.heads_3D import Head3DClassifier
+
 from fuseimg.datasets.knight import KNIGHT
 import torch.nn.functional as F
 import torch.nn as nn
@@ -29,8 +31,11 @@ from fuse.dl.lightning.pl_module import LightningModuleDefault
 import pytorch_lightning as pl
 from nnunet.network_architecture.generic_UNet import Generic_UNet
 from nnunet.network_architecture.initialization import InitWeights_He
-from torchvision.models import resnet50
-from fuse.dl.models.heads.heads_1D import Head1D
+# from torchvision.models import resnet50
+
+from resnet import resnet34, resnet18, resnet50
+
+# from fuse.dl.models.heads.heads_1D import Head1D
 ## Parameters:
 ##############################################################################
 # Data sources to use in model. Set {'imaging': True, 'clinical': False} for imaging only setting,
@@ -48,21 +53,29 @@ def make_model(use_data: dict, num_classes: int, imaging_dropout: float, fused_d
             conv_inputs = [("model.backbone_features", 2048)]
         elif not pretrained:
             backbone = BackboneResnet3D(in_channels=1)
+            state_dict = torch.load(open("/projects/msieve_dev3/usr/il018850/new_age_classification/base_resized/last.ckpt", "rb"))["state_dict"]
+            state_dict = {".".join(k.split(".")[2:]) : v for k, v in state_dict.items() if k.startswith("_model.backbone.")}
+            backbone.load_state_dict(state_dict)
             conv_inputs = [("model.backbone_features", 512)]
         else:
-            unet_path = "/data/usr/liam/nnUNet/3d_fullres/Task135_KiTS2021/nnUNetTrainerV2__nnUNetPlansv2.1/fold_0/model_final_checkpoint.model"
-            norm_op_kwargs = {'eps': 1e-05, 'affine': True}
-            dropout_op_kwargs = {'p': 0, 'inplace': True}
-            net_nonlin_kwargs = {'negative_slope': 0.01, 'inplace': True}
-            backbone = Generic_UNet(1, 32, 4, 5, 2, 2, nn.Conv3d, nn.InstanceNorm3d, norm_op_kwargs, nn.Dropout3d,dropout_op_kwargs,
-                        nn.LeakyReLU, net_nonlin_kwargs, True, False, None, InitWeights_He(1e-2),
-                        [[2, 2, 2], [2, 2, 2], [2, 2, 2], [2, 2, 2], [2, 2, 2]],
-                        [[3, 3, 3], [3, 3, 3], [3, 3, 3], [3, 3, 3], [3, 3, 3], [3, 3, 3]] 
-                        , True, True, True)
-            loaded_state_dict = torch.load(unet_path)["state_dict"]
-            backbone.load_state_dict(state_dict=loaded_state_dict)
-            backbone.inference_apply_nonlin = None
-            conv_inputs = [('model.backbone_features', 320)]
+            # unet_path = "/data/usr/liam/nnUNet/3d_fullres/Task135_KiTS2021/nnUNetTrainerV2__nnUNetPlansv2.1/fold_0/model_final_checkpoint.model"
+            # norm_op_kwargs = {'eps': 1e-05, 'affine': True}
[…]
