+++ fuseimg/datasets/knight.py
@@ -21,323 +21,38 @@ from functools import partial
 import torch
 import pandas as pd
 
-best_slices = [
-    312,
-    263,
-    172,
-    191,
-    43,
-    321,
-    47,
-    38,
-    131,
-    28,
-    12,
-    15,
-    23,
-    40,
-    112,
-    57,
-    58,
-    39,
-    68,
-    42,
-    29,
-    15,
-    336,
-    51,
-    38,
-    54,
-    183,
-    369,
-    31,
-    79,
-    19,
-    41,
-    55,
-    109,
-    47,
-    40,
-    54,
-    42,
-    23,
-    35,
-    67,
-    12,
-    198,
-    82,
-    33,
-    40,
-    63,
-    84,
-    28,
-    358,
-    23,
-    35,
-    157,
-    139,
-    31,
-    25,
-    31,
-    54,
-    28,
-    195,
-    57,
-    19,
-    35,
-    189,
-    28,
-    65,
-    189,
-    192,
-    297,
-    23,
-    31,
-    148,
-    68,
-    85,
-    30,
-    34,
-    30,
-    25,
-    205,
-    64,
-    26,
-    26,
-    30,
-    65,
-    65,
-    32,
-    52,
-    26,
-    50,
-    16,
-    48,
-    393,
-    30,
-    233,
-    24,
-    122,
-    412,
-    32,
-    192,
-    28,
-    192,
-    64,
-    149,
-    223,
-    46,
-    39,
-    34,
-    30,
-    34,
-    42,
-    20,
-    59,
-    84,
-    25,
-    179,
-    113,
-    215,
-    49,
-    374,
-    23,
-    217,
-    17,
-    22,
-    130,
-    150,
-    64,
-    45,
-    65,
-    66,
-    35,
-    33,
-    44,
-    258,
-    74,
-    21,
-    206,
-    32,
-    56,
-    53,
-    39,
-    140,
-    230,
-    169,
-    30,
-    63,
-    71,
-    162,
-    43,
-    10,
-    54,
-    54,
-    831,
-    17,
-    27,
-    354,
-    311,
-    411,
-    294,
-    291,
-    125,
-    101,
-    31,
-    60,
-    36,
-    32,
-    357,
-    54,
-    30,
-    28,
-    38,
-    226,
-    77,
-    36,
-    31,
-    47,
-    48,
-    35,
-    26,
-    33,
-    31,
-    32,
-    41,
-    35,
-    104,
-    73,
-    67,
-    70,
-    14,
-    163,
-    90,
-    72,
-    177,
-    81,
-    102,
-    80,
-    26,
-    76,
-    59,
-    145,
-    79,
-    28,
-    52,
-    126,
-    349,
-    43,
-    27,
-    40,
-    68,
-    42,
-    40,
-    24,
-    37,
-    45,
-    371,
-    210,
-    50,
-    286,
-    41,
-    77,
-    77,
-    70,
-    73,
-    28,
-    82,
-    29,
-    25,
-    49,
-    53,
-    22,
-    134,
-    30,
-    47,
-    24,
-    43,
-    59,
-    65,
-    47,
-    43,
-    35,
-    20,
-    65,
-    38,
-    15,
-    30,
-    35,
-    48,
-    271,
-    166,
-    51,
-    159,
-    124,
-    19,
-    35,
-    131,
-    122,
-    67,
-    55,
-    10,
-    62,
-    36,
-    38,
-    286,
-    40,
-    195,
-    162,
-    24,
-    34,
-    41,
-    79,
-    39,
-    35,
-    222,
-    339,
-    186,
-    54,
-    29,
-    220,
-    27,
-    49,
-    26,
-    27,
[…]
