+++ examples/fuse_examples/multimodality/ehr_transformer/dataset.py
@@ -1,0 +1,236 @@
+from typing import Sequence, Any
+import re
+import glob
+import os
+
+import numpy as np
+import pandas as pd
+from typing import Tuple
+
+from fuse.data import DatasetDefault, PipelineDefault, OpBase
+from fuse.data.ops.ops_read import OpReadDataframe
+#from fuse.data.ops.ops_common import OpCond, OpSet
+from fuse.data.utils.split import dataset_balanced_division_to_folds
+
+
+SOURCE = r'C:/D_Drive/Projects/EHR_Transformer/PhysioNet/predicting-mortality-of-icu-patients-the-physionetcomputing-in-cardiology-challenge-2012-1.0.0/predicting-mortality-of-icu-patients-the-physionet-computing-in-cardiology-challenge-2012-1.0.0'
+
+VALID_TESTS_ABOVE_ZERO = ['pH', 'Weight', 'Height', 'DiasABP', 'HR', 'NIMAP', 'MAP', 'NIDiasABP', 'NISysABP', 'PaCO2', 'PaO2', 'Temp', 'SaO2', 'RespRate', 'SysABP']
+
+STATIC_FIELDS = ['Age','Gender','Height','ICUType','Weight']
+
+
+class OpProcessTargetActivities(OpBase):
+    def __call__(self, sample_dict) -> Any:
+
+        sample_dict["activity.label"] = -1  # unknown
+
+        return sample_dict
+
+
+class OpProcessTargetActivities2(OpBase):
+    def __call__(self, sample_dict) -> Any:
+
+        sample_dict["activity.value"] = 1  # unknown
+
+        return sample_dict
+
+class PhysioNetCinC:
+
+    @staticmethod
+    def _read_raw_data(raw_data_path):
+        df = pd.DataFrame(columns=["PatientId", "Time", "Parameter", "Value"])
+        sub_sets = ["set-a", "set-b"]
+        for s in sub_sets:
+            csv_files = glob.glob(os.path.join(raw_data_path + '/' + s, "*.txt"))
+            for f in csv_files[0:5]:  #reducung the list temporarely for debugging
+                patient_id = os.path.splitext(os.path.basename(f))[0]
+                df_file = pd.read_csv(f)
+                df_file = df_file.drop(df_file[(df_file['Parameter'] == 'RecordID')].index).reset_index(drop=True)
+                df_file["PatientId"] = patient_id
+                df = df.append(df_file)
+
+        df.reset_index(inplace=True,drop=True)
+        return df
+
+    @staticmethod
+    def _drop_errors(df: pd.DataFrame) -> pd.DataFrame:
+        # drop records with measurements below or equal to zero for tests with only positive values
+        for v in VALID_TESTS_ABOVE_ZERO:
+            df = df.drop(df[(df['Parameter'] == v) & (df['Value'] <= 0)].index).reset_index(drop=True)
+
+        # drop gender values below zero
+        df = df.drop(df[(df['Parameter'] == 'Gender') & (df['Value'] < 0)].index).reset_index(drop=True)
+
+        # drop records with invalid Height and Weight values
+        df = df.drop(df[(df['Parameter'] == 'Weight') & (df['Value'] < 20)].index).reset_index(drop=True)
+        df = df.drop(df[(df['Parameter'] == 'Height') & (df['Value'] < 100)].index).reset_index(drop=True)
+        return df
+
+    @staticmethod
+    def _convert_time_to_datetime(df: pd.DataFrame) -> pd.DataFrame:
+        # dummy date is added for converting time to date time format
+        split = df['Time'].str.split(':', 1, True)
+        df['Year'] = "2020"
+        df['Month'] = "01"
[…]
