+++ examples/fuse_examples/multimodality/ehr_transformer/config_PhysioNetCinC.yaml
@@ -1,0 +1,125 @@
+dataset:
+  raw_data_path: 'C:/D_Drive/Projects/EHR_Transformer/PhysioNet/predicting-mortality-of-icu-patients-the-physionetcomputing-in-cardiology-challenge-2012-1.0.0/predicting-mortality-of-icu-patients-the-physionet-computing-in-cardiology-challenge-2012-1.0.0'
+  split_filename: None
+  num_folds: 5
+  train_folds: [ 0, 1, 2 ]
+  validation_folds: [ 3 ]
+  test_folds: [ 4 ]
+  seed: 2580
+  reset_split: True
+  num_percentiles : 4 #number of bins (percentiles) for converting floating lab/vital measurements to categorical values
+
+data:
+  #### Test/Val set defs
+  days_to_inddate: 60  #90 #None #180 #only visits before days_to_ind to index date (or after index date, if negative) are considered for test/val
+  days_to_inddate_start: null #only the visits between ind_date-days_to_inddate_start and ind_date-days_to_inddate are considered for test/val
+  event_prediction_window_days: null  #90
+  treatment_event_prediction_window_days: null  #60
+  #### Train set defs
+  # in case of train we take all visits (we want to train on as much data as we can), so days_to_inddate_tr and days_to_inddate_start_tr should be both null
+  days_to_inddate_tr: 60  #90 #None #180 #only visits before days_to_ind to index date (or after index date, if negative) are considered for train
+  days_to_inddate_start_tr: null #only the visits between days_to_inddate_start and days_to_inddate are considered for train
+  event_prediction_window_days_tr: null  #90
+
+  visit_days_resolution: 10 #consecutive visits that are less than visit_days_resolution days apart are merged into one visit.
+    #Note that merging stops only when number of days between consecutive visits is more than visit_days_resolution, so
+    #within a group of merged visits there may be some more than visit_days_resolution days apart.
+    #E.g. for visit days 1, 3, 5, 9, and for visit_days_resolution=2, days 1, 3, and 5 will be merged into a single visit, and day 10 will be a separate visit
+
+  limit_visits: null #210000 #null #50000000 #10000 #None #don't take too few (e.g. less than 10k), otherwise you'll get only a single visit per patient, which is not enough in case of next visit prediction
+  out_type: EVENT #EVENT or LABEL - determines the type of outcome GT, whether it is a list of events, or a binary label
+  task: CKD   #RC #"RC24" #"PD"#"BC"#'EP' #The task is only used output path naming and in db access (ehrtransformers.data_access.db2_data), to choose which dataset to read
+  subtask: ALL_SERV
+  task_type: outcome
+  output_name: admdate_visit
+  use_procedures: false #true #false
+  num_loader_workers: 1
+  input_column_names: [CCS, ]
+  min_visits_per_patient: 10 #patients with less visits are ignored
+
+  # When predicting next visit, this will map regular visit dictionary (codes to indices, index per code) to reduced
+  # visit dictionary (codes to indices, code groups mapped to (fewer) indices). When no mapping is used, this is null.
[…]
