+++ paper/paper.md
@@ -35,24 +35,23 @@ bibliography: paper.bib
 
 # Summary
 
-Machine Learning is at the forefront of scientific progress in Healthcare and Medicine. To accelerate scientific discovery, it is important to have tools that allow progress iterations to be collaborative, reproducible, reusable and easily built upon without "reinventing the wheel".  
+Machine Learning is at the forefront of scientific progress in Healthcare and Medicine. To accelerate scientific discovery, it is important to have tools that allow progress iterations to be collaborative, reproducible, reusable and easily built upon without "reinventing the wheel" for each task.  
 FuseMedML, or *fuse*, is a Python framework designed for accelerated Machine Learning (ML) based discovery in the medical domain. It is highly flexible and designed for easy collaboration, encouraging code reuse. Flexibility is enabled by a generic data object design where data is kept in a nested (hierarchical) Python dictionary (NDict), allowing to efficiently process and fuse information from multiple modalities. Functional components allow to specify input and output keys, to be read from and written to the nested dictionary.  
 Easy code reuse is enabled through key components implemented as standalone packages under the main *fuse* repo using the same design principles. These include *fuse.data* - a flexible data processing pipeline, *fuse.dl* - reusable Deep Learning (DL) model architecture components and loss functions, and *fuse.eval* - a library for evaluating ML models.  
 
 # Statement of need
-Medical related research projects span multiple modalities (e.g., imaging, clinical data, biochemical representations) and tasks (e.g., classification, segmentation, clinical conditions prediction).
-Through experience with many such projects we found three key challenges:
-1. Launching or implementing a new baseline model can take more time than it should. This is true even when very similar projects have already been done in the past by the same lab. 
-2. Porting individual components across projects is often painful, resulting in researchers “reinventing the wheel” time after time.
-3. Collaboration between projects across modalities as well as across domains such as imaging and molecules is very challenging.  
+Medical research often involves multiple modalities (e.g., imaging, clinical data, biochemical representations) and tasks (e.g., classification, segmentation, clinical condition prediction). In our experience working on numerous such projects, we have identified three key challenges:
+1. Setting up or implementing a new baseline model can be time-consuming, even when similar projects have already been completed by the same lab.
+2. Transferring individual components across projects can be difficult, leading to researchers frequently "reinventing the wheel."
+3. Collaborating between projects across modalities and domains, such as imaging and molecules, is often challenging.  
 
-FuseMedML was designed with the goal of alleviating these challenges.  
+To address these challenges, FuseMedML was developed with the goal of simplifying and streamlining medical research projects.  
 
 Before open sourcing it, we used *fuse* internally in multiple research projects [@raboh2022context], [@rabinovici2022early], [@rabinovici2022multimodal], [@jubran2021glimpse], [@tlusty2021pre], [@golts2022ensemble], [@radiology] and experienced significant improvement in development time, reusability and collaboration. 
 We were also able to meaningfully measure our progress and statistical significance of our results with off-the-shelf *fuse.eval* components that facilitate metrics' confidence interval calculation and model comparison. These tools have enabled us to organize two challenges as part of the 2022 International Symposium on Biomedical Imaging (ISBI) [@knight], [@bright].
 
 # State of the field
-FuseMedML is a comprehensive ML library with an emphasis on the biomedical domain. It provides a broad set of tools spanning the whole development pipeline, from data preparation, through model training to evaluation. It is built on top of leading ML frameworks such as PyTorch [@NEURIPS2019_bdbca288] and PyTorch Lightning [@Falcon_PyTorch_Lightning_2019] and complements them with additional flexible capabilities including domain specific ones.  
[…]
