+++ examples/fuse_examples/multimodality/ehr_transformer/dataset.py
@@ -11,6 +11,7 @@ from fuse.data import DatasetDefault, PipelineDefault, OpBase
 from fuse.data.ops.ops_read import OpReadDataframe
 #from fuse.data.ops.ops_common import OpCond, OpSet
 from fuse.data.utils.split import dataset_balanced_division_to_folds
+from ops_read_cinc import OpReadDataframeCinC
 
 
 SOURCE = r'C:/D_Drive/Projects/EHR_Transformer/PhysioNet/predicting-mortality-of-icu-patients-the-physionetcomputing-in-cardiology-challenge-2012-1.0.0/predicting-mortality-of-icu-patients-the-physionet-computing-in-cardiology-challenge-2012-1.0.0'
@@ -20,15 +21,22 @@ VALID_TESTS_ABOVE_ZERO = ['pH', 'Weight', 'Height', 'DiasABP', 'HR', 'NIMAP', 'M
 STATIC_FIELDS = ['Age','Gender','Height','ICUType','Weight']
 
 
-class OpProcessTargetActivities(OpBase):
+class OpAddBMI(OpBase):
     def __call__(self, sample_dict) -> Any:
 
-        sample_dict["activity.label"] = -1  # unknown
+        sample_dict["BMI"] = np.nan
 
+        if ("Height" in sample_dict.keys()) & ("Weight" in sample_dict.keys()):
+            height = sample_dict["Height"]
+            weight = sample_dict["Weight"]
+            if ~np.isnan(height) & ~np.isnan(weight):
+                sample_dict["BMI"] = 10000 * weight / (height*height)
+
+        print(sample_dict["BMI"])
         return sample_dict
 
 
-class OpProcessTargetActivities2(OpBase):
+class OpCorrectErrors(OpBase):
     def __call__(self, sample_dict) -> Any:
 
         sample_dict["activity.value"] = 1  # unknown
@@ -38,10 +46,11 @@ class OpProcessTargetActivities2(OpBase):
 class PhysioNetCinC:
 
     @staticmethod
-    def _read_raw_data(raw_data_path):
+    def _read_raw_data(raw_data_path:str)-> Tuple[pd.DataFrame,pd.DataFrame]:
+        #read patients info & tests
         df = pd.DataFrame(columns=["PatientId", "Time", "Parameter", "Value"])
-        sub_sets = ["set-a", "set-b"]
-        for s in sub_sets:
+        data_sub_sets = ["set-a", "set-b"]
+        for s in data_sub_sets:
             csv_files = glob.glob(os.path.join(raw_data_path + '/' + s, "*.txt"))
             for f in csv_files[0:5]:  #reducung the list temporarely for debugging
                 patient_id = os.path.splitext(os.path.basename(f))[0]
@@ -49,9 +58,18 @@ class PhysioNetCinC:
                 df_file = df_file.drop(df_file[(df_file['Parameter'] == 'RecordID')].index).reset_index(drop=True)
                 df_file["PatientId"] = patient_id
                 df = df.append(df_file)
+        df.reset_index(inplace=True, drop=True)
 
-        df.reset_index(inplace=True,drop=True)
-        return df
+        # read outcomes
+        df_outcomes = pd.DataFrame(columns=["RecordID", "In-hospital_death"])
+        outcomes = ['Outcomes-a.txt', 'Outcomes-b.txt']
+        for o in outcomes:
+            o_file = os.path.join(raw_data_path + '/' + o)
+            df_outcomes = df_outcomes.append(pd.read_csv(o_file)[["RecordID", "In-hospital_death"]]).reset_index(drop=True)
+        df_outcomes['RecordID'] = df_outcomes['RecordID'].astype(str)
[…]
