### 2022-12-26 #4943 @alex-golts
@anupamajha1 Thanks for these suggestions!
1. We added a bash script to the [KNIGHT example](https://github.com/BiomedSciAI/fuse-med-ml/tree/master/examples/fuse_examples/imaging/classification/knight) to download the data and edited the README file correspondingly.
The [STOIC21 example](https://github.com/BiomedSciAI/fuse-med-ml/tree/master/examples/fuse_examples/imaging/classification/stoic21) already contains in the code a class that downloads the data. For the remaining examples, obtaining the data involves some form of manual registration, so we couldn't easily automate that.

2. We revised the writing a bit, I hope it is somewhat better. you can see the changes in [this PR](https://github.com/BiomedSciAI/fuse-med-ml/pull/236). If you believe there are still any areas in the writing that could be improved upon, please feel free to let us know which ones specifically.
