+++ examples/fuse_examples/multimodality/ehr_transformer/dataset.py
@@ -1,3 +1,4 @@
+import pickle
 from typing import Sequence, Any
 import re
 import glob
@@ -12,6 +13,7 @@ from fuse.data.ops.ops_read import OpReadDataframe
 #from fuse.data.ops.ops_common import OpCond, OpSet
 from fuse.data.utils.split import dataset_balanced_division_to_folds
 from ops_read_cinc import OpReadDataframeCinC
+from fuse.data.utils.export import ExportDataset
 
 
 SOURCE = r'C:/D_Drive/Projects/EHR_Transformer/PhysioNet/predicting-mortality-of-icu-patients-the-physionetcomputing-in-cardiology-challenge-2012-1.0.0/predicting-mortality-of-icu-patients-the-physionet-computing-in-cardiology-challenge-2012-1.0.0'
@@ -33,14 +35,22 @@ class OpAddBMI(OpBase):
                 sample_dict["BMI"] = 10000 * weight / (height*height)
 
         print(sample_dict["BMI"])
+        print(sample_dict['Visits'].shape)
         return sample_dict
 
 
-class OpCorrectErrors(OpBase):
-    def __call__(self, sample_dict) -> Any:
+class OpCollectExamsStatistics(OpBase):
+    def __call__(self, sample_dict, percentiles:dict) -> Any:
 
-        sample_dict["activity.value"] = 1  # unknown
+        percentiles[sample_dict['PatientId']] = sample_dict['Age']
 
+        return sample_dict
+
+class OpMapToCategorical(OpBase):
+    def __call__(self, sample_dict, percentiles: dict) -> Any:
+
+        print(percentiles['HR'])
+        #print(percentiles.keys())
         return sample_dict
 
 class PhysioNetCinC:
@@ -59,6 +69,7 @@ class PhysioNetCinC:
                 df_file["PatientId"] = patient_id
                 df = df.append(df_file)
         df.reset_index(inplace=True, drop=True)
+        patient_ids = np.unique(df['PatientId'].values)
 
         # read outcomes
         df_outcomes = pd.DataFrame(columns=["RecordID", "In-hospital_death"])
@@ -69,7 +80,10 @@ class PhysioNetCinC:
         df_outcomes['RecordID'] = df_outcomes['RecordID'].astype(str)
         df_outcomes.rename(columns={'RecordID': 'PatientId'}, inplace=True)
 
-        return df, df_outcomes
+        #synchronize with patients data
+        df_outcomes = df_outcomes[df_outcomes['PatientId'].isin(patient_ids)]
+
+        return df, df_outcomes, patient_ids
 
     @staticmethod
     def _drop_errors(df: pd.DataFrame) -> pd.DataFrame:
@@ -116,37 +130,36 @@ class PhysioNetCinC:
             d_percentile[k] = np.percentile(d_values[k], percentiles)
 
         return d_percentile
-
-    @staticmethod
-    def _convert_to_patients_df(df: pd.DataFrame, df_outcomes: pd.DataFrame) -> Tuple[pd.DataFrame, dict]:
-        # dict of patients
-
-        statis_fields = ['Age', 'Height', 'Weight', 'Gender', 'ICUType', 'In-hospital_death']
-        df_patients = pd.DataFrame(columns=['PatientId', 'BMI']+statis_fields)
-        dict_patients_time_events = dict()
-        idx = 0
-        for pat_id, pat_records in df.groupby('PatientId'):
-            df_patients.loc[idx, 'PatientId'] = pat_id
-            df_static = pat_records[pat_records['Time'] == '00:00']
-            for f in statis_fields:
-                rec = df_static['Value'][(df_static['Parameter'] == f) & (df_static['Value'] >= 0)].reset_index(drop=True)
[…]
