+++ examples/fuse_examples/multimodality/ehr_transformer/dataset.py
@@ -10,17 +10,17 @@ from typing import Tuple
 
 from fuse.data import DatasetDefault, PipelineDefault, OpBase
 from fuse.data.ops.ops_read import OpReadDataframe
-#from fuse.data.ops.ops_common import OpCond, OpSet
+# from fuse.data.ops.ops_common import OpCond, OpSet
 from fuse.data.utils.split import dataset_balanced_division_to_folds
 from ops_read_cinc import OpReadDataframeCinC
 from fuse.data.utils.export import ExportDataset
 
-
 SOURCE = r'C:/D_Drive/Projects/EHR_Transformer/PhysioNet/predicting-mortality-of-icu-patients-the-physionetcomputing-in-cardiology-challenge-2012-1.0.0/predicting-mortality-of-icu-patients-the-physionet-computing-in-cardiology-challenge-2012-1.0.0'
 
-VALID_TESTS_ABOVE_ZERO = ['pH', 'Weight', 'Height', 'DiasABP', 'HR', 'NIMAP', 'MAP', 'NIDiasABP', 'NISysABP', 'PaCO2', 'PaO2', 'Temp', 'SaO2', 'RespRate', 'SysABP']
+VALID_TESTS_ABOVE_ZERO = ['pH', 'Weight', 'Height', 'DiasABP', 'HR', 'NIMAP', 'MAP', 'NIDiasABP', 'NISysABP', 'PaCO2',
+                          'PaO2', 'Temp', 'SaO2', 'RespRate', 'SysABP']
 
-STATIC_FIELDS = ['Age','Gender','Height','ICUType','Weight']
+STATIC_FIELDS = ['Age', 'Gender', 'Height', 'ICUType', 'Weight']
 
 
 class OpAddBMI(OpBase):
@@ -32,7 +32,7 @@ class OpAddBMI(OpBase):
             height = sample_dict["Height"]
             weight = sample_dict["Weight"]
             if ~np.isnan(height) & ~np.isnan(weight):
-                sample_dict["BMI"] = 10000 * weight / (height*height)
+                sample_dict["BMI"] = 10000 * weight / (height * height)
 
         print(sample_dict["BMI"])
         print(sample_dict['Visits'].shape)
@@ -40,29 +40,29 @@ class OpAddBMI(OpBase):
 
 
 class OpCollectExamsStatistics(OpBase):
-    def __call__(self, sample_dict, percentiles:dict) -> Any:
-
+    def __call__(self, sample_dict, percentiles: dict) -> Any:
         percentiles[sample_dict['PatientId']] = sample_dict['Age']
 
         return sample_dict
 
+
 class OpMapToCategorical(OpBase):
     def __call__(self, sample_dict, percentiles: dict) -> Any:
-
         print(percentiles['HR'])
-        #print(percentiles.keys())
+        # print(percentiles.keys())
         return sample_dict
+
 
 class PhysioNetCinC:
 
     @staticmethod
-    def _read_raw_data(raw_data_path:str)-> Tuple[pd.DataFrame,pd.DataFrame]:
-        #read patients info & tests
+    def _read_raw_data(raw_data_path: str) -> Tuple[pd.DataFrame, pd.DataFrame]:
+        # read patients info & tests
         df = pd.DataFrame(columns=["PatientId", "Time", "Parameter", "Value"])
         data_sub_sets = ["set-a", "set-b"]
         for s in data_sub_sets:
             csv_files = glob.glob(os.path.join(raw_data_path + '/' + s, "*.txt"))
-            for f in csv_files[0:5]:  #reducung the list temporarely for debugging
+            for f in csv_files[0:5]:  # reducung the list temporarely for debugging
[…]
