### 2022-12-27 #4943 @suragnair
Some more suggestions for improving the paper further. Firstly, it would be helpful to have a "design philosophy" focused figure that outlines `NDict` and input/output of fuse models (can be borrowed from the video in the `README`).

It would help to make the following improvements:
- Figure 1: It is not immediately apparent how `data.img` in `updated_sample` maps to `img` in sample, and so on. Adding arrows connecting them would be useful, like in the video.
- Figure 2: Why does 1 need to be specified in `('data.input.img', 1)`?
- Figure 3: EvaluatorDefault needs some explanation in the caption.
