### 2022-12-28 9019a29
auc metric rewrite + check the predictions are post softmax
