### 2023-01-01 47c41e5
support models norm in ensemble
