+++ fuse/data/datasets/caching/samples_cacher.py
@@ -122,15 +122,15 @@ class SamplesCacher:
                 if not os.path.isdir(found_dir):
                     continue
                 if os.path.basename(found_dir) != self._pipeline_desc_hash:
+                    new_desc = self._pipeline_desc_text
+                    new_file = os.path.join(found_dir, f'pipeline_{self._pipeline_desc_hash}_desc.txt')
+                    with open(new_file, 'wt') as f:
+                        f.write(new_desc)
+
                     pipeline_desc_file = os.path.join(found_dir, f'pipeline_{os.path.basename(found_dir)}_desc.txt')
                     if os.path.exists(pipeline_desc_file):
-                        print("*** Old pipeline description", new_file)
-                        new_desc = self._pipeline_desc_text
-                        new_file = os.path.join(found_dir, f'pipeline_{self._pipeline_desc_hash}_desc.txt')
-                        with open(new_file, 'wt') as f:
-                            f.write(new_desc)
-
-                        print("*** New pipeline description", new_file)
+                        print("*** Old pipeline description:", pipeline_desc_file)
+                        print("*** New pipeline description (does not match old pipeline):", new_file)
 
                     raise Exception(
                         f"Found samples cache for pipeline hash {os.path.basename(found_dir)} which is different from the current loaded pipeline hash {self._pipeline_desc_hash} !!\n"
