### 2023-01-01 71c3240
fix learning rate and added config options
