+++ fuseimg/datasets/knight.py
@@ -21,6 +21,19 @@ from functools import partial
 import torch
 import pandas as pd
 
+class OpNormalizeAgainstSelf(OpBase):
+
+    def __init__(self, **kwargs):
+        super().__init__(**kwargs)
+
+    def __call__(self, sample_dict: NDict, key: str):
+        img = sample_dict[key]
+        img = np.clip(img, *(np.percentile(img, [5,95]))) # truncate the intensities to the range of 0.5 to 99.5 percentiles 
+
+        img = (img - img.mean())/img.std()
+        sample_dict[key] = img
+
+        return sample_dict
 
 class OpKnightSampleIDDecode(OpBase):
     """
@@ -187,6 +200,8 @@ class KNIGHT:
                 # loading data
                 (OpLoadImage(data_path), dict(key_in="data.input.img_path", key_out="data.input.img", format="nib")),
                 # fixed image normalization
+                # (OpNormalizeAgainstSelf(), dict(key="data.input.img")),
+
                 (OpClip(), dict(key="data.input.img", clip=(-62, 301))),
                 (OpZScoreNorm(), dict(key="data.input.img", mean=104.0, std=75.3)),  # kits normalization
                 # transposing so the depth channel will be first
@@ -228,7 +243,7 @@ class KNIGHT:
                         translate=(RandInt(-15, 15), RandInt(-15, 15)),
                     ),
                 ),
-                (OpAugGaussian(), dict(key="data.input.img", std=0.01)),
+                (OpRandApply(OpAugGaussian(), 0.3), dict(key="data.input.img", std=0.01)),
                 # add channel dimension -> [C=1, D, H, W]
                 (OpLambda(partial(torch.unsqueeze, dim=0)), dict(key="data.input.img")),
             ],
