+++ fuseimg/datasets/knight.py
@@ -21,19 +21,22 @@ from functools import partial
 import torch
 import pandas as pd
 
-class OpNormalizeAgainstSelf(OpBase):
 
+class OpNormalizeAgainstSelf(OpBase):
     def __init__(self, **kwargs):
         super().__init__(**kwargs)
 
     def __call__(self, sample_dict: NDict, key: str):
         img = sample_dict[key]
-        img = np.clip(img, *(np.percentile(img, [5,95]))) # truncate the intensities to the range of 0.5 to 99.5 percentiles 
+        img = np.clip(
+            img, *(np.percentile(img, [5, 95]))
+        )  # truncate the intensities to the range of 0.5 to 99.5 percentiles
 
-        img = (img - img.mean())/img.std()
+        img = (img - img.mean()) / img.std()
         sample_dict[key] = img
 
         return sample_dict
+
 
 class OpKnightSampleIDDecode(OpBase):
     """
@@ -201,7 +204,6 @@ class KNIGHT:
                 (OpLoadImage(data_path), dict(key_in="data.input.img_path", key_out="data.input.img", format="nib")),
                 # fixed image normalization
                 # (OpNormalizeAgainstSelf(), dict(key="data.input.img")),
-
                 (OpClip(), dict(key="data.input.img", clip=(-62, 301))),
                 (OpZScoreNorm(), dict(key="data.input.img", mean=104.0, std=75.3)),  # kits normalization
                 # transposing so the depth channel will be first
