+++ examples/fuse_examples/imaging/classification/knight/baseline/fuse_baseline_improve.py
@@ -9,6 +9,7 @@ from fuse.dl.models import ModelMultiHead
 from fuse.dl.models.backbones.backbone_resnet_3d import BackboneResnet3D
 # from fuse.dl.models.heads.heads_3D import Head3D
 from fuse.dl.models.heads.heads_3D import Head3DClassifier
+from fuse.dl.models.heads.head_1D_classifier import Head1DClassifier
 
 from fuseimg.datasets.knight import KNIGHT
 import torch.nn.functional as F
@@ -53,29 +54,29 @@ def make_model(use_data: dict, num_classes: int, imaging_dropout: float, fused_d
             conv_inputs = [("model.backbone_features", 2048)]
         elif not pretrained:
             backbone = BackboneResnet3D(in_channels=1)
-            state_dict = torch.load(open("/projects/msieve_dev3/usr/il018850/new_age_classification/base_resized/last.ckpt", "rb"))["state_dict"]
+            state_dict = torch.load(open("/projects/msieve_dev3/usr/il018850/new_age_classification/base_all_resized/last.ckpt", "rb"))["state_dict"]
             state_dict = {".".join(k.split(".")[2:]) : v for k, v in state_dict.items() if k.startswith("_model.backbone.")}
             backbone.load_state_dict(state_dict)
             conv_inputs = [("model.backbone_features", 512)]
         else:
-            # unet_path = "/data/usr/liam/nnUNet/3d_fullres/Task135_KiTS2021/nnUNetTrainerV2__nnUNetPlansv2.1/fold_0/model_final_checkpoint.model"
-            # norm_op_kwargs = {'eps': 1e-05, 'affine': True}
-            # dropout_op_kwargs = {'p': 0, 'inplace': True}
-            # net_nonlin_kwargs = {'negative_slope': 0.01, 'inplace': True}
-            # backbone = Generic_UNet(1, 32, 4, 5, 2, 2, nn.Conv3d, nn.InstanceNorm3d, norm_op_kwargs, nn.Dropout3d,dropout_op_kwargs,
-            #             nn.LeakyReLU, net_nonlin_kwargs, True, False, None, InitWeights_He(1e-2),
-            #             [[2, 2, 2], [2, 2, 2], [2, 2, 2], [2, 2, 2], [2, 2, 2]],
-            #             [[3, 3, 3], [3, 3, 3], [3, 3, 3], [3, 3, 3], [3, 3, 3], [3, 3, 3]] 
-            #             , True, True, True)
-            # loaded_state_dict = torch.load(unet_path)["state_dict"]
-            # backbone.load_state_dict(state_dict=loaded_state_dict)
-            # backbone.inference_apply_nonlin = None
-            # conv_inputs = [('model.backbone_features', 320)]
-            state_dict = torch.load(open("/data/usr/liam/age_classification/pretrain/resnet_50_23dataset.pth", "rb"))["state_dict"]
-            state_dict = {k[7:]:v for k,v in state_dict.items()}
-            backbone = resnet50()#(shortcut_type='A')
-            backbone.load_state_dict(state_dict=state_dict)
-            conv_inputs = [("model.backbone_features", 2048)]
+            unet_path = "/data/usr/liam/nnUNet/3d_fullres/Task135_KiTS2021/nnUNetTrainerV2__nnUNetPlansv2.1/fold_0/model_final_checkpoint.model"
+            norm_op_kwargs = {'eps': 1e-05, 'affine': True}
+            dropout_op_kwargs = {'p': 0, 'inplace': True}
+            net_nonlin_kwargs = {'negative_slope': 0.01, 'inplace': True}
[…]
