+++ examples/fuse_examples/imaging/classification/knight/baseline/config.yaml
@@ -1,20 +1,20 @@
-exp_name : "testing-pretrained-med3d"
-experiment_num : 0 # 0 - self-norm | 2 - regular
+exp_name : "nnunet_pretrained-im&cl"
+experiment_num : 1 # 0 - self-norm | 2 - regular
 task_num : task_1 # task_1 or task_2 
 num_gpus : 1
-use_data : {"imaging": True, "clinical": True} 
-batch_size : 3
+use_data : {"imaging": True, "clinical": True}
+batch_size : 2
 resize_to : [70, 256, 256]
-num_epochs : 120
-learning_rate : 0.00005
+num_epochs : 100
+learning_rate : 0.0002
 imaging_dropout : 0.5
 fused_dropout : 0.5
-pretrained : True
-regression_head : False
+pretrained : False
+regression_head : True
 two_dim : False
-reset_cache : True
-gpus : [3]
-testing : False # for testing\continous integration
+reset_cache : False
+gpus : [5]
+testing : True # for testing\continous integration
 
 task_1:
   num_classes : 2
