+++ fuse/eval/metrics/metrics_common.py
@@ -228,10 +228,10 @@ class MetricCollector(MetricBase):
             # convert required ids to permutation
             original_ids = self._collected_ids
             required_ids = ids
-            original_ids_pos = {s : i  for (i,s) in enumerate(original_ids)}
+            original_ids_pos = {s: i for (i, s) in enumerate(original_ids)}
 
-            # permutation_old = [original_ids.index(sample_id) for sample_id in required_ids]  #16.399551391601562 
-            permutation = [original_ids_pos[sample_id] for sample_id in required_ids] #0.008132457733154297 seconds
+            # permutation_old = [original_ids.index(sample_id) for sample_id in required_ids]  #16.399551391601562
+            permutation = [original_ids_pos[sample_id] for sample_id in required_ids]  # 0.008132457733154297 seconds
 
             # create the permuted dictionary
             data = {}
