+++ fuse/dl/models/heads/__init__.py
@@ -1,5 +1,6 @@
 from .common import ClassifierFCN, ClassifierFCN3D, ClassifierMLP
 from .head_1D_classifier import Head1DClassifier
+from .heads_1D import Head1D
 from .heads_3D import Head3DClassifier
 from .head_dense_segmentation import HeadDenseSegmentation
 from .head_global_pooling_classifier import HeadGlobalPoolingClassifier
