+++ examples/fuse_examples/imaging/classification/knight/baseline/config.yaml
@@ -1,19 +1,19 @@
-exp_name : "nnunet_pretrained-im&cl"
-experiment_num : 1 # 0 - self-norm | 2 - regular
+exp_name : base-40slices
+experiment_num : 1 # 0 - [70, 256, 256] 1 - [40, 256, 256]
 task_num : task_1 # task_1 or task_2 
 num_gpus : 1
 use_data : {"imaging": True, "clinical": True}
-batch_size : 2
-resize_to : [70, 256, 256]
+batch_size : 8
+resize_to : [40, 256, 256]
 num_epochs : 100
-learning_rate : 0.0002
+learning_rate : 0.0004
 imaging_dropout : 0.5
 fused_dropout : 0.5
 pretrained : False
-regression_head : True
+regression_head : False
 two_dim : False
 reset_cache : False
-gpus : [5]
+gpus : [0]
 testing : True # for testing\continous integration
 
 task_1:
