+++ examples/fuse_examples/imaging/classification/knight/baseline/config.yaml
@@ -2,18 +2,18 @@ experiment_num : 0
 task_num : task_1 # task_1 or task_2
 num_gpus : 1
 use_data : {"imaging": True, "clinical": True}
-batch_size : 6
+batch_size : 6    # need to adjust barch size to your GPU available memory
 resize_to : [80, 256, 256]
 num_epochs : 100
 learning_rate : 0.0003
 imaging_dropout : 0.5
 fused_dropout : 0.5
-testing : False # for testing\continous integration
+testing : False   # for testing\continous integration
 reset_cache : False
 
 task_1:
   num_classes : 2
-  class_names: None
+  class_names: False
   target_name : "data.gt.gt_global.task_1_label"
   target_metric : "validation.metrics.auc"
 
