### 2023-01-02 9e27023
Simplified mnist example dramatically. Removed testing and eval stages, and model checkpointing.
