+++ fuse/dl/lightning/pl_module.py
@@ -15,11 +15,13 @@ limitations under the License.
 
 Created on June 30, 2021
 """
+from typing import Sequence, Dict, Callable
 
 import pytorch_lightning as pl
 from typing import Optional
 
 from fuse.dl.lightning.pl_funcs import *  # noqa
+from fuse.dl.models.model_wrapper import ModelWrapSeqToDict
 
 
 class LightningModuleDefault(pl.LightningModule):
@@ -169,3 +171,33 @@ class LightningModuleDefault(pl.LightningModule):
     def set_predictions_keys(self, keys: List[str]) -> None:
         """Define which keys to extract from batch_dict on prediction mode"""
         self._prediction_keys = keys
+
+
+class FuseModule(LightningModuleDefault, ModelWrapSeqToDict):
+
+    def __init__(
+        self,
+        # mutual args
+        model: torch.nn.Module = None,
+
+        # ModelWrapSeqToDict args
+        *,  # preventing positional args
+        model_inputs: Sequence[str] = None,
+        model_outputs: Sequence[str] = None,
+        pre_forward_processing_function: Callable = None,
+        post_forward_processing_function: Callable = None,
+        
+        # LightningModuleDefault args
+        model_dir: str,
+        losses: Optional[Dict[str, LossBase]] = None,
+        train_metrics: Optional[OrderedDict[str, MetricBase]] = None,
+        validation_metrics: Optional[OrderedDict[str, MetricBase]] = None,
+        test_metrics: Optional[OrderedDict[str, MetricBase]] = None,
+        optimizers_and_lr_schs: Any = None,
+        callbacks: Optional[Sequence[pl.Callback]] = None,
+        best_epoch_source: Optional[Union[Dict, List[Dict]]] = None,
+        save_hyperparameters: Optional[List[str]] = None,
+        tensorboard_sep: str = ".",
+        **kwargs: dict,
+    ) -> None:
+        pass
