### 2023-01-02 e58db33
return upper bound for pytochvision
