+++ fuse/dl/lightning/pl_module.py
@@ -15,13 +15,12 @@ limitations under the License.
 
 Created on June 30, 2021
 """
-from typing import Sequence, Dict, Callable
+from typing import Sequence, Dict
 
 import pytorch_lightning as pl
 from typing import Optional
 
 from fuse.dl.lightning.pl_funcs import *  # noqa
-from fuse.dl.models.model_wrapper import ModelWrapSeqToDict
 
 
 class LightningModuleDefault(pl.LightningModule):
