+++ examples/fuse_examples/multimodality/ehr_transformer/dataset.py
@@ -39,16 +39,62 @@ class OpAddBMI(OpBase):
         return sample_dict
 
 
-class OpCollectExamsStatistics(OpBase):
-    def __call__(self, sample_dict, percentiles: dict) -> Any:
-        percentiles[sample_dict['PatientId']] = sample_dict['Age']
-
-        return sample_dict
+# class OpCollectExamsStatistics(OpBase):
+#     def __call__(self, sample_dict, percentiles: dict) -> Any:
+#         percentiles[sample_dict['PatientId']] = sample_dict['Age']
+#
+#         return sample_dict
 
 
 class OpMapToCategorical(OpBase):
+    @staticmethod
+    def _digitize(sample_dict, percentiles):
+        #  sample_dict_mapped = np.deepcopy(sample_dict)
+        #
+        #  categorical_static = ['Gender', 'ICUType']
+        #  #fix categorical in dictionary not in handling
+        #  percentiles['Gender'] = [0,1]
+        #  percentiles['MechVent'] = [0, 1]
+        #
+        #
+        #  # mapping static clinical characteristics
+        #  for k in sample_dict.keys():
+        #      if k in categorical:
+        #          sample_dict_mapped[k] = k + '_' + sample_dict[k]
+        #      else: sample_dict_mapped[k] = np.digitize(sample_dict[k], percentiles[k])
+        #
+        #  df_visits = sample_dict['Visits']
+        #
+        #
+        #     list_remaining = [x for x in list_params if x not in special_handling]
+        #     for p in list_remaining:
+        #         new_values = np.digitize(d_values[p], d_percentiles[p])
+        #         new_deltas = np.digitize(d_delta[p], d_percentiles_deltas[p])
+        #         mapped_values = [p + '_' + str(x) for x in new_values]
+        #         mapped_deltas = [p + '_D_' + str(x) for x in new_deltas]
+        #         df_mapped.loc[d_values[p].index, 'MappedValue'] = mapped_values
+        #         df_mapped.loc[d_values[p].index, 'MappedDelta'] = mapped_deltas
+        #
+        #     if USE_DELTAS_FOR_MEASUREMETS:
+        #         df_mapped['FinalMappedValue'] = np.where(df_mapped['FirstIndicator'], df_mapped['MappedValue'],
+        #                                                  df_mapped['MappedDelta'])
+        #     else:
+        #         df_mapped['FinalMappedValue'] = df_mapped['MappedValue']
+        #
+        #     # remove records with patient ID
+        #     df_mapped = df_mapped.drop(df_mapped[(df_mapped['Parameter'] == 'RecordID')].index).reset_index(drop=True)
+        #
+        #     with open(RESULTS + 'mapped_data.pkl', 'bw') as f:
+        #         pickle.dump(df_mapped, f)
+        # else:
+        #     with open(RESULTS + 'mapped_data.pkl', 'rb') as f:
+        #         df_mapped = pickle.load(f)
+
+        return sample_dict
+
     def __call__(self, sample_dict, percentiles: dict) -> Any:
         print(percentiles['HR'])
+        sample_dict = OpMapToCategorical._digitize(sample_dict, percentiles)
         # print(percentiles.keys())
         return sample_dict
 
@@ -117,18 +163,38 @@ class PhysioNetCinC:
         return df
 
     @staticmethod
-    def _generate_percentiles(df: pd.DataFrame, num_percentiles: int) -> dict:
-        percentiles = range(0, 100 + int(100 / num_percentiles), int(100 / num_percentiles))
+    def _generate_percentiles(dataset: DatasetDefault, num_percentiles: int) -> dict:
 
-        list_params = np.unique(df[['Parameter']])
+        # calculate statistics of train set only and generate dictionary of percentiles for mapping
[…]
