+++ examples/fuse_examples/multimodality/ehr_transformer/dataset.py
@@ -26,16 +26,17 @@ STATIC_FIELDS = ['Age', 'Gender', 'Height', 'ICUType', 'Weight']
 class OpAddBMI(OpBase):
     def __call__(self, sample_dict) -> Any:
 
-        sample_dict["BMI"] = np.nan
+        d_static = sample_dict['StaticDetails']
+        d_static["BMI"] = np.nan
 
-        if ("Height" in sample_dict.keys()) & ("Weight" in sample_dict.keys()):
-            height = sample_dict["Height"]
-            weight = sample_dict["Weight"]
+        if ("Height" in d_static.keys()) & ("Weight" in d_static.keys()):
+            height = d_static["Height"]
+            weight = d_static["Weight"]
             if ~np.isnan(height) & ~np.isnan(weight):
-                sample_dict["BMI"] = 10000 * weight / (height * height)
+                d_static["BMI"] = 10000 * weight / (height * height)
 
-        print(sample_dict["BMI"])
-        print(sample_dict['Visits'].shape)
+        print(d_static["BMI"])
+        sample_dict['StaticDetails'] = d_static
         return sample_dict
 
 
@@ -48,53 +49,21 @@ class OpAddBMI(OpBase):
 
 class OpMapToCategorical(OpBase):
     @staticmethod
-    def _digitize(sample_dict, percentiles):
-        #  sample_dict_mapped = np.deepcopy(sample_dict)
-        #
-        #  categorical_static = ['Gender', 'ICUType']
-        #  #fix categorical in dictionary not in handling
-        #  percentiles['Gender'] = [0,1]
-        #  percentiles['MechVent'] = [0, 1]
-        #
-        #
-        #  # mapping static clinical characteristics
-        #  for k in sample_dict.keys():
-        #      if k in categorical:
-        #          sample_dict_mapped[k] = k + '_' + sample_dict[k]
-        #      else: sample_dict_mapped[k] = np.digitize(sample_dict[k], percentiles[k])
-        #
-        #  df_visits = sample_dict['Visits']
-        #
-        #
-        #     list_remaining = [x for x in list_params if x not in special_handling]
-        #     for p in list_remaining:
-        #         new_values = np.digitize(d_values[p], d_percentiles[p])
-        #         new_deltas = np.digitize(d_delta[p], d_percentiles_deltas[p])
-        #         mapped_values = [p + '_' + str(x) for x in new_values]
-        #         mapped_deltas = [p + '_D_' + str(x) for x in new_deltas]
-        #         df_mapped.loc[d_values[p].index, 'MappedValue'] = mapped_values
-        #         df_mapped.loc[d_values[p].index, 'MappedDelta'] = mapped_deltas
-        #
-        #     if USE_DELTAS_FOR_MEASUREMETS:
-        #         df_mapped['FinalMappedValue'] = np.where(df_mapped['FirstIndicator'], df_mapped['MappedValue'],
-        #                                                  df_mapped['MappedDelta'])
-        #     else:
-        #         df_mapped['FinalMappedValue'] = df_mapped['MappedValue']
-        #
-        #     # remove records with patient ID
-        #     df_mapped = df_mapped.drop(df_mapped[(df_mapped['Parameter'] == 'RecordID')].index).reset_index(drop=True)
-        #
-        #     with open(RESULTS + 'mapped_data.pkl', 'bw') as f:
-        #         pickle.dump(df_mapped, f)
-        # else:
-        #     with open(RESULTS + 'mapped_data.pkl', 'rb') as f:
-        #         df_mapped = pickle.load(f)
+    def _digitize_values(sample_dict, percentiles):
+        # mapping static clinical characteristics
+        for k in sample_dict['StaticDetails'].keys():
+            sample_dict['StaticDetails'][k] = k + '_' + \
[…]
