+++ examples/fuse_examples/tests/test_classification_knight.py
@@ -86,6 +86,7 @@ class KnightTestTestCase(unittest.TestCase):
 
         task_1:
             num_classes : 2
+            class_names: False
             target_name : "data.gt.gt_global.task_1_label"
             target_metric : "validation.metrics.auc"
         """
