### 2023-01-05 3c7bbc7
fix knight testing
