+++ examples/fuse_examples/tests/test_classification_knight.py
@@ -83,6 +83,7 @@ class KnightTestTestCase(unittest.TestCase):
         imaging_dropout : 0.5
         fused_dropout : 0.5
         testing : True
+        reset_cache : False
 
         task_1:
             num_classes : 2
