### 2023-01-05 d3441cc
fix learning rate and added config options (#238)
