### 2023-01-05 64671e1
fixed bugs (running large dataset) and added dropping of short patients
