+++ examples/fuse_examples/multimodality/ehr_transformer/dataset.py
@@ -27,7 +27,6 @@ class OpAddBMI(OpBase):
     def __call__(self, sample_dict) -> Any:
 
         d_static = sample_dict['StaticDetails']
-        d_static["BMI"] = np.nan
 
         if ("Height" in d_static.keys()) & ("Weight" in d_static.keys()):
             height = d_static["Height"]
@@ -35,7 +34,6 @@ class OpAddBMI(OpBase):
             if ~np.isnan(height) & ~np.isnan(weight):
                 d_static["BMI"] = 10000 * weight / (height * height)
 
-        print(d_static["BMI"])
         sample_dict['StaticDetails'] = d_static
         return sample_dict
 
@@ -48,23 +46,20 @@ class OpAddBMI(OpBase):
 
 
 class OpMapToCategorical(OpBase):
-    @staticmethod
-    def _digitize_values(sample_dict, percentiles):
+
+    def __call__(self, sample_dict, percentiles: dict) -> Any:
+
         # mapping static clinical characteristics
         for k in sample_dict['StaticDetails'].keys():
             sample_dict['StaticDetails'][k] = k + '_' + \
                                               str(np.digitize(sample_dict['StaticDetails'][k], percentiles[k]))
 
-        sample_dict['Visits']['Value'] = sample_dict['Visits'].\
-            apply(lambda row: row['Parameter'] + '_' + str(np.digitize(row['Value'],percentiles[row['Parameter']])),\
-                  axis=1)
+        if not sample_dict['Visits'].empty:
+            sample_dict['Visits']['Value'] = sample_dict['Visits']. \
+                apply(
+                lambda row: row['Parameter'] + '_' + str(np.digitize(row['Value'], percentiles[row['Parameter']])), \
+                axis=1)
 
-        return sample_dict
-
-    def __call__(self, sample_dict, percentiles: dict) -> Any:
-        print(percentiles['HR'])
-        sample_dict = OpMapToCategorical._digitize_values(sample_dict, percentiles)
-        # print(percentiles.keys())
         return sample_dict
 
 
@@ -77,7 +72,7 @@ class PhysioNetCinC:
         data_sub_sets = ["set-a", "set-b"]
         for s in data_sub_sets:
             csv_files = glob.glob(os.path.join(raw_data_path + '/' + s, "*.txt"))
-            for f in csv_files[0:5]:  # reducung the list temporarely for debugging
+            for f in csv_files:  # reducung the list temporarely for debugging
                 patient_id = os.path.splitext(os.path.basename(f))[0]
                 df_file = pd.read_csv(f)
                 df_file = df_file.drop(df_file[(df_file['Parameter'] == 'RecordID')].index).reset_index(drop=True)
@@ -99,10 +94,10 @@ class PhysioNetCinC:
         # synchronize with patients data
         df_outcomes = df_outcomes[df_outcomes['PatientId'].isin(patient_ids)]
 
-        return df, df_outcomes, patient_ids
+        return df, df_outcomes
 
     @staticmethod
-    def _drop_errors(df: pd.DataFrame) -> pd.DataFrame:
+    def _drop_records_with_errors(df: pd.DataFrame) -> pd.DataFrame:
         # drop records with measurements below or equal to zero for tests with only positive values
         for v in VALID_TESTS_ABOVE_ZERO:
             df = df.drop(df[(df['Parameter'] == v) & (df['Value'] <= 0)].index).reset_index(drop=True)
@@ -114,6 +109,18 @@ class PhysioNetCinC:
         df = df.drop(df[(df['Parameter'] == 'Weight') & (df['Value'] < 20)].index).reset_index(drop=True)
         df = df.drop(df[(df['Parameter'] == 'Height') & (df['Value'] < 100)].index).reset_index(drop=True)
         return df
+
+    @staticmethod
[…]
