+++ examples/fuse_examples/multimodality/ehr_transformer/config_PhysioNetCinC.yaml
@@ -9,6 +9,8 @@ dataset:
   seed: 2580
   reset_split: True
   num_percentiles : 4 #number of bins (percentiles) for converting floating lab/vital measurements to categorical values
+  categorical_max_num_of_values : 5 #max number of uniq values for categorical variable for not to be digitized
+  min_hours_in_hospital: 46
 
 data:
   #### Test/Val set defs
