+++ paper/paper.md
@@ -52,8 +52,10 @@ We were also able to meaningfully measure our progress and statistical significa
 
 # State of the field
 FuseMedML is a comprehensive machine learning library that focuses on the biomedical domain. It offers a range of tools covering the entire development process, including data preparation, model training, and evaluation. Built on top of popular machine learning frameworks such as PyTorch [@NEURIPS2019_bdbca288] and PyTorch Lightning [@Falcon_PyTorch_Lightning_2019], FuseMedML also includes flexible domain-specific capabilities to complement these frameworks. Overall, FuseMedML aims to facilitate machine learning discoveries within the healthcare and life science sectors.
-One way in which *fuse* can complement PyTorch is through its generic design concept of storing arbitrary types of data in a specialized nested dictionary. This is a key driver of flexibility, allowing minimal code modifications when moving building blocks between different projects. Concretely, *fuse* has a dataset class that extends the PyTorch dataset, and a model wrapper class that enables PyTorch models to operate on `batch_dict`s rather than tensors.  
+One way in which *fuse* can complement PyTorch is through its generic design concept (See \autoref{fig:fuse_design}) of storing arbitrary types of data in a specialized nested dictionary. This is a key driver of flexibility, allowing minimal code modifications when moving building blocks between different projects. Concretely, *fuse* has a dataset class that extends the PyTorch dataset, and a model wrapper class that enables PyTorch models to operate on `batch_dict`s rather than tensors.  
 In the case of PyTorch Lightning, *fuse* integrates with it directly as it builds upon its comprehensive trainer class, also allowing users to define their models and data modules in PyTorch Lightning style, with flexible levels of customizability.
+
+![This figure illustrates FuseMedML's design concept. A *fuse* component is instantiated with input and output keys. These keys refer to the `sample_dict`, the basic data sample structure of *fuse* represented by a special nested Python dictionary called "NDict".\label{fig:fuse_design}](figures/fuse_design.png){width=100%}
 
 There are existing PyTorch-based ML libraries that similarly to *fuse* cater to researchers in the biomedical domain. Two examples of such prominent libraries are MONAI [@Cardoso_MONAI_An_open-source_2022] and PyHealth [@DBLP:journals/corr/abs-2101-04209]. MONAI is primarily focused on medical imaging applications. PyHealth on the other hand mainly focuses on health records data. *fuse* is designed to support different types of medical data and multimodal use cases involving imaging, clinical and biochemical data.  
 As with generic ML frameworks like PyTorch and PyTorch Lightning, *fuse* can also coexist with the more specific libraries like MONAI, PyHealth or others. A user may opt to borrow parts from different libraries and complement them with components from *fuse*. As another example, a user may want to use the data *ops* of *fuse* which are generic and flexible, or its data caching mechanism, which allows to separate processing into a static and dynamic pipelines, controlling the desired stages to be cached.  
