+++ fuse/dl/optimizations/activation_checkpointing.py
@@ -1,0 +1,425 @@
+from torch import nn
+#from torch.utils.checkpoint import checkpoint
+import torch
+
+def custom_forward_and_backward(config, model, x, targets, regularization_terms,
+                                        do_backward=False,
+                                        use_cpu_ram_paging=False,
+                                        calc_loss=True):   
+    if do_backward and not calc_loss:
+        raise Exception('calculate_loss cannot be disabled when do_backward is enabled.')
+
+    sample_extractor = SampleExtractor([x], move_to_gpu=True)
+
+    # compute_vol_features = make_seq_func(
+    #    sample_extractor,
+    #    #lambda x: x.permute(1, 0, 2, 3),
+    #    (model, {'run': 'extract_features'}),
+    #    GlobalMaxPooling()
+    # )
+
+    use_feature_extractor_checkpointing = config['use_feature_extractor_checkpointing']
+    assert use_feature_extractor_checkpointing is not None
+
+    # build a list of sequences. A seuqnece is simply a python function
+
+    funcs_seq = []
+    #import ipdb;ipdb.set_trace()
+    for idx in range(len(use_feature_extractor_checkpointing)):
+
+        start = use_feature_extractor_checkpointing[idx][0]
+        end = use_feature_extractor_checkpointing[idx][1]
+
+        model_func_tuple = (model, {
+            'curr_part': (start, end),
+            'run': 'extract_features'
+        })
+
+        if idx == 0:
+            #for the first one, add the sample extractor
+            funcs_seq += [make_seq_func(
+                sample_extractor,
+                model_func_tuple
+            )]
+        else:
+            funcs_seq += [make_seq_func(
+                model_func_tuple)
+            ]
+    
+    ### add the feature reduction phase to the last function sequence
+    ### (note: assumes that model_func_tuple is unaltered after the loop)
+
+    funcs_seq[-1] = make_seq_func(
+        model_func_tuple,
+        # GlobalMaxPooling(),
+        (model, {'run': 'reduce_features'}),
+    )
+
+    funcs_seq += [
+        make_seq_func(
+            (model, {'run': 'classify'})
+        )
+    ]
+
+    def _calculate_loss(raw_predictions):
+        total_loss, individual_losses, regularization_loss = model.get_total_loss(raw_predictions, targets,
+                                                                                  **regularization_terms)
+        return total_loss, individual_losses, regularization_loss
+
+    def execute_func(f, x):
+        if isinstance(x, list) or isinstance(x, tuple):
+            # x = f(*x)
+            x = checkpoint(f, *x)
+        else:
+            # x = f(x)
+            x = checkpoint(f, x)
+        return x
+
+    
+    if do_backward:
+        
+        if False:
+            cprint('remove this - overriding checkpointing for debugging!!!!!\n' * 4, 'red')
+            assert False
+
+            x = 0
+
+            for f_idx in range(len(funcs_seq)):
+                x = execute_func(funcs_seq[f_idx], x)
+
+            total_loss, individual_losses, regularization_loss = _calculate_loss(x)
+
+            raw_predictions = x
+
+            total_loss.backward()
+        else:
+            raw_predictions, total_loss, individual_losses, regularization_loss = calculate_grads_simulating_flat_checkpointing(
+                funcs_seq,
+                # stored_features[slices_chunk_step_num].grad,
+                0,
+                loss_calc_func=_calculate_loss,
+                cpu_ram_paging=use_cpu_ram_paging
+            )
+    else:
+        with torch.set_grad_enabled(False):
[…]
