+++ paper/paper.md
@@ -84,7 +84,7 @@ FuseMedML's DL package works with PyTorch models, only modifying them to interac
 ## *fuse.eval*
 FuseMedML's evaluation package is a standalone library for evaluating machine learning models using various performance metrics and comparing the results between models. It offers advanced capabilities such as a generic confidence interval wrapper for any metric, a generic one-versus-all extension for converting any binary metric to a multi-class scenario, and metrics for comparing models while considering statistical significance. The package also includes model calibration tools and a pipeline for combining a sequence of metrics with possible dependencies. In addition, the evaluation package supports automatic per-fold evaluation and subgroup analysis, and can handle large data sets through batching and multiprocessing. See \autoref{fig:metric_pipeline} for an example of an evaluation metric pipeline that can be reused across projects.  
 
-![In this example a pipeline of evaluation metric components is shown. It consists of two metrics: the Area Under the receiver operating characteristic Curve and the Area Under the Precision-Recall Curve. Both metrics are wrapped with a Confidence Interval (CI) metric, resulting in a lower and upper bound for each metric. \label{fig:metric_pipeline}](figures/metric_pipeline.png){width=100%}
+![In this example a pipeline of evaluation metric components is shown. It consists of two metrics: the Area Under the receiver operating characteristic Curve and the Area Under the Precision-Recall Curve. Both metrics are wrapped with a Confidence Interval (CI) metric, resulting in a lower and upper bound for each metric. The metrics are executed by an instance of the EvaluatorDefault class, the basic *fuse.eval* class that combines input sources, evaluates using the specified metrics, generates a report and returns a dictionary with all the metrics results.\label{fig:metric_pipeline}](figures/metric_pipeline.png){width=100%}
 
 # Extensions
 The core technology of FuseMedML and its component packages is general, while domain-specific functionality is contained within extensions. One such extension, *fuse-imaging*, is currently available and extends the FuseMedML data package with operations useful for medical imaging, as well as implementations of public medical datasets.  
