### 2023-01-05 #4943 @alex-golts
Thanks for the suggestions @suragnair , I have addressed them as follows:

- Added a figure with a diagram explaining the design philosophy
- Added arrows as you suggested in Fig. 1 (now Fig. 2)
- The "1" argument represents the number of input channels. we keep it currently for backward compatibility. I added a comment to the code in the figure explaining what it is.
- Modified the caption of Fig. 3 (now Fig. 4) to explain EvaluatorDefault

Please let me know if you have any more questions or suggestions
