### 2023-01-05 7d50342
Ukbb changes -ensemble of multi-modal survival models (#231)
* - aug updates
- load 4 channels of the MRI image
- use mlflow
- initial ccc commands

* add pool layer, and fix color normalization

* remove stack

* minor

* select series on dynamic
use cosine and constant lr

* add descriptor to cacher/sampler tqdm

* Revert "add descriptor to cacher/sampler tqdm"

This reverts commit d3cbe0d32c10cf9be5881bba4a582c6d01e7f8f9.

* add descriptor to cacher/sampler tqdm

* fix for descriptor in caching

* support return None in no cache mode

* support missing image key in augmentations.

* revert 
[…]
