+++ fuse/requirements.txt
@@ -9,7 +9,7 @@ scikit-learn>=0.23.2
 termcolor>=1.1.0
 pycocotools>=2.0.1
 torch>=1.5.0
-torchvision>=0.8.1,<0.14.0
+torchvision>=0.8.1,<0.14.0  #removal of upper bound causes issues in tests?
 pytorch_lightning>=1.6
 tensorboard
 wget
