### 2023-01-08 9029c1a
support both tuple of str and int, and just str
