### 2023-01-08 7fe75b2
added trajectories generation + comments
