+++ examples/fuse_examples/multimodality/ehr_transformer/main_train.py
@@ -3,6 +3,7 @@ from typing import Any, Optional
 import hydra
 from omegaconf import DictConfig
 
+
 @hydra.main(config_path=".", config_name="config_PhysioNetCinC")
 def main(cfg: DictConfig):
     print(str(cfg))
@@ -10,6 +11,7 @@ def main(cfg: DictConfig):
     cfg = hydra.utils.instantiate(cfg)
 
     ds_train, ds_valid, _ = PhysioNetCinC.dataset(**cfg.dataset)
+
 
 if __name__ == "__main__":
     main()
