+++ examples/fuse_examples/multimodality/ehr_transformer/dataset.py
@@ -39,8 +39,9 @@ class OpAddBMI(OpBase):
         sample_dict['StaticDetails'] = d_static
         return sample_dict
 
+
 class OpConvertVisitToSentense(OpBase):
-    def __call__(self, sample_dict, visit_to_embed_static_variables,static_variables_to_embed) -> Any:
+    def __call__(self, sample_dict, visit_to_embed_static_variables, static_variables_to_embed) -> Any:
 
         df_static = sample_dict['StaticDetails']
         df_visits = sample_dict['Visits']
@@ -73,21 +74,23 @@ class OpConvertVisitToSentense(OpBase):
 
         return sample_dict
 
+
 class OpGenerateRandomTrajectoryOfVisits(OpBase):
     def __call__(self, sample_dict) -> Any:
 
         d_visits_sentences = sample_dict['VisitSentences']
         n_visits = len(d_visits_sentences)
-        # TODO add filtering patients with small number of visits
+        # TODO add filtering patients with small number of visits - add to configuration
         if n_visits < 10:
             return sample_dict
 
         # get random first and last visit for generating random trajectory
         # we use n_visits- 1 to keep the last visit for outcome in classifier
         # TODO: "3" should be configured
-        # TODO: add embeddings sentence here as we can cut out the first one
-        start_visit = randrange(int((n_visits-1) / 3))
-        stop_visit = n_visits - randrange(int((n_visits-1) / 3)) - 1
+        # TODO: for case with embedding od static info in first sentence only,
+        #  add embeddings sentence here as we can cut out the first one
+        start_visit = randrange(int((n_visits - 1) / 3))
+        stop_visit = n_visits - randrange(int((n_visits - 1) / 3)) - 1
         keys = list(d_visits_sentences.keys())
         trajectory_keys = keys[start_visit:stop_visit]
         next_visit = keys[stop_visit]
@@ -101,51 +104,39 @@ class OpGenerateRandomTrajectoryOfVisits(OpBase):
         sample_dict['Trajectory'] = trajectory_sentences
         sample_dict['NextVisit'] = d_visits_sentences[next_visit]
 
-
-
-
         return sample_dict
-
-# class OpCollectExamsStatistics(OpBase):
-#     def __call__(self, sample_dict, percentiles: dict) -> Any:
-#         percentiles[sample_dict['PatientId']] = sample_dict['Age']
-#
-#         return sample_dict
 
 
 class OpMapToCategorical(OpBase):
 
     def __call__(self, sample_dict, percentiles: dict) -> Any:
 
-        # mapping static clinical characteristics
+        # convert continuous measurements to categorical ones based on defined percentiles
+
+        # mapping static clinical characteristics (Age, Gender, ICU type, Height, etc)
         for k in sample_dict['StaticDetails'].keys():
             sample_dict['StaticDetails'][k] = k + '_' + \
                                               str(np.digitize(sample_dict['StaticDetails'][k], percentiles[k]))
-
+        # mapping labs exams and clinical characteristics captured during patients' stay in ICU
         if not sample_dict['Visits'].empty:
             sample_dict['Visits']['Value'] = sample_dict['Visits']. \
                 apply(
-                lambda row: row['Parameter'] + '_' + str(np.digitize(row['Value'], percentiles[row['Parameter']])), \
+                lambda row: row['Parameter'] + '_' + str(np.digitize(row['Value'], percentiles[row['Parameter']])),
                 axis=1)
 
         return sample_dict
 
-# class OpConvertVisitToSentense(OpBase):
-#
-#     def __call__(self, sample_dict) -> Any:
-#         print("test")
-#         return sample_dict
 
 class PhysioNetCinC:
[…]
