+++ fuse/dl/models/model_multihead.py
@@ -64,7 +64,7 @@ class ModelMultiHead(torch.nn.Module):
         self.add_module("heads", self.heads)
 
     def forward(self, batch_dict: NDict) -> Dict:
-        if isinstance(self.conv_input, str): # no number of input channels specified
+        if isinstance(self.conv_input, str):  # no number of input channels specified
             self.conv_input = (conv_input,)
         if self.conv_inputs is not None:
             conv_input = torch.cat([batch_dict[conv_input[0]] for conv_input in self.conv_inputs], 1)
