+++ fuse/dl/models/model_multihead.py
@@ -34,7 +34,7 @@ class ModelMultiHead(torch.nn.Module):
         backbone: torch.nn.Module,
         heads: Sequence[torch.nn.Module],
         conv_inputs: Tuple[Tuple[str, int], str] = None,
-        backbone_args: Tuple[Tuple[str, int], ...] = None,
+        backbone_args: Tuple[Tuple[str, int], str] = None,
         key_out_features: str = "model.backbone_features",
     ) -> None:
         """
@@ -65,7 +65,9 @@ class ModelMultiHead(torch.nn.Module):
 
     def forward(self, batch_dict: NDict) -> Dict:
         if isinstance(self.conv_input, str):  # no number of input channels specified
-            self.conv_input = (conv_input,)
+            self.conv_input = (self.conv_input,)
+        if isinstance(self.backbone_args, str):  # no number of input channels specified
+            self.backbone_args = (self.backbone_args,)
         if self.conv_inputs is not None:
             conv_input = torch.cat([batch_dict[conv_input[0]] for conv_input in self.conv_inputs], 1)
             backbone_features = self.backbone.forward(conv_input)
