### 2023-01-08 812f02d
fixed local/globals, added comments, restored mandatory paramaters to pl_module
