+++ examples/fuse_examples/imaging/classification/mnist/simple_mnist_starter.py
@@ -33,6 +33,7 @@ import torch.optim as optim
 from torch.utils.data.dataloader import DataLoader
 import pytorch_lightning as pl
 
+
 from fuse.eval.metrics.classification.metrics_thresholding_common import MetricApplyThresholds
 from fuse.eval.metrics.classification.metrics_classification_common import MetricAccuracy
 
@@ -46,19 +47,20 @@ from fuse.dl.lightning.pl_module import LightningModuleDefault
 from fuseimg.datasets.mnist import MNIST
 from examples.fuse_examples.imaging.classification.mnist import lenet
 
+## Paths and Hyperparameters ############################################
 ROOT = "_examples/mnist"  # TODO: fill path here
-model_dir = os.path.join(ROOT, "model_dir")
+MODEL_DIR = os.path.join(ROOT, "model_dir")
 PATHS = {
-    "model_dir": model_dir,
+    "model_dir": MODEL_DIR,
     "cache_dir": os.path.join(ROOT, "cache_dir"),
-    "inference_dir": os.path.join(model_dir, "infer_dir"),
-    "eval_dir": os.path.join(model_dir, "eval_dir"),
+    "inference_dir": os.path.join(MODEL_DIR, "infer_dir"),
+    "eval_dir": os.path.join(MODEL_DIR, "eval_dir"),
 }
-train_params = {
+TRAIN_PARAMS = {
     "data.batch_size": 100,
     "data.train_num_workers": 8,
     "data.validation_num_workers": 8,
-    "trainer.num_epochs": 12,
+    "trainer.num_epochs": 2,
     "trainer.num_devices": 1,
     "trainer.accelerator": "gpu",
     "trainer.strategy": "dp",
@@ -67,42 +69,61 @@ train_params = {
     "opt.weight_decay": 0.001,
 }
 
-
+## Model Definition #####################################################
 class FuseLitLenet(LightningModuleDefault):
-    def __init__(self):
-        super().__init__()
+    def __init__(self, paths, train_params):
+
         # wrap basic torch model to automatically read inputs from batch_dict and save its outputs to batch_dict
-        self._model = ModelWrapSeqToDict(
+        model = ModelWrapSeqToDict(
             model=lenet.LeNet(),
             model_inputs=["data.image"],
             post_forward_processing_function=lambda logits: (logits, F.softmax(logits, dim=1)),
             model_outputs=["model.logits.classification", "model.output.classification"],
         )
-        self._losses = {
+
+        # losses and metrics to track
+        losses = {
             "cls_loss": LossDefault(
                 pred="model.logits.classification", target="data.label", callable=F.cross_entropy, weight=1.0
             ),
         }
-        self._train_metrics = OrderedDict(
+        train_metrics = OrderedDict(
             [
                 ("operation_point", MetricApplyThresholds(pred="model.output.classification")),  # will apply argmax
                 ("accuracy", MetricAccuracy(pred="results:metrics.operation_point.cls_pred", target="data.label")),
             ]
         )
-        self._validation_metrics = copy.deepcopy(self._train_metrics)
+        validation_metrics = copy.deepcopy(train_metrics) # same as train metrics
+
+        # optimizer and learning rate scheduler
         optimizer = optim.Adam(
-            self._model.parameters(), lr=train_params["opt.lr"], weight_decay=train_params["opt.weight_decay"]
+            model.parameters(), lr=train_params["opt.lr"], weight_decay=train_params["opt.weight_decay"]
         )
         lr_scheduler = optim.lr_scheduler.ReduceLROnPlateau(optimizer)
         lr_sch_config = dict(scheduler=lr_scheduler, monitor="validation.losses.total_loss")
[…]
