+++ examples/fuse_examples/multimodality/ehr_transformer/main_train.py
@@ -2,6 +2,7 @@ from dataset import PhysioNetCinC
 from typing import Any, Optional
 import hydra
 from omegaconf import DictConfig
+from ehrtransformers.model.utils import WordVocab
 
 
 @hydra.main(config_path=".", config_name="config_PhysioNetCinC")
@@ -11,6 +12,7 @@ def main(cfg: DictConfig):
     cfg = hydra.utils.instantiate(cfg)
 
     corpus, ds_train, ds_valid, _ = PhysioNetCinC.dataset(**cfg.dataset)
+    token2idx = WordVocab(corpus, max_size=None, min_freq=1).get_stoi()
 
 
 if __name__ == "__main__":
