### 2023-01-08 cbec12a
Revert "black formatter applied"
This reverts commit b46b8ad6e84adf033044621394a1b36e2d216762.
