+++ fuseimg/datasets/kits21.py
@@ -258,7 +258,10 @@ class KITS21:
         )
 
         my_dataset = DatasetDefault(
-            sample_ids=sample_ids, static_pipeline=static_pipeline, dynamic_pipeline=dynamic_pipeline, cacher=cacher,
+            sample_ids=sample_ids,
+            static_pipeline=static_pipeline,
+            dynamic_pipeline=dynamic_pipeline,
+            cacher=cacher,
         )
         my_dataset.create()
         return my_dataset
