+++ fuse/utils/cpu_profiling/timer.py
@@ -21,7 +21,7 @@ if __name__ == "__main__":
 
     def foo(x):
         with Timer("banana"):
-            for _ in range(10 ** 5):
+            for _ in range(10**5):
                 x = x * 2
 
     foo(20)
