+++ examples/fuse_examples/multimodality/ehr_transformer/ehrtransformers/model/metrics.py
@@ -22,15 +22,13 @@ from torch import nn
 import numpy as np
 import torch
 
-
 def get_nontrivial_inds(label):
     """
 
     :param label: tensor of binary GT labels, that contains only '1.0' and '0.0'
     :return: indices of columns that contain both positive and negative classes
     """
-    return [ind for ind in range(label.shape[1]) if ((1.0 in label[:, ind]) and (0.0 in label[:, ind]))]
-
+    return [ind for ind in range(label.shape[1]) if ((1.0 in label[:,ind]) and (0.0 in label[:,ind]))]
 
 def precision(logits, label):
     if isinstance(label, list):
@@ -43,7 +41,9 @@ def precision(logits, label):
     sig = nn.Sigmoid()
     output = sig(logits)
     label, output = label.cpu(), output.detach().cpu()
-    tempprc = metrics.average_precision_score(label.numpy(), output.numpy(), average="macro")
+    tempprc = metrics.average_precision_score(
+        label.numpy(), output.numpy(), average="macro"
+    )
     return tempprc, output, label
 
 
@@ -77,7 +77,9 @@ def precision_test(logits, label):
 
     sig = nn.Sigmoid()
     output = sig(logits)
-    tempprc = metrics.average_precision_score(label.numpy(), output.numpy(), average="macro")
+    tempprc = metrics.average_precision_score(
+        label.numpy(), output.numpy(), average="macro"
+    )
     #     roc = metrics.roc_auc_score()
     return tempprc, output, label
 
