+++ examples/fuse_examples/multimodality/ehr_transformer/make_data.py
@@ -28,39 +28,34 @@ import numpy as np
 from ehrtransformers.configs.config import get_config
 from datetime import datetime
 
-
 def validate_naming_conventions(naming_conventions):
-    assert "patient_id_key" in naming_conventions
-    assert "index_date_key" in naming_conventions
-    assert (
-        "svc_date_key" in naming_conventions
-    )  # not strictly necessary - we can do without this column, if admission date column (adm_date_key) exists
-    assert "adm_date_key" in naming_conventions
-    assert "index_date_key" in naming_conventions
-    assert "age_key" in naming_conventions
+    assert 'patient_id_key' in naming_conventions
+    assert 'index_date_key' in naming_conventions
+    assert 'svc_date_key' in naming_conventions #not strictly necessary - we can do without this column, if admission date column (adm_date_key) exists
+    assert 'adm_date_key' in naming_conventions
+    assert 'index_date_key' in naming_conventions
+    assert 'age_key' in naming_conventions
     # assert 'age_month_key'  in naming_conventions
-    assert "date_birth_key" in naming_conventions
-    assert "gender_key" in naming_conventions
+    assert 'date_birth_key' in naming_conventions
+    assert 'gender_key' in naming_conventions
     # assert 'dxver_key'  in naming_conventions
-    assert "outcome_key" in naming_conventions
+    assert 'outcome_key' in naming_conventions
 
-    assert isinstance(naming_conventions["patient_id_key"], str)
-    assert isinstance(naming_conventions["index_date_key"], str)
-    assert isinstance(naming_conventions["svc_date_key"], str)
-    assert isinstance(naming_conventions["adm_date_key"], str)
-    assert isinstance(naming_conventions["age_key"], str)
+    assert isinstance(naming_conventions['patient_id_key'], str)
+    assert isinstance(naming_conventions['index_date_key'], str)
+    assert isinstance(naming_conventions['svc_date_key'], str)
+    assert isinstance(naming_conventions['adm_date_key'], str)
+    assert isinstance(naming_conventions['age_key'], str)
     # assert isinstance(naming_conventions['age_month_key'], str)
-    assert isinstance(naming_conventions["date_birth_key"], str)
-    assert isinstance(naming_conventions["gender_key"], str)
+    assert isinstance(naming_conventions['date_birth_key'], str)
+    assert isinstance(naming_conventions['gender_key'], str)
     # assert isinstance(naming_conventions['dxver_key'], str)
-    assert isinstance(naming_conventions["outcome_key"], str)
+    assert isinstance(naming_conventions['outcome_key'], str)
 
 
-class TimeFilter:
+class TimeFilter():
     date_column_name: str = None  # name of the column containing (visit) dates to be filtered
-    ref_date_column_name = (
-        None  # name of a column containing a date to be used as a reference for filtering (e.g. index date)
-    )
+    ref_date_column_name = None  # name of a column containing a date to be used as a reference for filtering (e.g. index date)
     start_days_before = None  # visits ealier than start_days_before before the reference date are filtered out. May be negative (in which case treated as days after ref. date)
     stop_days_before = None  # visits later than stop_days_before before reference date are filtered out. May be negative (in which case treated as days after ref. date)
 
@@ -78,7 +73,7 @@ class TimeFilter:
         self.start_days_before = start_days_before
         self.stop_days_before = stop_days_before
         self.ref_date_column_name = ref_date_column_name
-        self.date_format_string = "%Y-%m-%d"  # TODO: This may not be the format in the other input df's - be careful when switching to other data sources
+        self.date_format_string = '%Y-%m-%d' # TODO: This may not be the format in the other input df's - be careful when switching to other data sources
 
[…]
