+++ examples/fuse_examples/multimodality/ehr_transformer/ehrtransformers/data_access/icd_to_ccs.py
@@ -20,6 +20,7 @@ Created on June 30, 2021
 from os import environ
 import os
 import pandas as pd
+
 # from ibm_db import connect
 # from ibm_db_dbi import Connection
 import json
@@ -37,11 +38,12 @@ def get_repeat_codes_dict(icd_ccs_map_df, icd_to_use="icd9"):
     repeat_codes = repeat_codes[repeat_codes.ICD_VERSION == icd_to_use]
     return repeat_codes.set_index("CODE")["CCS_CATEGORY"].to_dict()
 
+
 def read_icd_ccs_dict(json_filepath=None):
     if json_filepath is None:
         log_path = os.path.abspath(__file__)
         code_path = os.path.dirname(os.path.dirname(os.path.dirname(log_path)))
-        json_filepath = os.path.join(code_path, 'ehrtransformers/data_access/icd_to_ccs.json')
+        json_filepath = os.path.join(code_path, "ehrtransformers/data_access/icd_to_ccs.json")
 
     with open(json_filepath) as f:
         tmp = json.load(f)
