+++ examples/fuse_examples/multimodality/ehr_transformer/ehrtransformers/data_access/utils.py
@@ -23,6 +23,7 @@ import json
 import pandas as pd
 import ehrtransformers.configs.naming as naming
 
+
 def seq_translate(tokens, translate_dict, mask_token=None):
     """
         returns a list of tokens translated using translate_dict
@@ -31,7 +32,16 @@ def seq_translate(tokens, translate_dict, mask_token=None):
         :param mask_token: special token that is always translated to 'UNK'
         :return:
         """
-    return tokens, [translate_dict.get(token, translate_dict[naming.unknown_token]) if token != mask_token else translate_dict[naming.unknown_token] for token in tokens]
+    return (
+        tokens,
+        [
+            translate_dict.get(token, translate_dict[naming.unknown_token])
+            if token != mask_token
+            else translate_dict[naming.unknown_token]
+            for token in tokens
+        ],
+    )
+
 
 def load_icd_to_ccs() -> dict:
     """Load the icd to ccs mapping supplied by the package.
@@ -94,7 +104,7 @@ def detector_idx(tokens, symbol=naming.separator_token):
     for token in tokens:
         detector.append(flag)
         if token == symbol:
-            flag = 1-flag
+            flag = 1 - flag
     return detector
 
 
@@ -114,6 +124,7 @@ def position_idx(tokens, symbol=naming.separator_token):
             flag += 1
     return group_inds
 
+
 def seq_pad(tokens, max_len, symbol=naming.padding_token):
     """
     Returns a list of tokens padded by symbol to length max_len.
@@ -125,20 +136,29 @@ def seq_pad(tokens, max_len, symbol=naming.padding_token):
     """
     token_len = len(tokens)
     if token_len < max_len:
-        return list(tokens) + [symbol] * (max_len-token_len)
+        return list(tokens) + [symbol] * (max_len - token_len)
 
     else:
         return tokens[:max_len]
 
-class DataAdder():
+
+class DataAdder:
     source_keys = None
     target_key = None
     mult_comb_mode = None
     source_on_keys = None
     target_on_keys = None
 
-    def __init__(self, source_data_dict, source_keys, target_key, source_on_keys, target_on_keys,
-                 mult_source_comb_mode=None, val_fillna=None):
+    def __init__(
+        self,
+        source_data_dict,
+        source_keys,
+        target_key,
+        source_on_keys,
+        target_on_keys,
+        mult_source_comb_mode=None,
+        val_fillna=None,
+    ):
         """
 
         :param source_data_dict: dictionary of set:path:
@@ -157,12 +177,12 @@ class DataAdder():
         if isinstance(source_keys, list):
             self.source_keys = source_keys
         elif isinstance(source_keys, str):
-            if source_keys == 'all':
-                self.source_keys = 'all'
+            if source_keys == "all":
+                self.source_keys = "all"
             else:
                 self.source_keys = [source_keys]
         else:
-            raise Exception('unexpected source_keys')
+            raise Exception("unexpected source_keys")
         self.target_key = target_key
         self.mult_comb_mode = mult_source_comb_mode
         self.source_on_keys = source_on_keys
@@ -173,20 +193,20 @@ class DataAdder():
     def get_num_classes(self):
         if isinstance(self.source_keys, list):
             return len(self.source_keys)
-        elif self.source_keys == 'all':
+        elif self.source_keys == "all":
             all_keys = set([])
             set_paths = []
             for set_name in self.source_data_dict:
                 set_name = set_name.lower()
-                if 'val' in set_name:
-                    if self.source_data_dict['val'] is not None:
-                        set_paths.append(self.source_data_dict['val'])
-                elif 'tr' in set_name:
-                    if self.source_data_dict['train'] is not None:
-                        set_paths.append(self.source_data_dict['train'])
[…]
