### 2023-01-08 67c0cf8
Made model args clear with annotations, commented model def.
