+++ examples/fuse_examples/multimodality/ehr_transformer/dataset.py
@@ -2,9 +2,8 @@ import pickle
 from collections import OrderedDict
 from random import randrange
 from typing import Sequence, Any
-import re
+
 import glob
-import os
 
 import numpy as np
 import pandas as pd
@@ -13,12 +12,8 @@ from typing import Tuple
 from fuse.data import DatasetDefault, PipelineDefault, OpBase
 from fuse.data.ops.ops_read import OpReadDataframe
 
-# from fuse.data.ops.ops_common import OpCond, OpSet
 from fuse.data.utils.split import dataset_balanced_division_to_folds
-from ops_read_cinc import OpReadDataframeCinC
 from fuse.data.utils.export import ExportDataset
-
-SOURCE = r"C:/D_Drive/Projects/EHR_Transformer/PhysioNet/predicting-mortality-of-icu-patients-the-physionetcomputing-in-cardiology-challenge-2012-1.0.0/predicting-mortality-of-icu-patients-the-physionet-computing-in-cardiology-challenge-2012-1.0.0"
 
 VALID_TESTS_ABOVE_ZERO = [
     "pH",
@@ -42,17 +37,31 @@ STATIC_FIELDS = ["Age", "Gender", "Height", "ICUType", "Weight"]
 
 
 class OpAddBMI(OpBase):
+    """
+    Apply BMI calculation using static patient's fields. Names of the fields are provided as parameters as well as
+    output field name.
+    Height expected to be in centimetres
+    Weight expected to be in kilograms
+
+    Example:
+    OpAddBMI(key_in_height="Height", key_in_weight="Weight", key_out_bmi="BMI")
+    """
+
+    def __init__(self, key_in_height: str, key_in_weight: str, key_out_bmi: str):
+
+        super().__init__()
+        self._key_height = key_in_height
+        self._key_weight = key_in_weight
+        self._key_bmi = key_out_bmi
+
     def __call__(self, sample_dict) -> Any:
 
-        d_static = sample_dict["StaticDetails"]
-
-        if ("Height" in d_static.keys()) & ("Weight" in d_static.keys()):
-            height = d_static["Height"]
-            weight = d_static["Weight"]
+        if (self._key_height in sample_dict["StaticDetails"]) & (self._key_weight in sample_dict["StaticDetails"]):
+            height = sample_dict["StaticDetails." + self._key_height]
+            weight = sample_dict["StaticDetails." + self._key_weight]
             if ~np.isnan(height) & ~np.isnan(weight):
-                d_static["BMI"] = 10000 * weight / (height * height)
+                sample_dict["StaticDetails." + self._key_bmi] = 10000 * weight / (height * height)
 
-        sample_dict["StaticDetails"] = d_static
         return sample_dict
 
 
@@ -96,15 +105,14 @@ class OpGenerateRandomTrajectoryOfVisits(OpBase):
 
         d_visits_sentences = sample_dict["VisitSentences"]
         n_visits = len(d_visits_sentences)
-        # TODO add filtering patients with small number of visits - add to configuration
-        if n_visits < 10:
-            return sample_dict
 
         # get random first and last visit for generating random trajectory
         # we use n_visits- 1 to keep the last visit for outcome in classifier
         # TODO: "3" should be configured
-        # TODO: for case with embedding od static info in first sentence only,
+        # TODO: for case with embedding of static info in first sentence only,
         #  add embeddings sentence here as we can cut out the first one
+        #  other option to drop randomly X visits in random places
+        #  verify that the resulted trajectory is less than max_len (need to check in Vadim's code)
         start_visit = randrange(int((n_visits - 1) / 3))
         stop_visit = n_visits - randrange(int((n_visits - 1) / 3)) - 1
[…]
