+++ examples/fuse_examples/multimodality/ehr_transformer/config_PhysioNetCinC.yaml
@@ -1,6 +1,7 @@
 dataset:
-  raw_data_path: 'C:/D_Drive/Projects/EHR_Transformer/PhysioNet/predicting-mortality-of-icu-patients-the-physionetcomputing-in-cardiology-challenge-2012-1.0.0/predicting-mortality-of-icu-patients-the-physionet-computing-in-cardiology-challenge-2012-1.0.0'
+  #raw_data_path: 'C:/D_Drive/Projects/EHR_Transformer/PhysioNet/predicting-mortality-of-icu-patients-the-physionetcomputing-in-cardiology-challenge-2012-1.0.0/predicting-mortality-of-icu-patients-the-physionet-computing-in-cardiology-challenge-2012-1.0.0'
   #raw_data_pickle: 'raw_data.pkl'
+  raw_data_path: ${oc.env:CINC_DATA_PATH}
   split_filename: None
   num_folds: 5
   train_folds: [ 0, 1, 2 ]
@@ -11,20 +12,21 @@ dataset:
   num_percentiles : 4 #number of bins (percentiles) for converting floating lab/vital measurements to categorical values
   categorical_max_num_of_values : 5 #max number of uniq values for categorical variable for not to be digitized
   min_hours_in_hospital: 46
+  min_number_of_visits: 10
   visit_to_embed_static_variables: FIRST #embed to first (FIRST) or to all visits (ALL) or null (not to embed)
   static_variables_to_embed: ['Age','ICUType','Height','Weight','BMI',]
 
 data:
   #### Test/Val set defs
-  days_to_inddate: 60  #90 #None #180 #only visits before days_to_ind to index date (or after index date, if negative) are considered for test/val
-  days_to_inddate_start: null #only the visits between ind_date-days_to_inddate_start and ind_date-days_to_inddate are considered for test/val
-  event_prediction_window_days: null  #90
-  treatment_event_prediction_window_days: null  #60
-  #### Train set defs
-  # in case of train we take all visits (we want to train on as much data as we can), so days_to_inddate_tr and days_to_inddate_start_tr should be both null
-  days_to_inddate_tr: 60  #90 #None #180 #only visits before days_to_ind to index date (or after index date, if negative) are considered for train
-  days_to_inddate_start_tr: null #only the visits between days_to_inddate_start and days_to_inddate are considered for train
-  event_prediction_window_days_tr: null  #90
+#  days_to_inddate: 60  #90 #None #180 #only visits before days_to_ind to index date (or after index date, if negative) are considered for test/val
+#  days_to_inddate_start: null #only the visits between ind_date-days_to_inddate_start and ind_date-days_to_inddate are considered for test/val
+#  event_prediction_window_days: null  #90
+#  treatment_event_prediction_window_days: null  #60
+#  #### Train set defs
+#  # in case of train we take all visits (we want to train on as much data as we can), so days_to_inddate_tr and days_to_inddate_start_tr should be both null
+#  days_to_inddate_tr: 60  #90 #None #180 #only visits before days_to_ind to index date (or after index date, if negative) are considered for train
+#  days_to_inddate_start_tr: null #only the visits between days_to_inddate_start and days_to_inddate are considered for train
+#  event_prediction_window_days_tr: null  #90
 
   visit_days_resolution: 10 #consecutive visits that are less than visit_days_resolution days apart are merged into one visit.
[…]
