### 2023-01-09 #4943 @suragnair
This is slightly open-ended and subjective. The Hello World notebook says "By the end of the session we hope you'll be familiar with basic Fuse's workflow and _acknowledge its potential_." However, after going through the paper, docs and notebooks, I am still not able to fully grasp the power of FuseMedML over typical workflows.

I fully understand that FuseMedML implements a lot of handy functions, especially for evaluation, and I can see how that by itself is useful for ML projects. However, I'm not fully convinced how the NDict concept enhances reproducibility by itself. One could argue that well documented individual functions/classes would probably be just as useful?

Would be interested to hear your thoughts on this. It would also be very helpful to add some more comments to the Hello World notebook, especially specifying how FuseMedML concepts enhance reproducibility over how someone would do it without FuseMedML.
