### 2023-01-09 087acdf
updated Op of generation trajectories, added postions, and indexes
