+++ examples/fuse_examples/multimodality/ehr_transformer/dataset.py
@@ -4,7 +4,7 @@ from random import randrange
 from typing import Sequence, Any
 
 import glob
-
+import os
 import numpy as np
 import pandas as pd
 from typing import Tuple
@@ -14,6 +14,8 @@ from fuse.data.ops.ops_read import OpReadDataframe
 
 from fuse.data.utils.split import dataset_balanced_division_to_folds
 from fuse.data.utils.export import ExportDataset
+from utils import seq_translate, position_idx, special_tokens, seq_pad
+from ehrtransformers.model.utils import WordVocab
 
 VALID_TESTS_ABOVE_ZERO = [
     "pH",
@@ -66,31 +68,30 @@ class OpAddBMI(OpBase):
 
 
 class OpConvertVisitToSentense(OpBase):
-    def __call__(self, sample_dict, visit_to_embed_static_variables, static_variables_to_embed) -> Any:
+    def __init__(self, static_variables_to_embed: list):
+
+        super().__init__()
+        self._static_variables_to_embed = static_variables_to_embed
+
+    def __call__(self, sample_dict) -> Any:
 
         df_static = sample_dict["StaticDetails"]
         df_visits = sample_dict["Visits"]
+
         # convert statis details for embedding to list
         static_embeddings = []
-        if visit_to_embed_static_variables is not None:
+        if len(self._static_variables_to_embed) > 0:
             for k in df_static.keys():
-                if k in static_variables_to_embed:
+                if k in self._static_variables_to_embed:
                     static_embeddings += [df_static[k]]
 
         d_visit_sentences = OrderedDict()
-        first_visit_embedded = False
 
         for visit_time, df_visit in df_visits.groupby("DateTime", sort=True):
 
             d_visit_sentences[visit_time] = []
-
-            if visit_to_embed_static_variables == "FIRST":
-                if not first_visit_embedded:
-                    d_visit_sentences[visit_time].extend(static_embeddings)
-                    first_visit_embedded = True
-            else:
-                if visit_to_embed_static_variables == "ALL":
-                    d_visit_sentences[visit_time].extend(static_embeddings)
+            if len(static_embeddings) > 0:
+                d_visit_sentences[visit_time] = static_embeddings.copy()
 
             d_visit_sentences[visit_time].extend(df_visit["Value"].to_list())
             d_visit_sentences[visit_time].extend(["SEP"])
@@ -101,6 +102,12 @@ class OpConvertVisitToSentense(OpBase):
 
 
 class OpGenerateRandomTrajectoryOfVisits(OpBase):
+    def __init__(self, max_len: int, vocab: dict):
+
+        super().__init__()
+        self._max_len = max_len
+        self._vocab = vocab
+
     def __call__(self, sample_dict) -> Any:
 
         d_visits_sentences = sample_dict["VisitSentences"]
@@ -108,27 +115,32 @@ class OpGenerateRandomTrajectoryOfVisits(OpBase):
 
         # get random first and last visit for generating random trajectory
         # we use n_visits- 1 to keep the last visit for outcome in classifier
-        # TODO: "3" should be configured
-        # TODO: for case with embedding of static info in first sentence only,
-        #  add embeddings sentence here as we can cut out the first one
+        # TODO: "2" should be configured
         #  other option to drop randomly X visits in random places
-        #  verify that the resulted trajectory is less than max_len (need to check in Vadim's code)
-        start_visit = randrange(int((n_visits - 1) / 3))
-        stop_visit = n_visits - randrange(int((n_visits - 1) / 3)) - 1
+        start_visit = randrange(int((n_visits - 1) / 2))
+        stop_visit = n_visits - randrange(int((n_visits - 1) / 2)) - 1
         keys = list(d_visits_sentences.keys())
[…]
