### 2023-01-09 e7051e8
updated Op of generation trajectories, added postions, and indexes
