### 2023-01-10 83d4dcc
add down sample option to BatchSamplerDefault
