+++ fuse/data/utils/samplers.py
@@ -61,7 +61,9 @@ class BatchSamplerDefault(BatchSampler):
                                         In mode 'exact' the weights should be integers that sums up to batch size.
                                         In mode 'approx' the weights should be floats that sums up to ~1
                                         If not specified, equal number of samples from each class will be used.
-        :param num_batches: optional - if set will force num_batches, otherwise num_batches will be automatically to go over each sample at least once (exactly or approximately).
+        :param num_batches: optional - if set to an int will force num_batches, 
+                                       if set to None or "over_sample"  - num_batches will be automatically to go over each sample at least once (exactly or approximately).
+                                       if set to "down_sample"  - num_batches will be automatically to go over each sample in one of the classes (exactly or approximately) and the rest will be down sampled.
         :param verbose:
         :param dataset_get_multi_kwargs: extra parameters for dataset.get_multi() to optimize the running time.
         """
@@ -106,6 +108,10 @@ class BatchSamplerDefault(BatchSampler):
         if self._balanced_class_weights is None:
             self._balanced_class_weights_list = None
         else:
+            # add missing weights (set probability / num instances in a batch to zero)
+            self._balanced_class_weights = {
+                value: self._balanced_class_weights.get(value, 0) for value in self._balanced_class_values
+            }
             self._balanced_class_weights_list = [
                 self._balanced_class_weights.get(value, 0) for value in self._balanced_class_values
             ]
@@ -139,9 +145,9 @@ class BatchSamplerDefault(BatchSampler):
                 raise Exception("Error: in mode 'approx', batch size must be set.")
             if self._balanced_class_weights_list is not None:
                 for weight in self._balanced_class_weights_list:
-                    if not isinstance(weight, float):
+                    if weight > 1.0:
                         raise Exception(
-                            f"Error: in mode 'exact', expecting only floats in balanced_class_weights, got {type(weight)}"
+                            f"Error: in mode 'exact', expecting probabilities, got {weight}"
                         )
                 if not math.isclose(sum(self._balanced_class_weights_list), 1.0):
                     raise Exception(
@@ -196,7 +202,7 @@ class BatchSamplerDefault(BatchSampler):
 
         # Calculate num batches. Number of batches to iterate over all data at least once (exactly or approximately)
         # Calculate only if not directly specified by the user
-        if self._num_batches is None:
+        if self._num_batches is None or self._num_batches in ["over_sample", "down_sample"]:
             if self._mode == "exact":
                 samples_per_batch = self._balanced_class_weights
             else:  # mode is approx
@@ -210,8 +216,13 @@ class BatchSamplerDefault(BatchSampler):
                 else 0
                 for cls_i in self._balanced_class_sizes
             ]
-            bigger_balanced_class_weighted_size = max(balanced_class_weighted_sizes)
-            self._num_batches = int(bigger_balanced_class_weighted_size)
+            if self._num_batches == "down_sample":
+                balanced_class_weighted_size = min(balanced_class_weighted_sizes)
+            else: # over sample
+                balanced_class_weighted_size = max(balanced_class_weighted_sizes)
+            
+            self._num_batches = int(balanced_class_weighted_size)
+
 
         # pointers per class
         self._cls_pointers = {cls_i: 0 for cls_i in self._balanced_class_values}
