### 2023-01-10 19cb648
bug fixes in generating trajectory of visits Op
