+++ examples/fuse_examples/multimodality/ehr_transformer/main_train.py
@@ -1,5 +1,4 @@
 from dataset import PhysioNetCinC
-from typing import Any, Optional
 import hydra
 from omegaconf import DictConfig
 
