### 2023-01-10 5b65e69
Exposed max_allowed_used_space to SamplesCacher __init__ and removed hard coded num_workers from knight dataset.
