### 2023-01-10 8b98bf0
removed hard-coded num_workers
