### 2023-01-10 24e61c5
moved WordVocab to local "utils.py"; added path to data pkl to config; other misc changes
