+++ examples/fuse_examples/multimodality/ehr_transformer/dataset.py
@@ -14,8 +14,9 @@ from fuse.data.ops.ops_read import OpReadDataframe
 
 from fuse.data.utils.split import dataset_balanced_division_to_folds
 from fuse.data.utils.export import ExportDataset
-from utils import seq_translate, position_idx, special_tokens, seq_pad
-from ehrtransformers.model.utils import WordVocab
+from utils import seq_translate, position_idx, special_tokens, seq_pad, WordVocab
+
+# from ehrtransformers.model.utils import WordVocab
 
 VALID_TESTS_ABOVE_ZERO = [
     "pH",
@@ -137,16 +138,16 @@ class OpGenerateRandomTrajectoryOfVisits(OpBase):
         tokens = [special_tokens["cls"]] + tokens
         tokens = seq_pad(tokens, self._max_len)
         positions = position_idx(tokens)
-        indexes = seq_translate(tokens, self._vocab)
+        token_ids = seq_translate(tokens, self._vocab)
 
         sample_dict["Tokens"] = tokens
         sample_dict["Positions"] = positions
-        sample_dict["Indexes"] = indexes
+        sample_dict["Indexes"] = token_ids
 
         # outcomes of next visit prediction
         # TODO - add to configuration, discussion should be discussed
         sample_dict["NextVisitTokens"] = d_visits_sentences[next_visit]
-        sample_dict["NextVisitIndexes"] = seq_translate(sample_dict["NextVisitTokens"], self._vocab)
+        sample_dict["NextVisitTokenIds"] = seq_translate(sample_dict["NextVisitTokens"], self._vocab)
 
         # TODO - need to handle in configuration the details of other heads (to discuss with Moshiko):
         #    predicting mortality - outcome in sample_dict['Target']
@@ -156,19 +157,18 @@ class OpGenerateRandomTrajectoryOfVisits(OpBase):
 
 
 class OpMapToCategorical(OpBase):
-    def __call__(self, sample_dict, percentiles: dict) -> Any:       
+    def __call__(self, sample_dict, bins: dict) -> Any:
 
-        # convert continuous measurements to categorical ones based on defined percentiles
+        # convert continuous measurements to categorical ones based on defined bins
         # mapping static clinical characteristics (Age, Gender, ICU type, Height, etc)
         for k in sample_dict["StaticDetails"]:
-            sample_dict["StaticDetails"][k] = (
-                k + "_" + str(np.digitize(sample_dict["StaticDetails"][k], percentiles[k]))
-            )
+            sample_dict["StaticDetails"][k] = k + "_" + str(np.digitize(sample_dict["StaticDetails"][k], bins[k]))
         # mapping labs exams and clinical characteristics captured during patients' stay in ICU
         if not sample_dict["Visits"].empty:
             sample_dict["Visits"]["Value"] = sample_dict["Visits"].apply(
-                lambda row: row["Parameter"] + "_" + str(np.digitize(row["Value"], percentiles[row["Parameter"]])) \
-                    if  row["Parameter"] in percentiles else special_tokens["unknown"], 
+                lambda row: row["Parameter"] + "_" + str(np.digitize(row["Value"], bins[row["Parameter"]]))
+                if row["Parameter"] in bins
+                else special_tokens["unknown"],
                 axis=1,
             )
 
@@ -183,13 +183,12 @@ class PhysioNetCinC:
         data_sub_sets = ["set-a", "set-b"]
         for s in data_sub_sets:
             csv_files = glob.glob(os.path.join(raw_data_path + "/" + s, "*.txt"))
-            for f in csv_files:  # reducing the list temporarily for debugging
[…]
