### 2023-01-10 462bdb4
made cacheing optional in KNIGHT and made a notebook to compare caching vs not caching
