### 2023-01-10 bfd4d31
single head script with vanila transformer
