+++ examples/fuse_examples/multimodality/ehr_transformer/main_train.py
@@ -1,15 +1,170 @@
 from dataset import PhysioNetCinC
+from typing import Any, Optional, List, Tuple
 import hydra
 from omegaconf import DictConfig
 
+import torch
+import torch.nn.functional as F
 
-@hydra.main(config_path=".", config_name="config_PhysioNetCinC")
+import pytorch_lightning as pl
+
+from torch.utils.data.dataloader import DataLoader
+
+from fuse.data.utils.collates import CollateDefault
+from fuse.data.utils.samplers import BatchSamplerDefault
+from fuse.dl.models.model_wrapper import ModelWrapSeqToDict
+from fuse.dl.lightning.pl_module import LightningModuleDefault
+from fuse.dl.models.heads import Head1D
+from fuse.eval.metrics.metrics_common import Filter
+from fuse.eval.metrics.classification.metrics_classification_common import MetricAUCROC
+from fuse.dl.losses import LossDefault
+from fuse.utils import NDict
+from fuse.data import DatasetDefault
+
+
+from fuse_examples.multimodality.ehr_transformer.model import Embed, TransformerEncoder
+
+
+def data(
+    dataset_cfg: dict, target_key: str, batch_size: int, 
+    data_loader_train: dict, data_loader_valid: dict
+) -> Tuple[DatasetDefault, DataLoader, DataLoader]:
+    
+    token2idx, ds_train, ds_valid, _ = PhysioNetCinC.dataset(**dataset_cfg)
+
+    ds_train[0].print_tree()
+
+    dl_train = DataLoader(
+        ds_train,
+        collate_fn=CollateDefault(keep_keys=["data.sample_id", "Target", "Indexes"]),
+        batch_sampler=BatchSamplerDefault(
+            ds_train,
+            balanced_class_name=target_key,
+            batch_size=batch_size,
+            mode="approx",
+            num_batches=500,
+        ),
+        **data_loader_train,
+    )
+    dl_valid = DataLoader(
+        ds_valid, collate_fn=CollateDefault(keep_keys=["data.sample_id", "Target", "Indexes"]), **data_loader_valid
+    )
+
+    return token2idx, ds_train, dl_train, dl_valid
+
+def model(
+    embed: dict,
+    classifier_head: dict,
+    z_dim: int,
+    transformer_encoder: dict,
+    vocab_size: int,
+):
+    embed = Embed(
+        key_in="Indexes", key_out="model.embedding", n_vocab=vocab_size, **embed
+    )
+
+    encoder_model = TransformerEncoder(**transformer_encoder)
+
+    model = torch.nn.Sequential(
+        embed,
+        ModelWrapSeqToDict(
+            model=encoder_model,
+            model_inputs=["model.embedding"],
+            model_outputs=["model.z", None],
+        ),
+        Head1D(head_name="cls", conv_inputs=[("model.z", z_dim)], **classifier_head),
+    )
+
+    return model
+
+
+
+def train(
+    model: torch.nn.Module,
+    dl_train: DataLoader,
+    dl_valid: DataLoader,
+    model_dir: str,
+    opt: callable,
+    trainer_kwargs: dict,
+    target_key: str,
+    lr_scheduler: callable = None,
+    track_clearml: Optional[dict] = None,
+):
+
+    if track_clearml is not None:
+        from fuse.dl.lightning.pl_funcs import start_clearml_logger
+
+        start_clearml_logger(**track_clearml)
+
+    #  Loss
+    losses = {
+        "ce": LossDefault(
+            pred="model.logits.cls",
+            target=target_key,
+            callable=F.cross_entropy,
+        ),
+    }
+
+    # Metrics
[…]
