### 2023-01-10 f68b92b
starting modification of of old mnist example test
