+++ fuse/dl/models/backbones/backbone_transformer.py
@@ -1,14 +1,17 @@
 import torch
 from torch import nn
+from typing import Optional
 from vit_pytorch.vit import Transformer as _Transformer
 from vit_pytorch.vit import repeat
+import torch.nn.functional as F
+from x_transformers import Encoder, CrossAttender, TransformerWrapper
 
 
 class Transformer(nn.Module):
     """
     Transformer backbone.
     Gets a [batch_size, num_tokens, token_dim] shaped tensor
-    Returns a [batch_size, num_tokens + 1, token_dim] shaped tensor, where the first token is the CLS token
+    Returns a [batch_size, num_tokens + num_cls_tokens, token_dim] shaped tensor, where the first tokens are the CLS tokens
     """
 
     def __init__(
@@ -35,7 +38,7 @@ class Transformer(nn.Module):
     def forward(self, x: torch.Tensor) -> torch.Tensor:
         """
         :param x: [batch_size, num_tokens, token_dim] shaped tensor
-        :return: [batch_size, num_tokens + 1, token_dim] shaped tensor, where the first token is the CLS token
+        :return: [batch_size, num_tokens + num_cls_tokens, token_dim] shaped tensor, where the first tokens are the CLS tokens
         """
         b, n, _ = x.shape
         cls_tokens = repeat(self.cls_token, "1 a d -> b a d", b=b)
@@ -44,3 +47,94 @@ class Transformer(nn.Module):
         x = self.dropout(x)
         x = self.transformer(x)
         return x
+
+
+class CrossAttentionTransformer(nn.Module):
+    """
+    "a" := first modality
+    "b" := second modality
+
+
+    """
+
+    def __init__(
+        self,
+        emb_dim: int,
+        num_tokens_a: int,
+        num_tokens_b: int,
+        max_seq_len_a: int,
+        max_seq_len_b: int,
+        depth_a: int = 6,
+        depth_b: int = 6,
+        depth_cross_attn: int = 6,
+        heads_a: int = 9,
+        heads_b: int = 9,
+        output_dim: Optional[int] = None,
+        context: str = "seq_b",
+    ):
+        """
+        TODO receive params as tuples? 
+        :param emb_dim: model dimension
+        :param context:
+
+        drug -> a
+        target -> b
+        """
+        super().__init__()
+
+        if output_dim is None:
+            output_dim = emb_dim
+
+        assert context in ["seq_a", "seq_b", "both"]
+        self._context = context
+
+        self.enc_a = TransformerWrapper(
+            num_tokens=num_tokens_a,
+            max_seq_len=max_seq_len_a,
+            attn_layers=Encoder(dim=emb_dim, depth=depth_a, heads=heads_a),
+        )
+        self.enc_b = TransformerWrapper(
+            num_tokens=num_tokens_b,
+            max_seq_len=max_seq_len_b,
+            attn_layers=Encoder(dim=emb_dim, depth=depth_b, heads=heads_b),
+        )
+
+        if self._context in ["a", "b"]:
+            self.cross_attn = CrossAttender(dim=emb_dim, depth=depth_cross_attn)
+
+        else:  # both
+            self.cross_attn_a_as_context = CrossAttender(dim=emb_dim, depth=depth_cross_attn)
+            self.cross_attn_b_as_context = CrossAttender(dim=emb_dim, depth=depth_cross_attn)
+            self.ff = nn.Linear(emb_dim * 2, emb_dim)
+
+        self.last_linear = nn.Linear(emb_dim, output_dim)
+
+    def forward(self, xa, xb):
+        enc_xa = self.enc_a(xa, return_embeddings=True)
+        enc_xb = self.enc_b(xb, return_embeddings=True)
+
[…]
