### 2023-01-10 885135d
deleted old mnist starter, edited test to use new one
