+++ examples/fuse_examples/tests/test_classification_mnist.py
@@ -30,8 +30,6 @@ from fuseimg.datasets.mnist import MNIST
 from fuse.utils.rand.seed import Seed
 
 from fuse_examples.imaging.classification.mnist.simple_mnist_starter import (
-    ROOT,
-    MODEL_DIR,
     PATHS,
     TRAIN_PARAMS,
     run_train,
@@ -40,9 +38,10 @@ from fuse_examples.imaging.classification.mnist.simple_mnist_starter import (
 
 def run_mnist(root: str) -> None:
     Seed.set_seed(0, False)  # previous test (in the pipeline) changed the deterministic behavior to True
-    run_train(train_dataset, validation_dataset, paths, train_common_params)
-    results = run_eval(paths, eval_common_params)
-    assert results["metrics.auc.macro_avg"] >= 0.95, "Error: expecting higher performence"
+    try:
+        run_train(PATHS, TRAIN_PARAMS)
+    except:
+        raise Exception("Training MNIST Failed.")
 
 
 class ClassificationMnistTestCase(unittest.TestCase):
