### 2023-01-10 2f12ba7
add down sample option to BatchSamplerDefault (#246)
