### 2023-01-11 b038bd8
added option of adding static details as a first special visit
