+++ examples/fuse_examples/multimodality/ehr_transformer/dataset.py
@@ -15,7 +15,8 @@ from fuse.data.ops.ops_read import OpReadDataframe
 from fuse.data.utils.split import dataset_balanced_division_to_folds
 from fuse.data.utils.export import ExportDataset
 
-from fuse_examples.multimodality.ehr_transformer.utils import seq_translate, position_idx, special_tokens, seq_pad, WordVocab
+# from examples.fuse_examples.multimodality.ehr_transformer.utils import seq_translate, position_idx, special_tokens, seq_pad, WordVocab
+from utils import seq_translate, position_idx, special_tokens, seq_pad, WordVocab
 
 # from ehrtransformers.model.utils import WordVocab
 
@@ -70,30 +71,31 @@ class OpAddBMI(OpBase):
 
 
 class OpConvertVisitToSentence(OpBase):
-    def __init__(self, static_variables_to_embed: list):
+    def __init__(self):
 
         super().__init__()
-        self._static_variables_to_embed = static_variables_to_embed
+        #self._static_variables_to_embed = static_variables_to_embed
 
     def __call__(self, sample_dict) -> Any:
 
-        df_static = sample_dict["StaticDetails"]
+        #df_static = sample_dict["StaticDetails"]
         df_visits = sample_dict["Visits"]
 
-        # convert statis details for embedding to list
-        static_embeddings = []
-        if len(self._static_variables_to_embed) > 0:
-            for k in df_static.keys():
-                if k in self._static_variables_to_embed:
-                    static_embeddings += [df_static[k]]
+        # # convert statis details for embedding to list
+        # static_embeddings = []
+        # if len(self._static_variables_to_embed) > 0:
+        #     for k in df_static.keys():
+        #         if k in self._static_variables_to_embed:
+        #             static_embeddings += [df_static[k]]
 
         d_visit_sentences = OrderedDict()
 
-        for visit_time, df_visit in df_visits.groupby("DateTime", sort=True):
+       # for visit_time, df_visit in df_visits.groupby("DateTime", sort=True):
+        for visit_time, df_visit in df_visits.groupby("Time", sort=True):
 
             d_visit_sentences[visit_time] = []
-            if len(static_embeddings) > 0:
-                d_visit_sentences[visit_time] = static_embeddings.copy()
+            # if len(static_embeddings) > 0:
+            #     d_visit_sentences[visit_time] = static_embeddings.copy()
 
             d_visit_sentences[visit_time].extend(df_visit["Value"].to_list())
             d_visit_sentences[visit_time].extend(["SEP"])
@@ -104,11 +106,13 @@ class OpConvertVisitToSentence(OpBase):
 
 
 class OpGenerateRandomTrajectoryOfVisits(OpBase):
-    def __init__(self, max_len: int, vocab: dict):
+    def __init__(self, max_len: int, vocab: dict, embed_static_in_all_visits: int, static_variables_to_embed: list):
 
         super().__init__()
         self._max_len = max_len
         self._vocab = vocab
+        self._static_variables_to_embed = static_variables_to_embed
+        self._embed_static_in_all_visits = embed_static_in_all_visits
 
     def __call__(self, sample_dict) -> Any:
 
@@ -125,18 +129,37 @@ class OpGenerateRandomTrajectoryOfVisits(OpBase):
         trajectory_keys = keys[start_visit:stop_visit]
         next_visit = keys[stop_visit]
 
+        # add static variables as a first special visit
+        df_static = sample_dict["StaticDetails"]
+        static_embeddings = []
+        if len(self._static_variables_to_embed) > 0:
+            for k in df_static.keys():
+                if k in self._static_variables_to_embed:
+                    static_embeddings += [df_static[k]]
+
[…]
