+++ fuse/dl/models/backbones/backbone_transformer.py
@@ -3,7 +3,6 @@ from torch import nn
 from typing import Optional
 from vit_pytorch.vit import Transformer as _Transformer
 from vit_pytorch.vit import repeat
-import torch.nn.functional as F
 from x_transformers import Encoder, CrossAttender, TransformerWrapper
 
 
@@ -52,9 +51,9 @@ class Transformer(nn.Module):
 
 class CrossAttentionTransformer(nn.Module):
     """
+
     "a" := first modality
     "b" := second modality
-
 
     """
 
@@ -106,7 +105,7 @@ class CrossAttentionTransformer(nn.Module):
             attn_layers=Encoder(dim=emb_dim, depth=depth_b, heads=heads_b),
         )
 
-        if self._context in ["a", "b"]:
+        if self._context in ["seq_a", "seq_b"]:
             self.cross_attn = CrossAttender(dim=emb_dim, depth=depth_cross_attn)
 
         else:  # both
@@ -120,11 +119,11 @@ class CrossAttentionTransformer(nn.Module):
         enc_xa = self.enc_a(xa, return_embeddings=True)
         enc_xb = self.enc_b(xb, return_embeddings=True)
 
-        if self._context == "a":
+        if self._context == "seq_a":
             x = self.cross_attn(enc_xb, context=enc_xa)
             x = x[:, 0]
 
-        if self._context == "b":
+        if self._context == "seq_b":
             x = self.cross_attn(enc_xa, context=enc_xb)
             x = x[:, 0]
 
@@ -133,15 +132,17 @@ class CrossAttentionTransformer(nn.Module):
             x_bca = self.cross_attn_a_as_context(enc_xb, context=enc_xa)
             x = torch.cat((x_acb[:, 0], x_bca[:, 0]), dim=1)
             x = self.ff(x)
+
+        x = self.last_linear(x)
         # ---
         # request output dim
         # two diff emb dim
-        x = self.head(x)
-        logits = x
 
-        cls_preds = F.softmax(logits, dim=1)
-        return logits, cls_preds
+        # TODO delete
+        # x = self.head(x)
+        # logits = x
 
+        # cls_preds = F.softmax(logits, dim=1)
+        # return logits, cls_preds
 
-if __name__ == "__main__":
-    model = CrossAttentionTransformer(512, 20000, 20000, 1024, 1024)
+        return x
