### 2023-01-11 fcd324a
deleted code of the old version of transforment implementation, added comments to files
