+++ examples/fuse_examples/multimodality/ehr_transformer/ehrtransformers/configs/__init__.py
@@ -1,3 +1,0 @@
-from .config import get_config
-
-__all__ = ["get_config"]
