+++ examples/fuse_examples/multimodality/ehr_transformer/ehrtransformers/configs/configmodels.py
@@ -1,190 +1,0 @@
-"""
-(C) Copyright 2021 IBM Corp.
-
-Licensed under the Apache License, Version 2.0 (the "License");
-you may not use this file except in compliance with the License.
-You may obtain a copy of the License at
-
-   http://www.apache.org/licenses/LICENSE-2.0
-
-Unless required by applicable law or agreed to in writing, software
-distributed under the License is distributed on an "AS IS" BASIS,
-WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-See the License for the specific language governing permissions and
-limitations under the License.
-
-Created on June 30, 2021
-
-"""
-
-from typing import Optional
-from pydantic import BaseSettings
-from pydantic import Field
-from typing import Any
-import ehrtransformers.configs.naming as naming
-
-
-class DataSettings(BaseSettings):
-    days_to_inddate: Optional[
-        int
-    ] = 180  # days before index date when the latest visit is considered (later visits are discarded)
-    days_to_inddate_start: Optional[
-        int
-    ] = 180  # days before index date when the first visit is considered (earlier visits are discarded)
-    event_prediction_window_days: Optional[int] = None  # within what window to predict an event
-    treatment_event_prediction_window_days: Optional[int] = None  # within what window to predict treatment event
-
-    days_to_inddate_tr: Optional[
-        int
-    ] = 180  # days before index date when the latest visit is considered (later visits are discarded) in the train set (in case we want to do it differently for train and test)
-    days_to_inddate_start_tr: Optional[
-        int
-    ] = 180  # days before index date when the earliest visit is considered (earlier visits are discarded) in the train set (in case we want to do it differently for train and test)
-    event_prediction_window_days_tr: Optional[int] = None
-    limit_visits: Optional[int] = None
-    out_type: Optional[str] = "LABEL"
-    task: Optional[str] = "PD"
-    subtask: Optional[str] = "MEDICARE_INPATIENT"
-    task_type: Optional[str] = "outcome"
-    output_name: Optional[str] = "admdate_visit"
-    use_procedures: Optional[
-        bool
-    ] = False  # Whether or not to use procedure codes in addition to diagnosis codes when constructing an input
-    num_loader_workers: Optional[int] = 3
-    visit_days_resolution: Optional[
-        int
-    ] = 2  # consecutive visits that are less than visit_days_resolution days apart are merged into one visit.
-    # Note that merging stops only when number of days between consecutive visits is more than visit_days_resolution, so
-    # within a group of merged visits there may be some more than visit_days_resolution days apart.
-    # E.g. for visit days 1, 3, 5, 9, and for visit_days_resolution=2, days 1, 3, and 5 will be merged into a single visit, and day 10 will be a separate visit
-    input_column_names: Optional[list] = None
-    min_visits_per_patient: Optional[int] = 5  # patients with less visits are ignored
-
-    # When predicting next visit, this will map regular visit dictionary (codes to indices, index per code) to reduced
-    # visit dictionary (codes to indices, code groups mapped to (fewer) indices). When no mapping is used, this is null.
-    reduce_pred_visit_vocab_mapping_path: Optional[str] = "/ehrtransformers/data_access/icd_to_ccs.json"  # null
-    data_source: Optional[str] = "csv"  # sql_db #csv
-    data_source_str: Optional[
-        str
[…]
