+++ examples/fuse_examples/multimodality/ehr_transformer/ehrtransformers/model/OutcomeFuse.py
@@ -1,332 +1,0 @@
-"""
-(C) Copyright 2021 IBM Corp.
-
-Licensed under the Apache License, Version 2.0 (the "License");
-you may not use this file except in compliance with the License.
-You may obtain a copy of the License at
-
-   http://www.apache.org/licenses/LICENSE-2.0
-
-Unless required by applicable law or agreed to in writing, software
-distributed under the License is distributed on an "AS IS" BASIS,
-WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-See the License for the specific language governing permissions and
-limitations under the License.
-
-Created on June 30, 2021
-
-"""
-
-import sys
-
-sys.path.insert(0, "../")
-
-import numpy as np
-import torch
-from torch.utils.data.dataset import Dataset
-from torch.utils.data import DataLoader
-from datetime import datetime
-import ehrtransformers.configs.naming as naming
-
-from ehrtransformers.data_access.utils import (
-    detector_idx,
-    position_idx,
-    seq_pad,
-    seq_translate,
-)
-from ehrtransformers.utils.common import load_pkl
-from ehrtransformers.utils.stats import StatsAnalyzer
-
-from fuse.data.datasets.dataset_wrap_seq_to_dict import DatasetWrapSeqToDict
-from fuse.data.utils.samplers import BatchSamplerDefault
-from fuse.data.utils.collates import CollateDefault
-
-
-# Set logs/outputs dirs, and save source code relevant to this run
-
-
-def load_outcome_data(
-    file_name,
-    dset_name,
-    naming_conventions,
-    head_config,
-    model_config,
-    age_month_resolution,
-    ageVocab,
-    token2idx,
-    max_sequence_len,
-    batch_size,
-    shuffle,
-    reverse_inputs=False,
-    num_workers=3,
-    do_contrastive_sampling=False,
-):
-    """
-
-    :param file_name: path of the input dataset
-    :param dset_name: name of the dataset
-    :param naming_conventions: naming conventions dict
-    :param head_config: head config class as defined in ehrtransformers/configs/head_config.py
-    :param age_month_resolution: 1: age in months, 12: age in years
-    :param ageVocab: age vocabulary
-    :param token2idx: input (diagnoses, procedures) vocabulary
-    :param max_sequence_len: max number of tokens in an input sequence
-    :param batch_size: batch size
-    :param shuffle: boolean, True - shuffles the data, False - preserves data order
-    :param reverse_inputs: boolean, if True the input sequence is reversed, i.e. the last visit is the first input.
-        Since the first input embedding is used as the entire patient representation, it makes more sense to make it
-        the last visit rather than the first one.
-    :return:
-    """
-
-    data_dict = load_pkl(file_name)
-    data_df = rename_columns_for_behrt(
-        data_dict["data_df"],
-        naming_conf=naming_conventions,
-        age_month_resolution=age_month_resolution,
-    )
-    # add relevant external GT:
-    data_df = head_config.add_external_gt_to_df(data_df, [dset_name])
-
-    stats_inst_tst = StatsAnalyzer(
-        data_dict["data_df"],
-        naming_conventions,
-        title=dset_name,
-    )
-    stats_inst_tst.print_stats()
-    Dset = Outcome(
-        head_config=head_config,
-        token2idx=token2idx,
-        age2idx=ageVocab,
-        dataframe=data_df,
-        max_len=max_sequence_len,
-        reverse_inputs=reverse_inputs,
-    )
-
-    out_mapping = [
-        "data.age",
-        "data.code",
-        "data.position",
-        "data.segment",
-        "data.mask",
-        "data.patid",
-        "data.vis_date",
[…]

[…]
