### 2023-01-12 ea04710
more comments and flake8 fixes added
