+++ examples/fuse_examples/multimodality/ehr_transformer/dataset.py
@@ -71,9 +71,9 @@ class OpAddBMI(OpBase):
 class OpConvertVisitToSentence(OpBase):
     """
     Converts the observations presented in data frame
-     to sentence by visit time. Sentence will be presented as list of values
-     per visit (converted to categorical strings) and separator at the end of the sentence
-
+     to sentence by visit time. Sentence will be presented as
+     list of values per visit (converted to categorical strings)
+     and separator at the end of the sentence
     """
 
     def __call__(self, sample_dict) -> Any:
@@ -118,8 +118,9 @@ class OpGenerateFinalTrajectoryOfVisits(OpBase):
         d_visits_sentences = sample_dict["VisitSentences"]
         n_visits = len(d_visits_sentences)
 
-        # extract sub-set of trajectory by random selection of start and stop keys
-        # we use n_visits- 1 to keep the last visit for outcome in classifier
+        # extract sub-set of trajectory by random selection of
+        # start and stop keys we use n_visits- 1 to keep the last
+        # visit for outcome in classifier
         start_visit = randrange(int((n_visits - 1) / 2))
         stop_visit = n_visits - randrange(int((n_visits - 1) / 2)) - 1
         keys = list(d_visits_sentences.keys())
@@ -135,13 +136,15 @@ class OpGenerateFinalTrajectoryOfVisits(OpBase):
                     static_embeddings += [df_static[k]]
 
         # Build trajectory of visits
-        # appends sentences in the reverse order from stop to start in order to get X last sentences than less
-        # than max_len
+        # appends sentences in the reverse order from
+        # stop to start in order to get X last sentences that
+        # less than max_len
         trajectory_tokens = []
         for k in reversed(trajectory_keys):
             visit_sentence = d_visits_sentences[k]
 
-            if self._embed_static_in_all_visits:  # add static details to every visit
+            # add static details to every visit
+            if self._embed_static_in_all_visits:
                 visit_sentence = static_embeddings + visit_sentence
 
             if (len(trajectory_tokens) + len(visit_sentence)) < self._max_len:
@@ -173,18 +176,25 @@ class OpGenerateFinalTrajectoryOfVisits(OpBase):
 
 class OpMapToCategorical(OpBase):
     """
-    Converts continuous values of observations to the category based on calculated "bins"
-         each pair of name + value will be converted to the word "Name_binN"
-         Example: Age = 80 will be converted to "Age_3" (third bin in calculated distribution)
+    Converts continuous values of observations
+    to the category based on calculated "bins"
+    each pair of name + value will be converted
+    to the word "Name_binN"
+
+     Example: Age = 80 will be converted to "Age_3"
+      (third bin in calculated distribution)
     """
 
     def __call__(self, sample_dict, bins: dict) -> Any:
 
-        # convert continuous measurements to categorical ones based on defined bins
-        # mapping static clinical characteristics (Age, Gender, ICU type, Height, etc)
+        # convert continuous measurements to categorical ones based
+        # on defined bins mapping static clinical characteristics
+        # (Age, Gender, ICU type, Height, etc)
         for k in sample_dict["StaticDetails"]:
             sample_dict["StaticDetails"][k] = k + "_" + str(np.digitize(sample_dict["StaticDetails"][k], bins[k]))
-        # mapping labs exams and clinical characteristics captured during patients' stay in ICU
+
+        # mapping labs exams and clinical characteristics captured
+        # during patients' stay in ICU
         if not sample_dict["Visits"].empty:
             sample_dict["Visits"]["Value"] = sample_dict["Visits"].apply(
                 lambda row: row["Parameter"] + "_" + str(np.digitize(row["Value"], bins[row["Parameter"]]))
[…]
