+++ examples/fuse_examples/multimodality/ehr_transformer/dataset.py
@@ -10,6 +10,7 @@ import pandas as pd
 
 from fuse.data import DatasetDefault, PipelineDefault, OpBase
 from fuse.data.ops.ops_read import OpReadDataframe
+from fuse.data.ops.ops_common import OpLookup, OpSetIfNotExist 
 
 from fuse.data.utils.split import dataset_balanced_division_to_folds
 from fuse.data.utils.export import ExportDataset
@@ -170,6 +171,8 @@ class OpGenerateFinalTrajectoryOfVisits(OpBase):
         # outcomes of next visit prediction
         sample_dict["NextVisitTokens"] = d_visits_sentences[next_visit]
         sample_dict["NextVisitTokenIds"] = seq_translate(sample_dict["NextVisitTokens"], self._vocab)
+        sample_dict["NextVisitLabels"] = np.zeros(len(self._vocab))
+        sample_dict["NextVisitLabels"][sample_dict["NextVisitTokenIds"]] = 1.0
 
         return sample_dict
 
@@ -471,6 +474,8 @@ class PhysioNetCinC:
                 ),
                 dict(),
             ),
+            (OpSetIfNotExist(), dict(key="StaticDetails.Gender", value=-1)),
+            (OpLookup(map={"Gender_0": 0, "Gender_1": 1, -1:-1}), dict(key_in="StaticDetails.Gender", key_out="Gender"))
         ]
 
     @staticmethod
