+++ examples/fuse_examples/multimodality/ehr_transformer/main_train.py
@@ -2,6 +2,7 @@ from dataset import PhysioNetCinC
 from typing import Any, Optional, List, Tuple
 import hydra
 from omegaconf import DictConfig
+from copy import deepcopy
 
 import torch
 import torch.nn.functional as F
@@ -24,6 +25,19 @@ from fuse.data import DatasetDefault
 
 from examples.fuse_examples.multimodality.ehr_transformer.model import Embed, TransformerEncoder
 
+def filter_gender_label_unknown_for_loss(batch_dict: NDict) -> NDict:
+    # filter out samples
+    keep_indices = batch_dict["Gender"] != -1
+    batch_dict = batch_dict.indices(
+        keep_indices
+    )  # this will keep only a subset of elements per each key in the dictionary (where applicable)
+    return batch_dict
+
+
+def filter_gender_label_unknown_for_metric(sample_dict: NDict) -> NDict:
+    # filter out samples
+    sample_dict["filter"] = sample_dict["Gender"] == -1
+    return sample_dict
 
 def data(
     dataset_cfg: dict, target_key: str, batch_size: int, data_loader_train: dict, data_loader_valid: dict
@@ -31,11 +45,9 @@ def data(
 
     token2idx, ds_train, ds_valid, _ = PhysioNetCinC.dataset(**dataset_cfg)
 
-    ds_train[0].print_tree()
-
     dl_train = DataLoader(
         ds_train,
-        collate_fn=CollateDefault(keep_keys=["data.sample_id", "Target", "Indexes"]),
+        collate_fn=CollateDefault(keep_keys=["data.sample_id", "Target", "Indexes", "Gender", "NextVisitLabels"]),
         batch_sampler=BatchSamplerDefault(
             ds_train,
             balanced_class_name=target_key,
@@ -46,7 +58,7 @@ def data(
         **data_loader_train,
     )
     dl_valid = DataLoader(
-        ds_valid, collate_fn=CollateDefault(keep_keys=["data.sample_id", "Target", "Indexes"]), **data_loader_valid
+        ds_valid, collate_fn=CollateDefault(keep_keys=["data.sample_id", "Target", "Indexes", "Gender", "NextVisitLabels"]), **data_loader_valid
     )
 
     return token2idx, ds_train, dl_train, dl_valid
@@ -58,12 +70,16 @@ def model(
     z_dim: int,
     transformer_encoder: dict,
     vocab_size: int,
+    aux_gender_classification: bool,
+    classifier_gender_head: dict,
+    aux_next_vis_classification: bool,
+    classifier_next_vis_head: dict
 ):
     embed = Embed(key_in="Indexes", key_out="model.embedding", n_vocab=vocab_size, **embed)
 
     encoder_model = TransformerEncoder(**transformer_encoder)
 
-    model = torch.nn.Sequential(
+    models_sequence = [
         embed,
         ModelWrapSeqToDict(
             model=encoder_model,
@@ -71,8 +87,16 @@ def model(
             model_outputs=["model.z", None],
         ),
         Head1D(head_name="cls", conv_inputs=[("model.z", z_dim)], **classifier_head),
-    )
+    ]
 
+    # append auxiliary head for gender 
+    if aux_gender_classification:
+        models_sequence.append(Head1D(head_name="gender", conv_inputs=[("model.z", z_dim)], **classifier_gender_head))
+
+    if aux_next_vis_classification:
+        models_sequence.append(Head1D(head_name="next_vis", conv_inputs=[("model.z", z_dim)], num_outputs=vocab_size, **classifier_next_vis_head))
+
+    model = torch.nn.Sequential(*models_sequence)
     return model
 
 
@@ -84,13 +108,17 @@ def train(
[…]
