### 2023-01-12 #4943 @alex-golts
The "hello world" notebook is meant to be simple and in simple cases like a standard MNIST classification pipeline it can be indeed a bit more challenging to demonstrate the particular usefulness of the fuse "philosophy" vs. standard approaches. But to give one example that may emphasize some of the benefits, even looking at just this simple example:

In the instantiation of `BatchSamplerDefault`, we pass a key name (string) that specifies the label by which to balance samples. This assumes the notion of a `sample_dict` that represents a data sample that will have this key. Once users across different projects agree on this methodology and use `sample_dict`s to store their data, it makes it easier to reuse code. For example, the code for creating a balanced sampler in another project would be exactly the same (with maybe only the key name being different). In a standard approach, different projects would have the data stored in different ways, so they might need to have a different implementation of this custom sampler per project, or they will need to agree on another convention to make the code fully reusable.
The `BatchSamplerDefault` is indeed reusable between different examples that we have, you can see it used in the same way for example [here](https://github.com/BiomedSciAI/fuse-med-ml/blob/master/examples/fuse_examples/imaging/classification/knight/baseline/fuse_baseline.py) and [here](https://github.com/BiomedSciAI/fuse-med-ml/blob/master/examples/fuse_examples/imaging/classification/stoic21/runner_stoic21.py).

I agree that it's possible to think of different conventions than `NDict` of course, and as long as those are used, it could be equally convenient to use other types of well documented classes or functions. So in that sense, `NDict` is a choice that we made that we find convenient. but the usefulness is in having many generic building blocks that all "comply" with this methodology. and if users decide to use it for their project from the start, then we believe they will find it convenient to reuse code from existing projects that use the same methodology.

We will add some comments along the "hello world" notebook to make this point more clear. And we also have a couple more thoughts about how to make this notebook better at showcasing the way fuse facilitates better code reuse. we're still discussing it and plan to make some improvements going forward.
