+++ examples/fuse_examples/multimodality/ehr_transformer/config.yaml
@@ -1,4 +1,4 @@
-name: ehr_transformer_basic_with_aux
+name: ehr_transformer_4
 root: "."
 
 target_key: "Target"
@@ -26,7 +26,7 @@ data:
     embed_static_in_all_visits: 0
   
  
-  batch_size: 256
+  batch_size: 128
   target_key: ${target_key}
   
   data_loader_train: # Dataloader constructor parameters
@@ -38,7 +38,9 @@ data:
 
 
 model:
-  transformer_encoder:
+  encoder_type: "transformer" # "bert" "transformer"
+
+  transformer_encoder: # used when encoder type is "transformer"
     num_tokens: ${max_len_seq} 
     token_dim: ${model.embed.emb_dim}
     depth: 4
@@ -47,9 +49,19 @@ model:
     dropout: 0.0
     emb_dropout: 0.0
     num_cls_tokens: 1
-  
+
+  bert_config_kwargs:
+    hidden_size: ${model.z_dim} #48  #240 #72 #120 #288, # word embedding and seg embedding hidden size (needs to be a multiple of attention heads)
+    num_hidden_layers: 6 # number of multi-head attention layers required
+    num_attention_heads: 24 # number of attention heads
+    intermediate_size: 512 #128 #512 # the size of the "intermediate" layer in the transformer encoder
+    hidden_act: gelu # activation function ("gelu", 'relu', 'swish')
+    hidden_dropout_prob: 0.2 # dropout rate
+    attention_probs_dropout_prob: 0.22 # multi-head attention dropout rate
+    initializer_range: 0.02 # parameter weight initializer range
+    
   embed:
-    emb_dim: 80
+    emb_dim: ${model.z_dim}
 
   classifier_head:
     num_outputs: 2
@@ -65,42 +77,55 @@ model:
   aux_gender_classification: ${aux_gender_classification}
   aux_next_vis_classification: ${aux_next_vis_classification}
   
-  z_dim: 80
+  z_dim: 48
 
 # train
 train: # arguments for train() in classifiers_main_train.py
   model_dir: ${root}/${name}
   target_key: ${target_key}
-  target_loss_weight: 0.0006
+  target_loss_weight: 0.8
   aux_gender_classification: ${aux_gender_classification}
-  gender_loss_weight:  0.0002
+  gender_loss_weight:  0.1
   aux_next_vis_classification: ${aux_next_vis_classification}
-  next_vis_loss_weight: 1.0
+  next_vis_loss_weight: 0.1
 
-  track_clearml:
-    project_name: "pdes"
-    task_name: ${name}
-    tags: "classifiers"
-    reuse_last_task_id: True
-    continue_last_task: False
+  # uncomment to track in clearml
+  # track_clearml:
+  #   project_name: "ehr_transformer"
+  #   task_name: ${name}
+  #   tags: "fuse_example"
+  #   reuse_last_task_id: True
+  #   continue_last_task: False
 
+  
+  # SGD
   opt:
     _partial_: true
-    # unmask to use SGD instead
     _target_: torch.optim.SGD
     momentum: 0.99
     nesterov: True
     lr: 0.001
-    # weight_decay: 1e-5
+  #   # weight_decay: 1e-5
 
-    # Adam
-    # _target_: torch.optim.Adam
-    # lr: 1e-3
+  # AdamW
+  # opt:
+  #   _partial_: true
+  #   _target_: torch.optim.AdamW
+  #   lr: 1e-3
+  
+  # lr_scheduler:
+  #   _partial_: True
+  #   _target_: transformers.get_linear_schedule_with_warmup
+  #   num_warmup_steps: 1000
[…]
