### 2023-01-12 c2b17ff
fixes of Readme files and better configuration of pickle file
