+++ examples/fuse_examples/multimodality/ehr_transformer/README.md
@@ -41,7 +41,7 @@ The network was implemented in a manner that allows several output ”heads”, 
 
 The patients’ data is passing a set of preprocessing steps before entering the modeling stage, which includes the exclusion of illegal patients, mapping continuous values of observation to categorical values by applying digitization, and combining both static data (general information) and time series in one input patient trajectory sequence. 
 
-As mentioned, the network has multiple prediction heads, related to multiple prediction tasks. We trained and tested multiple networks consisting of a combination of the following tasks:
+As mentioned, the network has multiple prediction heads, related to multiple prediction tasks. Using multiple heads aims to improve patients' representation.  We trained and tested multiple networks consisting of a combination of the following tasks:
 
  <br/>- **Next Visit Prediction** A multi-label classification task of predicting the observation values that will appear in the visit following the last visit of the trajectory.
 <br/>- **Gender Prediction**: A binary classifier that predicts the biologically determined gender of the
@@ -58,8 +58,8 @@ Download the following data to a designated folder in your system:
 
 In the configuration file config.yaml hydra file:
 1.	Define environment variable CINC_DATA_PATH to the data folder. 
-2.	There is a possibility to define an environment variable CINC_DATA_PKL of a path to the pickle file that will manage the loaded data for the next runs. 
+2.	There is a possibility to define an environment variable CINC_DATA_PKL of a path to the pickle file that will manage the loaded data for the next runs. Use 'null' (that is the default) in configuration if you are not using this option
 3.	Configure network hyperparameters
 
-Run:  python3 runner.py
+Run:  python3 main_train.py
 
