+++ examples/fuse_examples/multimodality/ehr_transformer/config.yaml
@@ -1,13 +1,13 @@
-name: ehr_transformer_4
+name: ehr_transformer
 root: "."
 
-target_key: "Target"
+target_key: "Target" # key that points to labels in sample_dict
 max_len_seq: 350 # maximal number of tokens in the trajectory
-aux_gender_classification: True
-aux_next_vis_classification: True
+aux_gender_classification: True # set to True to enable gender auxiliary head 
+aux_next_vis_classification: True # set to True to enable next visit auxiliary head
 
 data: 
-  dataset_cfg:
+  dataset_cfg: # See PhysioNetCinC.dataset for details about the arguments 
     raw_data_pkl: null # in case we want to use pickle for managing raw data use ${oc.env:CINC_DATA_PKL} to define the
     #path to the pickle file
     raw_data_path: ${oc.env:CINC_DATA_PATH}
@@ -39,9 +39,9 @@ data:
 
 
 model:
-  encoder_type: "transformer" # "bert" "transformer"
+  encoder_type: "bert" # supported values: "bert" "transformer"
 
-  transformer_encoder: # used when encoder type is "transformer"
+  transformer_encoder: # TransformerEncoder constructor arguments - used when encoder type is "transformer"
     num_tokens: ${max_len_seq} 
     token_dim: ${model.embed.emb_dim}
     depth: 4
@@ -51,28 +51,28 @@ model:
     emb_dropout: 0.0
     num_cls_tokens: 1
 
-  bert_config_kwargs:
-    hidden_size: ${model.z_dim} #48  #240 #72 #120 #288, # word embedding and seg embedding hidden size (needs to be a multiple of attention heads)
+  bert_config_kwargs: # BertConfig constructor arguments - used when encoder type is "bert" 
+    hidden_size: ${model.z_dim}  # word embedding and seg embedding hidden size (needs to be a multiple of attention heads)
     num_hidden_layers: 6 # number of multi-head attention layers required
     num_attention_heads: 24 # number of attention heads
-    intermediate_size: 512 #128 #512 # the size of the "intermediate" layer in the transformer encoder
-    hidden_act: gelu # activation function ("gelu", 'relu', 'swish')
-    hidden_dropout_prob: 0.2 # dropout rate
-    attention_probs_dropout_prob: 0.22 # multi-head attention dropout rate
-    initializer_range: 0.02 # parameter weight initializer range
+    intermediate_size: 50 # the size of the "intermediate" layer in the transformer encoder
+    hidden_act: relu # activation function ("gelu", 'relu', 'swish')
+    hidden_dropout_prob: 0.0 # dropout rate
+    attention_probs_dropout_prob: 0.0 # multi-head attention dropout rate
+    initializer_range: null # parameter weight initializer range
     
-  embed:
+  embed: # Embed constructor arguments
     emb_dim: ${model.z_dim}
 
-  classifier_head:
+  classifier_head: # HeadD1 constructor arguments - used for main classifer head
     num_outputs: 2
     layers_description: [256]
 
-  classifier_gender_head:
+  classifier_gender_head: # HeadD1 constructor arguments - used for gender classifer head
     num_outputs: 2
     layers_description: [256]
 
-  classifier_next_vis_head:
+  classifier_next_vis_head: # HeadD1 constructor arguments - used for next visit classifer head
     layers_description: [256]
 
   aux_gender_classification: ${aux_gender_classification}
[…]
