### 2023-01-12 b8e2689
change default to bert
