### 2023-01-13 3390fda
Ehr transformer ICU example (#245)
