+++ fuse/dl/tests/test_cat.py
@@ -23,10 +23,7 @@ from fuse.dl.models.backbones.backbone_transformer import CrossAttentionTransfor
 
 
 class TestCrossAttentionTransformer(unittest.TestCase):
-    def test_basic(self) -> None:
-        """
-        Basic testing for the CrossAttentionTransformer model
-        """
+    def test_all_contexts(self) -> None:
 
         # model parameters
         model_params = {
@@ -37,6 +34,15 @@ class TestCrossAttentionTransformer(unittest.TestCase):
             "max_seq_len_b": 1024,
             "output_dim": 256,
         }
+
+        for context in ["seq_a", "seq_b", "both"]:
+            model_params["context"] = context
+            self.validate_model_with_params(model_params)
+
+    def validate_model_with_params(self, model_params: dict) -> None:
+        """
+        Basic testing for the CrossAttentionTransformer model
+        """
 
         # init model
         model = CrossAttentionTransformer(**model_params)
@@ -51,7 +57,9 @@ class TestCrossAttentionTransformer(unittest.TestCase):
         # validation
         assert output.shape[0] == 1
         if output.shape[1] != model_params["output_dim"]:
-            raise Exception(f"Expected output dimension to be {model_params['output_dim']}, but got: {output.shape[1]}")
+            raise Exception(
+                f"Expected output dimension to be {model_params['output_dim']}, but got: {output.shape[1]}. model parameters are: {model_params}."
+            )
 
 
 if __name__ == "__main__":
