+++ fuse/dl/tests/test_cat.py
@@ -19,11 +19,15 @@ Created on Jan 11, 2023
 
 import torch
 import unittest
+import random
 from fuse.dl.models.backbones.backbone_transformer import CrossAttentionTransformer
 
 
 class TestCrossAttentionTransformer(unittest.TestCase):
     def test_all_contexts(self) -> None:
+        """
+        test each of the three options for context: "seq_a", "seq_b" and "both using the helper function
+        """
 
         # model parameters
         model_params = {
@@ -41,15 +45,19 @@ class TestCrossAttentionTransformer(unittest.TestCase):
 
     def validate_model_with_params(self, model_params: dict) -> None:
         """
-        Basic testing for the CrossAttentionTransformer model
+        Basic validation for the CrossAttentionTransformer model
+
+        :param model_params: A dictionary of the model's parameters to validate
         """
 
         # init model
         model = CrossAttentionTransformer(**model_params)
 
         # init random sequences that don't exceed max sequences length
-        s1 = torch.randint(0, model_params["num_tokens_a"], (1, 400))
-        s2 = torch.randint(0, model_params["num_tokens_b"], (1, 1000))
+        seq_a_len = random.randint(0, model_params["max_seq_len_a"])
+        seq_b_len = random.randint(0, model_params["max_seq_len_b"])
+        s1 = torch.randint(0, model_params["num_tokens_a"], (1, seq_a_len))
+        s2 = torch.randint(0, model_params["num_tokens_b"], (1, seq_b_len))
 
         # processing sample
         output = model(s1, s2)
@@ -58,7 +66,7 @@ class TestCrossAttentionTransformer(unittest.TestCase):
         assert output.shape[0] == 1
         if output.shape[1] != model_params["output_dim"]:
             raise Exception(
-                f"Expected output dimension to be {model_params['output_dim']}, but got: {output.shape[1]}. model parameters are: {model_params}."
+                f"Expected output dimension to be {model_params['output_dim']}, but got: {output.shape[1]}. used model parameters: {model_params}."
             )
 
 
