### 2023-01-18 bed2fd6
Adding unitest for ehr_transformer
