### 2023-01-18 3cce8e6
Adding unitest for ehr_transformer
