+++ examples/fuse_examples/multimodality/ehr_transformer/main_train.py
@@ -1,4 +1,3 @@
-from dataset import PhysioNetCinC
 from typing import Any, Optional, Tuple
 import hydra
 from omegaconf import DictConfig
@@ -23,6 +22,7 @@ from fuse.utils import NDict
 
 
 from examples.fuse_examples.multimodality.ehr_transformer.model import Embed, TransformerEncoder, Bert, BertConfig
+from examples.fuse_examples.multimodality.ehr_transformer.dataset import PhysioNetCinC
 
 
 def filter_gender_label_unknown_for_loss(batch_dict: NDict) -> NDict:
