+++ examples/fuse_examples/multimodality/ehr_transformer/main_train.py
@@ -20,7 +20,6 @@ from fuse.eval.metrics.classification.metrics_classification_common import Metri
 from fuse.dl.losses import LossDefault
 from fuse.utils import NDict
 
-
 from examples.fuse_examples.multimodality.ehr_transformer.model import Embed, TransformerEncoder, Bert, BertConfig
 from examples.fuse_examples.multimodality.ehr_transformer.dataset import PhysioNetCinC
 
