+++ examples/fuse_examples/multimodality/ehr_transformer/main_train.py
@@ -1,7 +1,3 @@
-<<<<<<< HEAD
-=======
-from dataset import PhysioNetCinC
->>>>>>> a86e706094d3dfbce51cc090f65055d386a5730c
 from typing import Any, Optional, Tuple
 import hydra
 from omegaconf import DictConfig
@@ -24,13 +20,8 @@ from fuse.eval.metrics.classification.metrics_classification_common import Metri
 from fuse.dl.losses import LossDefault
 from fuse.utils import NDict
 
-<<<<<<< HEAD
 from examples.fuse_examples.multimodality.ehr_transformer.model import Embed, TransformerEncoder, Bert, BertConfig
 from examples.fuse_examples.multimodality.ehr_transformer.dataset import PhysioNetCinC
-=======
-
-from examples.fuse_examples.multimodality.ehr_transformer.model import Embed, TransformerEncoder, Bert, BertConfig
->>>>>>> a86e706094d3dfbce51cc090f65055d386a5730c
 
 
 def filter_gender_label_unknown_for_loss(batch_dict: NDict) -> NDict:
