### 2023-01-18 22a9d8e
added random batch size to test
