### 2023-01-18 c45ad6b
a more configureable resnet3d implementation (controls conv channels, strides and stem kernel size)
