### 2023-01-19 1fa5c2f
Ehr transformer example (#256)
* Adding unitest for ehr_transformer

* Adding unitest for ehr_transformer

* Fixed to unitest

* Fixed to unitest

* Fixed to unitest

* Fixed to unitest

* Fixed merging messages

* Added temp dir for the parallel  tests

Co-authored-by: Moshe Raboh Moshiko.Raboh@ibm.com <moshiko@cccxl016.pok.ibm.com>
Co-authored-by: Moshe Raboh Moshiko.Raboh@ibm.com <moshiko@cccxc409.pok.ibm.com>
Co-authored-by: Moshiko Raboh <86309179+mosheraboh@users.noreply.github.com>
