### 2023-01-19 #4943 @alex-golts
@jmschrei I have fixed the invalid DOI.

can you please clarify what exactly you mean by "Zenodo DOI for the paper"? I see according to the [submission instructions](https://joss.readthedocs.io/en/latest/submitting.html#submission-requirements) that we're expected to deposit the **repository** to a service like Zenodo. did you mean that? in that case we already have it [here](https://zenodo.org/record/5146491).
Or did you mean to just deposit the paper? If that's the case, should it be the PDF? the raw "paper.md" and "paper.bib" files?
