### 2023-01-19 547d1ec
adding time_unit, time_followup to expected_cindex
