### 2023-01-21 #4943 @Kevin-Mattheus-Moerman
@alex-golts and co-authors, I am the AEiC on this track and here to help process final steps towards acceptance in JOSS. Before we can proceed, below are some some minor issues that require your attention:

On the paper (You can call `@editorialbot generate pdf` to update the draft):
- [ ] The paper mentions

> A data pipeline may consist of a `static_pipline` and a `dynamic_pipeline`.

Check if `static_pipline` should read `static_pipeline`.

- [ ] Fix spelling for `acording`, which should read `according`

On the archive:
- [ ] Please manually edit ZENODO archive title to match the paper title
- [ ] Please also manually edit the ZENODO archive author list to match the paper authors and order (you may need to remove any contributors that were added on the ZENODO list)
- [ ] Please manually edit the ZENODO listed license (currently it says `Other (Open)`) to match your actual software's open source license.
