+++ fuse/data/datasets/dataset_default.py
@@ -187,8 +187,9 @@ class DatasetDefault(DatasetBase):
         # get sample id
         if not self._explicit_sample_ids_mode:
             sample_id = item
-            if sample_id >= self._final_sample_ids:
-                raise IndexError
+            if isinstance(sample_id, (int, np.integer)):
+                if sample_id >= self._final_sample_ids:
+                    raise IndexError
 
         elif not isinstance(item, (int, np.integer)):
             sample_id = item
