### 2023-01-22 0ef8273
allow non-int sample ids in DatasetDefault
