### 2023-01-22 #4943 @alex-golts
@Kevin-Mattheus-Moerman we fixed the typos, thanks for pointing them out.
we also edited the title, author list and license on Zenodo according to your comments.
please let me know if you have more comments.
