### 2023-01-22 806f4bf
explicitly support no sample ids
