### 2023-01-22 a079521
add strict boolean argument. when strict=False, it allows to patchify incompatible shapes.
