### 2023-01-23 #4943 @alex-golts
Thank you all for the thorough and helpful reviews and for the editorial work! it was a great experience to go through the unique thoughtful and transparent review process of JOSS. we feel it definitely helped us improve both the paper and code.
